Frozen Rush Data Storage

Freezes the Redis data base of an ArchiveAsync to a data.table::data.table(). No further points can be added to the archive but the data can be accessed and analyzed. Useful when the Redis data base is not permanently available. Use the callback bbotk.async_freeze_archive to freeze the archive after the optimization has finished.

S3 Methods

as.data.table(archive)
ArchiveAsync -> data.table::data.table()
Returns a tabular view of all performed function calls of the Objective. The x_domain column is unnested to separate columns.

See also

Super classes

Archive -> ArchiveAsync -> ArchiveAsyncFrozen

Active bindings

data

(data.table::data.table)
Data table with all finished points.

queued_data

(data.table::data.table)
Data table with all queued points.

running_data

(data.table::data.table)
Data table with all running points.

finished_data

(data.table::data.table)
Data table with all finished points.

failed_data

(data.table::data.table)
Data table with all failed points.

n_queued

(integer(1))
Number of queued points.

n_running

(integer(1))
Number of running points.

n_finished

(integer(1))
Number of finished points.

n_failed

(integer(1))
Number of failed points.

n_evals

(integer(1))
Number of evaluations stored in the archive.

Methods

Public methods

ArchiveAsyncFrozen$new()

ArchiveAsyncFrozen$push_points()

ArchiveAsyncFrozen$push_point()

ArchiveAsyncFrozen$push_running_points()

ArchiveAsyncFrozen$push_running_point()

ArchiveAsyncFrozen$push_finished_points()

ArchiveAsyncFrozen$push_finished_point()

ArchiveAsyncFrozen$push_failed_points()

ArchiveAsyncFrozen$push_failed_point()

ArchiveAsyncFrozen$pop_point()

ArchiveAsyncFrozen$finish_points()

ArchiveAsyncFrozen$finish_point()

ArchiveAsyncFrozen$fail_points()

ArchiveAsyncFrozen$fail_point()

ArchiveAsyncFrozen$push_result()

ArchiveAsyncFrozen$data_with_state()

ArchiveAsyncFrozen$clear()

ArchiveAsyncFrozen$clone()

Inherited methods

Archive$format()

Archive$help()

Archive$print()

ArchiveAsync$best()

ArchiveAsync$nds_selection()

ArchiveAsyncFrozen$new()

Creates a new instance of this R6 class.

Usage

ArchiveAsyncFrozen$new(archive)

Arguments

archive

(ArchiveAsync)
The archive to freeze.

ArchiveAsyncFrozen$push_points()

Push queued points to the archive.

Usage

ArchiveAsyncFrozen$push_points(xss, xss_extra = NULL, extra = NULL)

Arguments

xss

(list of named list())
List of named lists of point values.

xss_extra

(list of named list() | NULL)
List of named lists of additional information.

extra

(list of named list() | NULL)
Deprecated argument for additional information. Use xss_extra instead.

ArchiveAsyncFrozen$push_point()

Push a single queued point to the archive.

Usage

ArchiveAsyncFrozen$push_point(xs, xs_extra = NULL, extra = NULL)

Arguments

xs

(named list())
Named list of point values.

xs_extra

(named list() | NULL)
Named list of additional information.

extra

(named list() | NULL)
Deprecated argument for additional information. Use xs_extra instead.

ArchiveAsyncFrozen$push_running_points()

Push running points to the archive.

Usage

ArchiveAsyncFrozen$push_running_points(xss, xss_extra = NULL, extra = NULL)

Arguments

xss

(list of named list())
List of named lists of point values.

xss_extra

(list of named list() | NULL)
List of named lists of additional information.

extra

(list of named list() | NULL)
Deprecated argument for additional information. Use xss_extra instead.

ArchiveAsyncFrozen$push_running_point()

Push running point to the archive.

Usage

ArchiveAsyncFrozen$push_running_point(xs, xs_extra = NULL, extra = NULL)

Arguments

xs

(named list())
Named list of point values.

xs_extra

(named list() | NULL)
Named list of additional information.

extra

(named list() | NULL)
Deprecated argument for additional information. Use xs_extra instead.

ArchiveAsyncFrozen$push_finished_points()

Push finished points to the archive.

Usage

ArchiveAsyncFrozen$push_finished_points(
  xss,
  yss,
  xss_extra = NULL,
  yss_extra = NULL
)

Arguments

xss

(list of named list())
List of named lists of point values.

yss

(list of named list())
List of named lists of results.

xss_extra

(list of named list() | NULL)
List of named lists of additional information.

yss_extra

(list of named list() | NULL)
List of named lists of additional information.

ArchiveAsyncFrozen$push_finished_point()

Push a single finished point to the archive.

Usage

ArchiveAsyncFrozen$push_finished_point(
  xs,
  ys,
  xs_extra = NULL,
  ys_extra = NULL
)

Arguments

xs

(named list())
Named list of point values.

ys

(named list())
Named list of results.

xs_extra

(named list() | NULL)
Named list of additional information.

ys_extra

(named list() | NULL)
Named list of additional information.

ArchiveAsyncFrozen$push_failed_points()

Push failed points to the archive.

Usage

ArchiveAsyncFrozen$push_failed_points(xss, xss_extra = NULL, conditions = NULL)

Arguments

xss

(list of named list())
List of named lists of point values.

xss_extra

(list of named list() | NULL)
List of named lists of additional information.

conditions

(list | NULL)
List of conditions for each failed point. If NULL, a generic error message is used.

ArchiveAsyncFrozen$push_failed_point()

Push a single failed point to the archive.

Usage

ArchiveAsyncFrozen$push_failed_point(xs, xs_extra = NULL, condition = NULL)

Arguments

xs

(named list())
Named list of point values.

xs_extra

(named list() | NULL)
Named list of additional information.

condition

(any | NULL)
Condition of the failed point. If NULL, a generic error message is used.

ArchiveAsyncFrozen$pop_point()

Pop a point from the queue.

Usage

ArchiveAsyncFrozen$pop_point()

ArchiveAsyncFrozen$finish_points()

Save the results of multiple running points and move them to the finished points.

Usage

ArchiveAsyncFrozen$finish_points(
  keys,
  yss,
  x_domains,
  yss_extra = NULL,
  extra = NULL
)

Arguments

keys

(character())
Keys of the points.

yss

(list of named list())
List of named lists of results.

x_domains

(list())
List of named lists of transformed point values.

yss_extra

(list of named list() | NULL)
List of named lists of additional information.

extra

(list of named list() | NULL)
Deprecated argument for additional information. Use yss_extra instead.

ArchiveAsyncFrozen$finish_point()

Save the results of a running point and move it to the finished points.

Usage

ArchiveAsyncFrozen$finish_point(
  key,
  ys,
  x_domain,
  ys_extra = NULL,
  extra = NULL
)

Arguments

key

(character(1))
Key of the point.

ys

(named list())
Named list of results.

x_domain

(named list())
Named list of transformed point values.

ys_extra

(named list() | NULL)
Named list of additional information.

extra

(named list() | NULL)
Deprecated argument for additional information. Use ys_extra instead.

ArchiveAsyncFrozen$fail_points()

Move multiple running points to the failed points.

Usage

ArchiveAsyncFrozen$fail_points(keys, conditions = NULL)

Arguments

keys

(character())
Keys of the points.

conditions

(list() | NULL)
Conditions of the failed points.

ArchiveAsyncFrozen$fail_point()

Move a running point to the failed points.

Usage

ArchiveAsyncFrozen$fail_point(key, condition = NULL)

Arguments

key

(character(1))
Key of the point.

condition

(any | NULL)
Condition of the failed point.

ArchiveAsyncFrozen$push_result()

Deprecated. Use $finish_point() instead.

Usage

ArchiveAsyncFrozen$push_result(key, ys, x_domain, extra = NULL)

Arguments

key

(character())
Key of the point.

ys

(list())
Named list of results.

x_domain

(list())
Named list of transformed point values.

extra

(list())
Named list of additional information.

ArchiveAsyncFrozen$data_with_state()

Fetch points with a specific state.

Usage

ArchiveAsyncFrozen$data_with_state(
  fields = c("xs", "ys", "xs_extra", "worker_extra", "ys_extra", "condition"),
  states = c("queued", "running", "finished", "failed"),
  reset_cache = FALSE
)

Arguments

fields

(character())
Fields to fetch. Defaults to c("xs", "ys", "xs_extra", "worker_extra", "ys_extra").

states

(character())
States of the tasks to be fetched. Defaults to c("queued", "running", "finished", "failed").

reset_cache

(logical(1))
Whether to reset the cache of the finished points.

ArchiveAsyncFrozen$clear()

Clear all evaluation results from archive.

Usage

ArchiveAsyncFrozen$clear()

ArchiveAsyncFrozen$clone()

The objects of this class are cloneable with this method.

Usage

ArchiveAsyncFrozen$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# example only runs if a Redis server is available
if (mlr3misc::require_namespaces(c("rush", "redux", "mirai"), quietly = TRUE) &&
  redux::redis_available()) {
# define the objective function
fun = function(xs) {
  list(y = - (xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10)
}

# set domain
domain = ps(
  x1 = p_dbl(-10, 10),
  x2 = p_dbl(-5, 5)
)

# set codomain
codomain = ps(
  y = p_dbl(tags = "maximize")
)

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = codomain,
  properties = "deterministic"
)

# start workers
rush::rush_plan(worker_type = "mirai")
mirai::daemons(1)

# initialize instance
instance = oi_async(
  objective = objective,
  terminator = trm("evals", n_evals = 20),
  callback = clbk("bbotk.async_freeze_archive")
)

# load optimizer
optimizer = opt("async_random_search")

# trigger optimization
optimizer$optimize(instance)

# frozen archive
instance$archive

# best performing configuration
instance$archive$best()

# covert to data.table
as.data.table(instance$archive)
}
#>        state          x1         x2           y        timestamp_xs
#>       <char>       <num>      <num>       <num>              <POSc>
#>  1: finished -6.21358293  0.2769503  -68.201348 2026-06-27 12:46:43
#>  2: finished -2.62264757  4.2779729  -64.337761 2026-06-27 12:46:43
#>  3: finished  6.84682656  3.2309502  -52.316468 2026-06-27 12:46:43
#>  4: finished -1.96786549 -3.4438612   -5.940969 2026-06-27 12:46:43
#>  5: finished -9.86571338  2.5243351 -161.313432 2026-06-27 12:46:43
#>  6: finished  0.01208465  3.1304511  -31.534238 2026-06-27 12:46:43
#>  7: finished -4.77998488  1.4299587  -55.592729 2026-06-27 12:46:43
#>  8: finished  7.57503606  1.3830058  -40.291767 2026-06-27 12:46:43
#>  9: finished  8.01326577 -2.8537699  -26.180748 2026-06-27 12:46:43
#> 10: finished  5.39871055  2.9134685  -36.520343 2026-06-27 12:46:43
#> 11: finished  4.82894640 -0.2106949   -5.783161 2026-06-27 12:46:43
#> 12: finished  2.92831418 -0.4491262    2.631276 2026-06-27 12:46:44
#> 13: finished -2.48881904  1.6091926  -31.394153 2026-06-27 12:46:44
#> 14: finished  2.11332211  1.0427817   -6.356925 2026-06-27 12:46:44
#> 15: finished  0.87403046  4.3072392  -44.663553 2026-06-27 12:46:44
#> 16: finished -5.88500581 -0.9916228  -56.206896 2026-06-27 12:46:44
#> 17: finished  4.08136954  3.9785972  -43.032918 2026-06-27 12:46:44
#> 18: finished -7.80489880 -3.6260806  -86.528017 2026-06-27 12:46:44
#> 19: finished -9.37102033 -3.4632716 -119.514724 2026-06-27 12:46:44
#> 20: finished  5.22167868  0.7561287  -14.487716 2026-06-27 12:46:44
#>        state          x1         x2           y        timestamp_xs
#>       <char>       <num>      <num>       <num>              <POSc>
#>                             worker_id        timestamp_ys
#>                                <char>              <POSc>
#>  1: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#>  2: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#>  3: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#>  4: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#>  5: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#>  6: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#>  7: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#>  8: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#>  9: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#> 10: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#> 11: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:43
#> 12: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:44
#> 13: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:44
#> 14: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:44
#> 15: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:44
#> 16: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:44
#> 17: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:44
#> 18: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:44
#> 19: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:44
#> 20: artsycraftsy_easteuropeanshepherd 2026-06-27 12:46:44
#>                             worker_id        timestamp_ys
#>                                <char>              <POSc>
#>                                     keys x_domain_x1 x_domain_x2
#>                                   <char>       <num>       <num>
#>  1: 8ab9c7ba-b36f-4090-9439-37bf931b9356 -6.21358293   0.2769503
#>  2: 97b0553e-c436-4350-b0c5-244f93384063 -2.62264757   4.2779729
#>  3: 28e5f83a-ce62-4555-bc9c-6b786fe0aba4  6.84682656   3.2309502
#>  4: 86fc67e1-2218-4ad4-8787-5370c9f55e18 -1.96786549  -3.4438612
#>  5: d9dc7e38-a728-423d-bce4-558028ea0c69 -9.86571338   2.5243351
#>  6: 1100eac1-256e-49c8-b4a7-970dabf9d394  0.01208465   3.1304511
#>  7: 1a45adf5-28c7-4b08-8d45-0702f05f5af9 -4.77998488   1.4299587
#>  8: f4a7196b-707a-4228-a909-e9cd08e6fb1f  7.57503606   1.3830058
#>  9: 64eda53a-f8dc-4fae-9116-40d12a63915c  8.01326577  -2.8537699
#> 10: da1cb18a-8364-4a94-907c-31c794d5ba7f  5.39871055   2.9134685
#> 11: 35123c86-df87-4bd3-b5af-42da43ebc482  4.82894640  -0.2106949
#> 12: 46999064-9140-4c6c-afac-a2660dee4539  2.92831418  -0.4491262
#> 13: a9459bcc-9b67-49dc-ba3b-ee0f6071ffc6 -2.48881904   1.6091926
#> 14: b33ccea4-482a-4fe6-b6c0-a8bd22b304ec  2.11332211   1.0427817
#> 15: a05b948d-8d94-42ae-83bd-eaaf934433f5  0.87403046   4.3072392
#> 16: 490278aa-6f0a-401f-b01c-f7dfa78cb635 -5.88500581  -0.9916228
#> 17: 1f394de8-e3d8-4b6a-84b4-92861435f3b8  4.08136954   3.9785972
#> 18: b3882fb4-de27-41f5-a206-52d67f8aedda -7.80489880  -3.6260806
#> 19: 5fe1ecb8-93d5-43e4-bbff-6b02faafad2c -9.37102033  -3.4632716
#> 20: 4a1e0c10-6c3b-46fe-88cb-4cf590805075  5.22167868   0.7561287
#>                                     keys x_domain_x1 x_domain_x2
#>                                   <char>       <num>       <num>