Optimization Instance for Asynchronous Optimization

The OptimInstanceAsync specifies an optimization problem for an OptimizerAsync. The function oi_async() creates an OptimInstanceAsyncSingleCrit or OptimInstanceAsyncMultiCrit.

Details

OptimInstanceAsync is an abstract base class that implements the base functionality each instance must provide.

Super class

bbotk::OptimInstance -> OptimInstanceAsync

Public fields

rush

(Rush)
Rush controller for parallel optimization.

Methods

Public methods

OptimInstanceAsync$new()

OptimInstanceAsync$print()

OptimInstanceAsync$clear()

Inherited methods

bbotk::OptimInstance$assign_result()

bbotk::OptimInstance$format()

Method new()

Creates a new instance of this R6 class.

Usage

OptimInstanceAsync$new(
  objective,
  search_space = NULL,
  terminator,
  check_values = FALSE,
  callbacks = NULL,
  archive = NULL,
  rush = NULL,
  label = NA_character_,
  man = NA_character_
)

Arguments

objective

(Objective)
Objective function.

search_space

(paradox::ParamSet)
Specifies the search space for the Optimizer. The paradox::ParamSet describes either a subset of the domain of the Objective or it describes a set of parameters together with a trafo function that transforms values from the search space to values of the domain. Depending on the context, this value defaults to the domain of the objective.

terminator

Terminator
Termination criterion.

check_values

(logical(1))
Should points before the evaluation and the results be checked for validity?

callbacks

(list of mlr3misc::Callback)
List of callbacks.

archive

(Archive).

rush

(Rush)
If a rush instance is supplied, the tuning runs without batches.

label

(character(1))
Label for this object. Can be used in tables, plot and text output instead of the ID.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Method print()

Printer.

Usage

OptimInstanceAsync$print(...)

Arguments

...

(ignored).

Method clear()

Reset terminator and clear all evaluation results from archive and results.

Usage

OptimInstanceAsync$clear()