Single Criterion Optimization Instance for Asynchronous Optimization

The OptimInstanceAsyncSingleCrit specifies an optimization problem for an OptimizerAsync. The function oi_async() creates an OptimInstanceAsyncSingleCrit.

Super classes

bbotk::OptimInstance -> bbotk::OptimInstanceAsync -> OptimInstanceAsyncSingleCrit

Active bindings

result_x_domain

(list())
(transformed) x part of the result in the domain space of the objective.

result_y

(numeric())
Optimal outcome.

Methods

Public methods

OptimInstanceAsyncSingleCrit$new()

OptimInstanceAsyncSingleCrit$assign_result()

OptimInstanceAsyncSingleCrit$clone()

Inherited methods

bbotk::OptimInstance$format()

bbotk::OptimInstanceAsync$clear()

bbotk::OptimInstanceAsync$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimInstanceAsyncSingleCrit$new(
  objective,
  search_space = NULL,
  terminator,
  check_values = FALSE,
  callbacks = NULL,
  archive = NULL,
  rush = NULL
)

Arguments

objective

(Objective)
Objective function.

search_space

(paradox::ParamSet)
Specifies the search space for the Optimizer. The paradox::ParamSet describes either a subset of the domain of the Objective or it describes a set of parameters together with a trafo function that transforms values from the search space to values of the domain. Depending on the context, this value defaults to the domain of the objective.

terminator

Terminator
Termination criterion.

check_values

(logical(1))
Should points before the evaluation and the results be checked for validity?

callbacks

(list of mlr3misc::Callback)
List of callbacks.

archive

(Archive).

rush

(Rush)
If a rush instance is supplied, the tuning runs without batches.

Method assign_result()

The OptimizerAsync object writes the best found point and estimated performance value here. For internal use.

Usage

OptimInstanceAsyncSingleCrit$assign_result(xdt, y)

Arguments

xdt

(data.table::data.table())
Set of untransformed points / points from the search space. One point per row, e.g. data.table(x1 = c(1, 3), x2 = c(2, 4)). Column names have to match ids of the search_space. However, xdt can contain additional columns.

y

(numeric(1))
Optimal outcome.

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimInstanceAsyncSingleCrit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.