Optimization Instance with budget and archive

Wraps a single-criteria Objective function with extra services for convenient evaluation. Inherits from OptimInstance.

Automatic storing of results in an Archive after evaluation.

Automatic checking for termination. Evaluations of design points are performed in batches. Before a batch is evaluated, the Terminator is queried for the remaining budget. If the available budget is exhausted, an exception is raised, and no further evaluations can be performed from this point on.

Super class

bbotk::OptimInstance -> OptimInstanceSingleCrit

Methods

Public methods

OptimInstanceSingleCrit$new()

OptimInstanceSingleCrit$assign_result()

OptimInstanceSingleCrit$clone()

Inherited methods

bbotk::OptimInstance$eval_batch()

bbotk::OptimInstance$format()

bbotk::OptimInstance$objective_function()

bbotk::OptimInstance$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimInstanceSingleCrit$new(objective, search_space = NULL, terminator)

Arguments

objective

(Objective).

search_space

(paradox::ParamSet). If no search space is provided, search space is set to domain of objective.

terminator

(Terminator).

Method assign_result()

The Optimizer object writes the best found point and estimated performance value here. For internal use.

Usage

OptimInstanceSingleCrit$assign_result(xdt, y)

Arguments

xdt

(data.table)
x values as data.table with one row. Contains the value in the search space of the OptimInstance object. Can contain additional columns for extra information.

y

(numeric(1))
Optimal outcome.

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimInstanceSingleCrit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.