Abstract Terminator Class

Abstract Terminator class that implements the base functionality each terminator must provide. A terminator is an object that determines when to stop the optimization.

Termination of optimization works as follows:

Evaluations in a instance are performed in batches.

Before each batch evaluation, the Terminator is checked, and if it is positive, we stop.

The optimization algorithm itself might decide not to produce any more points, or even might decide to do a smaller batch in its last evaluation.

Therefore the following note seems in order: While it is definitely possible to execute a fine-grained control for termination, and for many optimization algorithms we can specify exactly when to stop, it might happen that too few or even too many evaluations are performed, especially if multiple points are evaluated in a single batch (c.f. batch size parameter of many optimization algorithms). So it is advised to check the size of the returned archive, in particular if you are benchmarking multiple optimization algorithms.

Usage,NULL

Technical details

Terminator subclasses can overwrite .status() to support progress bars via the package progressr. The method must return the maximum number of steps (max_steps) and the currently achieved number of steps (current_steps) as a named integer vector.

See also

Public fields

param_set

paradox::ParamSet
Set of control parameters for terminator.

properties

(character())
Set of properties.

unit

(character())
Unit of steps.

Methods

Public methods

Terminator$new()

Terminator$format()

Terminator$print()

Terminator$status()

Terminator$remaining_time()

Terminator$clone()

Method new()

Creates a new instance of this R6 class.

Usage

Terminator$new(param_set = ps(), properties = character())

Arguments

param_set

(paradox::ParamSet)
Set of control parameters for terminator.

properties

(character())
Set of properties.

Method format()

Helper for print outputs.

Usage

Terminator$format(with_params = FALSE)

Arguments

with_params

(logical(1))
Add parameter values to format string.

Method print()

Printer.

Usage

Terminator$print(...)

Arguments

...

(ignored).

Method status()

Returns how many progression steps are made (current_steps) and the amount steps needed for termination (max_steps).

Usage

Terminator$status(archive)

Arguments

archive

(Archive).

Returns

named integer(2).

Method remaining_time()

Returns remaining runtime in seconds. If the terminator is not time-based, the reaming runtime is Inf.

Usage

Terminator$remaining_time(archive)

Arguments

archive

(Archive).

Returns

integer(1).

Method clone()

The objects of this class are cloneable with this method.

Usage

Terminator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.