Reflections for bbotk

Environment which stores various information to allow objects to examine and introspect their structure and properties (c.f. Reflections).

This environment be modified by third-party packages.

The following objects are set by bbotk:

optimizer_properties (character())
Properties of an optimizer, e.g. "dependencies".

objective_properties (character())
Properties of an objective, e.g. "noisy".

bbotk_reflections

Format

environment.

Examples

ls.str(bbotk_reflections)

#> objective_properties :  chr [1:4] "noisy" "single-crit" "multi-crit" "deterministic"
#> optimizer_properties :  chr [1:3] "dependencies" "single-crit" "multi-crit"
#> terminator_properties :  chr [1:2] "single-crit" "multi-crit"