Optimization via Generalized Simulated Annealing

OptimizerGenSA class that implements generalized simulated annealing. Calls GenSA::GenSA() from package GenSA.

Source

Tsallis C, Stariolo DA (1996). “Generalized simulated annealing.” Physica A: Statistical Mechanics and its Applications, 233(1-2), 395--406. doi: 10.1016/s0378-4371(96)00271-3 .

Xiang Y, Gubian S, Suomela B, Hoeng J (2013). “Generalized Simulated Annealing for Global Optimization: The GenSA Package.” The R Journal, 5(1), 13. doi: 10.32614/rj-2013-002 .

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("gensa")
opt("gensa")

Parameters

smooth

logical(1)

temperature

numeric(1)

acceptance.param

numeric(1)

verbose

logical(1)

trace.mat

logical(1)

For the meaning of the control parameters, see GenSA::GenSA(). Note that we have removed all control parameters which refer to the termination of the algorithm and where our terminators allow to obtain the same behavior.

Super class

bbotk::Optimizer -> OptimizerGenSA

Methods

Public methods

OptimizerGenSA$new()

OptimizerGenSA$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerGenSA$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerGenSA$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(paradox)

domain = ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))

search_space = ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))

codomain = ParamSet$new(list(ParamDbl$new("y", tags = "minimize")))

objective_function = function(xs) {
  list(y = as.numeric(xs)^2)
}

objective = ObjectiveRFun$new(fun = objective_function,
                              domain = domain,
                              codomain = codomain)
terminator = trm("evals", n_evals = 10)
instance = OptimInstanceSingleCrit$new(objective = objective,
                             search_space = search_space,
                             terminator = terminator)

optimizer = opt("gensa")

# Modifies the instance by reference
optimizer$optimize(instance)

#>                x x_domain            y
#> 1: -7.903696e-05   <list> 6.246841e-09

# Returns best scoring evaluation
instance$result

#>                x x_domain            y
#> 1: -7.903696e-05   <list> 6.246841e-09

# Allows access of data.table of full path of all evaluations
instance$archive$data()

#>                 x            y x_domain           timestamp batch_nr
#>  1: -8.384997e-01 7.030818e-01   <list> 2020-07-19 05:00:56        1
#>  2: -7.903696e-02 6.246841e-03   <list> 2020-07-19 05:00:56        2
#>  3:  7.992503e-01 6.388011e-01   <list> 2020-07-19 05:00:56        3
#>  4: -7.903696e-02 6.246841e-03   <list> 2020-07-19 05:00:56        4
#>  5: -7.903596e-02 6.246683e-03   <list> 2020-07-19 05:00:56        5
#>  6: -7.903796e-02 6.246999e-03   <list> 2020-07-19 05:00:56        6
#>  7:  7.903696e-02 6.246841e-03   <list> 2020-07-19 05:00:56        7
#>  8:  7.903796e-02 6.246999e-03   <list> 2020-07-19 05:00:56        8
#>  9:  7.903596e-02 6.246683e-03   <list> 2020-07-19 05:00:56        9
#> 10: -7.903696e-05 6.246841e-09   <list> 2020-07-19 05:00:56       10