Optimization via Local Search

OptimizerBatchLocalSearch class that implements a simple Local Search. Local Search starts by determining the n_initial_points initial best points present in the Archive of the OptimInstance. If fewer points than n_initial_points are present, additional initial_random_sample_size points sampled uniformly at random are evaluated and the best n_initial_points initial points are determined.

In each iteration, for each of the n_initial_points initial best points, neighbors_per_point neighbors are generated by local mutation. Local mutation generates a neighbor by sampling a single parameter that is to be mutated and then proceeds as follows: Double parameters (paradox::p_dbl()) are mutated via Gaussian mutation (with a prior standardization to [0, 1] and retransformation after mutation). Integer parameters (paradox::p_int()) undergo the same mutation but are rounded to the closest integer after mutation. Categorical parameters (paradox::p_fct() and paradox::p_lgl()) are mutated via uniform mutation. Note that parameters that are conditioned on (i.e., they are parents of a paradox::Condition, see the dependencies of the search space) are not mutated.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("local_search")
opt("local_search")

Parameters

n_initial_points

integer(1)
Size of the set of initial best points which are used as starting points for the Local Search. Default is 10.

initial_random_sample_size

integer(1)
Number of points that are sampled uniformly at random before the best n_initial_points initial points are determined, if fewer points than n_initial_points are present in the Archive of the OptimInstance. Default is 100.

neighbors_per_point

integer(1)
Number of neighboring points to generate for each of the n_initial_points best starting points in each iteration. Default is 100.

mutation_sd

numeric(1)
Standard deviation used to create neighbors during mutation of numeric parameters on the standardized [0, 1] scale. Default is 0.1.

Archive

The Archive holds the following additional column that is specific to the algorithm:

.point_id (integer(1))
The id (1, ..., n_initial_points) indicating from which of the n_initial_points best points the evaluated point was generated from.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super classes

bbotk::Optimizer -> bbotk::OptimizerBatch -> OptimizerBatchLocalSearch

Methods

Public methods

OptimizerBatchLocalSearch$new()

OptimizerBatchLocalSearch$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$print()

bbotk::OptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerBatchLocalSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerBatchLocalSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

search_space = domain = ps(x = p_dbl(lower = -1, upper = 1))

codomain = ps(y = p_dbl(tags = "minimize"))

objective_function = function(xs) {
  list(y = as.numeric(xs)^2)
}

objective = ObjectiveRFun$new(
 fun = objective_function,
 domain = domain,
 codomain = codomain)

instance = OptimInstanceBatchSingleCrit$new(
 objective = objective,
 search_space = search_space,
 terminator = trm("evals", n_evals = 100))

# evaluate an initial sample of 10 points uniformly at random
# choose the best 3 points as the initial points
# for each of these points generate 10 neighbors
# repeat this process
optimizer = opt("local_search",
  n_initial_points = 3,
  initial_random_sample_size = 10,
  neighbors_per_point = 10)

# modifies the instance by reference
optimizer$optimize(instance)
#>               x  x_domain            y
#>           <num>    <list>        <num>
#> 1: 0.0005011361 <list[1]> 2.511374e-07

# returns best scoring evaluation
instance$result
#>               x  x_domain            y
#>           <num>    <list>        <num>
#> 1: 0.0005011361 <list[1]> 2.511374e-07

# allows access of data.table of full path of all evaluations
as.data.table(instance$archive$data)
#>                  x            y  x_domain           timestamp batch_nr
#>              <num>        <num>    <list>              <POSc>    <int>
#>   1:  0.6633798382 4.400728e-01 <list[1]> 2024-12-10 09:42:36        1
#>   2:  0.1868167301 3.490049e-02 <list[1]> 2024-12-10 09:42:36        1
#>   3:  0.5579412673 3.112985e-01 <list[1]> 2024-12-10 09:42:36        1
#>   4: -0.2044567061 4.180254e-02 <list[1]> 2024-12-10 09:42:36        1
#>   5:  0.6997655481 4.896718e-01 <list[1]> 2024-12-10 09:42:36        1
#>   6:  0.4836911629 2.339571e-01 <list[1]> 2024-12-10 09:42:36        1
#>   7: -0.3644195613 1.328016e-01 <list[1]> 2024-12-10 09:42:36        1
#>   8: -0.7766395523 6.031690e-01 <list[1]> 2024-12-10 09:42:36        1
#>   9: -0.7978092693 6.364996e-01 <list[1]> 2024-12-10 09:42:36        1
#>  10:  0.6001954963 3.602346e-01 <list[1]> 2024-12-10 09:42:36        1
#>  11: -0.1370253994 1.877596e-02 <list[1]> 2024-12-10 09:42:36        2
#>  12: -0.0215169461 4.629790e-04 <list[1]> 2024-12-10 09:42:36        2
#>  13:  0.5445898025 2.965781e-01 <list[1]> 2024-12-10 09:42:36        2
#>  14:  0.4920044611 2.420684e-01 <list[1]> 2024-12-10 09:42:36        2
#>  15:  0.4269822230 1.823138e-01 <list[1]> 2024-12-10 09:42:36        2
#>  16:  0.1559925822 2.433369e-02 <list[1]> 2024-12-10 09:42:36        2
#>  17:  0.0902813743 8.150727e-03 <list[1]> 2024-12-10 09:42:36        2
#>  18:  0.4153557058 1.725204e-01 <list[1]> 2024-12-10 09:42:36        2
#>  19:  0.3997818691 1.598255e-01 <list[1]> 2024-12-10 09:42:36        2
#>  20:  0.0221645000 4.912651e-04 <list[1]> 2024-12-10 09:42:36        2
#>  21:  0.0314772621 9.908180e-04 <list[1]> 2024-12-10 09:42:36        2
#>  22: -0.1778582768 3.163357e-02 <list[1]> 2024-12-10 09:42:36        2
#>  23: -0.0832811451 6.935749e-03 <list[1]> 2024-12-10 09:42:36        2
#>  24:  0.0437959090 1.918082e-03 <list[1]> 2024-12-10 09:42:36        2
#>  25: -0.5583333488 3.117361e-01 <list[1]> 2024-12-10 09:42:36        2
#>  26:  0.0676355593 4.574569e-03 <list[1]> 2024-12-10 09:42:36        2
#>  27: -0.2628358915 6.908271e-02 <list[1]> 2024-12-10 09:42:36        2
#>  28: -0.6937927119 4.813483e-01 <list[1]> 2024-12-10 09:42:36        2
#>  29: -0.0021132379 4.465775e-06 <list[1]> 2024-12-10 09:42:36        2
#>  30: -0.3310923412 1.096221e-01 <list[1]> 2024-12-10 09:42:36        2
#>  31: -0.6360801218 4.045979e-01 <list[1]> 2024-12-10 09:42:36        2
#>  32: -0.5950992386 3.541431e-01 <list[1]> 2024-12-10 09:42:36        2
#>  33: -0.3789507075 1.436036e-01 <list[1]> 2024-12-10 09:42:36        2
#>  34: -0.6185221834 3.825697e-01 <list[1]> 2024-12-10 09:42:36        2
#>  35: -0.3632844360 1.319756e-01 <list[1]> 2024-12-10 09:42:36        2
#>  36: -0.2527166770 6.386572e-02 <list[1]> 2024-12-10 09:42:36        2
#>  37: -0.1987589247 3.950511e-02 <list[1]> 2024-12-10 09:42:36        2
#>  38: -0.1761783370 3.103881e-02 <list[1]> 2024-12-10 09:42:36        2
#>  39: -0.3526675372 1.243744e-01 <list[1]> 2024-12-10 09:42:36        2
#>  40: -0.3563786829 1.270058e-01 <list[1]> 2024-12-10 09:42:36        2
#>  41:  0.0552805788 3.055942e-03 <list[1]> 2024-12-10 09:42:36        3
#>  42:  0.2251610669 5.069751e-02 <list[1]> 2024-12-10 09:42:36        3
#>  43: -0.1504886259 2.264683e-02 <list[1]> 2024-12-10 09:42:36        3
#>  44:  0.1202336660 1.445613e-02 <list[1]> 2024-12-10 09:42:36        3
#>  45: -0.1478261467 2.185257e-02 <list[1]> 2024-12-10 09:42:36        3
#>  46: -0.2840797954 8.070133e-02 <list[1]> 2024-12-10 09:42:36        3
#>  47: -0.0711566573 5.063270e-03 <list[1]> 2024-12-10 09:42:36        3
#>  48: -0.0073062742 5.338164e-05 <list[1]> 2024-12-10 09:42:36        3
#>  49:  0.0005011361 2.511374e-07 <list[1]> 2024-12-10 09:42:36        3
#>  50:  0.1135319935 1.288951e-02 <list[1]> 2024-12-10 09:42:36        3
#>  51:  0.2066745412 4.271437e-02 <list[1]> 2024-12-10 09:42:36        3
#>  52: -0.1826762278 3.337060e-02 <list[1]> 2024-12-10 09:42:36        3
#>  53: -0.1755334509 3.081199e-02 <list[1]> 2024-12-10 09:42:36        3
#>  54:  0.2386403299 5.694921e-02 <list[1]> 2024-12-10 09:42:36        3
#>  55:  0.0732157784 5.360550e-03 <list[1]> 2024-12-10 09:42:36        3
#>  56: -0.0014880499 2.214292e-06 <list[1]> 2024-12-10 09:42:36        3
#>  57:  0.2393805916 5.730307e-02 <list[1]> 2024-12-10 09:42:36        3
#>  58:  0.0236713413 5.603324e-04 <list[1]> 2024-12-10 09:42:36        3
#>  59: -0.2997644265 8.985871e-02 <list[1]> 2024-12-10 09:42:36        3
#>  60: -0.2408414743 5.800462e-02 <list[1]> 2024-12-10 09:42:36        3
#>  61: -0.3983185203 1.586576e-01 <list[1]> 2024-12-10 09:42:36        3
#>  62: -0.3418990436 1.168950e-01 <list[1]> 2024-12-10 09:42:36        3
#>  63: -0.2056122033 4.227638e-02 <list[1]> 2024-12-10 09:42:36        3
#>  64: -0.2971442078 8.829468e-02 <list[1]> 2024-12-10 09:42:36        3
#>  65: -0.5627570063 3.166954e-01 <list[1]> 2024-12-10 09:42:36        3
#>  66: -0.4628425571 2.142232e-01 <list[1]> 2024-12-10 09:42:36        3
#>  67: -0.3037045426 9.223645e-02 <list[1]> 2024-12-10 09:42:36        3
#>  68: -0.3574463727 1.277679e-01 <list[1]> 2024-12-10 09:42:36        3
#>  69: -0.0547147830 2.993707e-03 <list[1]> 2024-12-10 09:42:36        3
#>  70: -0.3698035044 1.367546e-01 <list[1]> 2024-12-10 09:42:36        3
#>  71:  0.0630691002 3.977711e-03 <list[1]> 2024-12-10 09:42:36        4
#>  72:  0.0126809148 1.608056e-04 <list[1]> 2024-12-10 09:42:36        4
#>  73:  0.3459077198 1.196522e-01 <list[1]> 2024-12-10 09:42:36        4
#>  74:  0.0229600936 5.271659e-04 <list[1]> 2024-12-10 09:42:36        4
#>  75: -0.2879237525 8.290009e-02 <list[1]> 2024-12-10 09:42:36        4
#>  76:  0.4322524469 1.868422e-01 <list[1]> 2024-12-10 09:42:36        4
#>  77:  0.0329207734 1.083777e-03 <list[1]> 2024-12-10 09:42:36        4
#>  78: -0.0611411480 3.738240e-03 <list[1]> 2024-12-10 09:42:36        4
#>  79: -0.0212242601 4.504692e-04 <list[1]> 2024-12-10 09:42:36        4
#>  80:  0.1131771516 1.280907e-02 <list[1]> 2024-12-10 09:42:36        4
#>  81:  0.1877085104 3.523448e-02 <list[1]> 2024-12-10 09:42:36        4
#>  82:  0.2244789808 5.039081e-02 <list[1]> 2024-12-10 09:42:36        4
#>  83: -0.2471450311 6.108067e-02 <list[1]> 2024-12-10 09:42:36        4
#>  84:  0.2815944087 7.929541e-02 <list[1]> 2024-12-10 09:42:36        4
#>  85: -0.4774730835 2.279805e-01 <list[1]> 2024-12-10 09:42:36        4
#>  86: -0.0458369533 2.101026e-03 <list[1]> 2024-12-10 09:42:36        4
#>  87: -0.3448624261 1.189301e-01 <list[1]> 2024-12-10 09:42:36        4
#>  88:  0.3193934158 1.020122e-01 <list[1]> 2024-12-10 09:42:36        4
#>  89:  0.1411337936 1.991875e-02 <list[1]> 2024-12-10 09:42:36        4
#>  90:  0.1738674154 3.022988e-02 <list[1]> 2024-12-10 09:42:36        4
#>  91: -0.1319433298 1.740904e-02 <list[1]> 2024-12-10 09:42:36        4
#>  92: -0.1259396662 1.586080e-02 <list[1]> 2024-12-10 09:42:36        4
#>  93:  0.0392718518 1.542278e-03 <list[1]> 2024-12-10 09:42:36        4
#>  94:  0.1816003303 3.297868e-02 <list[1]> 2024-12-10 09:42:36        4
#>  95:  0.1992232667 3.968991e-02 <list[1]> 2024-12-10 09:42:36        4
#>  96:  0.0685160731 4.694452e-03 <list[1]> 2024-12-10 09:42:36        4
#>  97: -0.2968106554 8.809657e-02 <list[1]> 2024-12-10 09:42:36        4
#>  98: -0.3724388923 1.387107e-01 <list[1]> 2024-12-10 09:42:36        4
#>  99:  0.0013052224 1.703605e-06 <list[1]> 2024-12-10 09:42:36        4
#> 100:  0.0571064147 3.261143e-03 <list[1]> 2024-12-10 09:42:36        4
#>                  x            y  x_domain           timestamp batch_nr
#>      .point_id
#>          <int>
#>   1:        NA
#>   2:        NA
#>   3:        NA
#>   4:        NA
#>   5:        NA
#>   6:        NA
#>   7:        NA
#>   8:        NA
#>   9:        NA
#>  10:        NA
#>  11:         1
#>  12:         1
#>  13:         1
#>  14:         1
#>  15:         1
#>  16:         1
#>  17:         1
#>  18:         1
#>  19:         1
#>  20:         1
#>  21:         2
#>  22:         2
#>  23:         2
#>  24:         2
#>  25:         2
#>  26:         2
#>  27:         2
#>  28:         2
#>  29:         2
#>  30:         2
#>  31:         3
#>  32:         3
#>  33:         3
#>  34:         3
#>  35:         3
#>  36:         3
#>  37:         3
#>  38:         3
#>  39:         3
#>  40:         3
#>  41:         1
#>  42:         1
#>  43:         1
#>  44:         1
#>  45:         1
#>  46:         1
#>  47:         1
#>  48:         1
#>  49:         1
#>  50:         1
#>  51:         2
#>  52:         2
#>  53:         2
#>  54:         2
#>  55:         2
#>  56:         2
#>  57:         2
#>  58:         2
#>  59:         2
#>  60:         2
#>  61:         3
#>  62:         3
#>  63:         3
#>  64:         3
#>  65:         3
#>  66:         3
#>  67:         3
#>  68:         3
#>  69:         3
#>  70:         3
#>  71:         1
#>  72:         1
#>  73:         1
#>  74:         1
#>  75:         1
#>  76:         1
#>  77:         1
#>  78:         1
#>  79:         1
#>  80:         1
#>  81:         2
#>  82:         2
#>  83:         2
#>  84:         2
#>  85:         2
#>  86:         2
#>  87:         2
#>  88:         2
#>  89:         2
#>  90:         2
#>  91:         3
#>  92:         3
#>  93:         3
#>  94:         3
#>  95:         3
#>  96:         3
#>  97:         3
#>  98:         3
#>  99:         3
#> 100:         3
#>      .point_id