Optimization via Non-linear Optimization

OptimizerNLoptr class that implements non-linear optimization. Calls nloptr::nloptr() from package nloptr.

Source

Johnson, G S (2020). “The NLopt nonlinear-optimization package.” https://github.com/stevengj/nlopt.

Parameters

algorithm

character(1)

eval_g_ineq

function()

xtol_rel

numeric(1)

xtol_abs

numeric(1)

ftol_rel

numeric(1)

ftol_abs

numeric(1)

start_values

character(1)
Create random start values or based on center of search space? In the latter case, it is the center of the parameters before a trafo is applied.

For the meaning of the control parameters, see nloptr::nloptr() and nloptr::nloptr.print.options().

The termination conditions stopval, maxtime and maxeval of nloptr::nloptr() are deactivated and replaced by the Terminator subclasses. The x and function value tolerance termination conditions (xtol_rel = 10^-4, xtol_abs = rep(0.0, length(x0)), ftol_rel = 0.0 and ftol_abs = 0.0) are still available and implemented with their package defaults. To deactivate these conditions, set them to -1.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerNLoptr

Methods

Public methods

OptimizerNLoptr$new()

OptimizerNLoptr$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerNLoptr$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerNLoptr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
if (requireNamespace("nloptr")) {

  search_space = domain = ps(x = p_dbl(lower = -1, upper = 1))

  codomain = ps(y = p_dbl(tags = "minimize"))

  objective_function = function(xs) {
    list(y = as.numeric(xs)^2)
  }

  objective = ObjectiveRFun$new(
    fun = objective_function,
    domain = domain,
    codomain = codomain)

  # We use the internal termination criterion xtol_rel
  terminator = trm("none")
  instance = OptimInstanceSingleCrit$new(
    objective = objective,
    search_space = search_space,
    terminator = terminator)

  optimizer = opt("nloptr", algorithm = "NLOPT_LN_BOBYQA")

  # Modifies the instance by reference
  optimizer$optimize(instance)

  # Returns best scoring evaluation
  instance$result

  # Allows access of data.table of full path of all evaluations
  as.data.table(instance$archive)
}
#>                 x            y           timestamp batch_nr    x_domain_x
#>  1: -5.746524e-01 3.302254e-01 2022-07-13 04:25:11        1 -5.746524e-01
#>  2: -5.746524e-01 3.302254e-01 2022-07-13 04:25:11        2 -5.746524e-01
#>  3: -5.746524e-01 3.302254e-01 2022-07-13 04:25:11        3 -5.746524e-01
#>  4: -2.556417e-01 6.535267e-02 2022-07-13 04:25:11        4 -2.556417e-01
#>  5: -8.936631e-01 7.986337e-01 2022-07-13 04:25:11        5 -8.936631e-01
#>  6: -1.110223e-16 1.232595e-32 2022-07-13 04:25:11        6 -1.110223e-16
#>  7:  3.190107e-02 1.017678e-03 2022-07-13 04:25:11        7  3.190107e-02
#>  8: -3.190107e-03 1.017678e-05 2022-07-13 04:25:11        8 -3.190107e-03
#>  9:  3.190107e-04 1.017678e-07 2022-07-13 04:25:11        9  3.190107e-04
#> 10: -3.190107e-04 1.017678e-07 2022-07-13 04:25:11       10 -3.190107e-04
#> 11: -7.954135e-21 6.326826e-41 2022-07-13 04:25:11       11 -7.954135e-21
# }