Optimization via Random Search

OptimizerRandomSearch class that implements a simple Random Search.

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("random_search")
opt("random_search")

Parameters

batch_size

integer(1)
Maximum number of points to try in a batch.

Super class

bbotk::Optimizer -> OptimizerRandomSearch

Methods

Public methods

OptimizerRandomSearch$new()

OptimizerRandomSearch$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(paradox)

domain = ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))

search_space = ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))

codomain = ParamSet$new(list(ParamDbl$new("y", tags = "minimize")))

objective_function = function(xs) {
  list(y = as.numeric(xs)^2)
}

objective = ObjectiveRFun$new(fun = objective_function,
                              domain = domain,
                              codomain = codomain)
terminator = term("evals", n_evals = 10)
instance = OptimInstanceSingleCrit$new(objective = objective,
                             search_space = search_space,
                             terminator = terminator)

optimizer = opt("random_search")

# Modifies the instance by reference
optimizer$optimize(instance)

# Returns best scoring evaluation
instance$result

#>             x x_domain          y
#> 1: -0.1929238   <list> 0.03721958

# Allows access of data.table of full path of all evaluations
instance$archive$data()

#>              x          y x_domain           timestamp batch_nr
#>  1: -0.1929238 0.03721958   <list> 2020-07-15 09:01:15        1
#>  2: -0.8726771 0.76156530   <list> 2020-07-15 09:01:15        2
#>  3: -0.2225974 0.04954959   <list> 2020-07-15 09:01:16        3
#>  4:  0.9510957 0.90458297   <list> 2020-07-15 09:01:16        4
#>  5: -0.4202154 0.17658099   <list> 2020-07-15 09:01:16        5
#>  6:  0.3567609 0.12727831   <list> 2020-07-15 09:01:16        6
#>  7:  0.4706392 0.22150125   <list> 2020-07-15 09:01:16        7
#>  8: -0.6080865 0.36976923   <list> 2020-07-15 09:01:16        8
#>  9:  0.9610793 0.92367352   <list> 2020-07-15 09:01:16        9
#> 10:  0.4830431 0.23333060   <list> 2020-07-15 09:01:16       10