Dictionary of Terminators

A simple mlr3misc::Dictionary storing objects of class Terminator. Each terminator has an associated help page, see mlr_terminators_[id].

This dictionary can get populated with additional terminators by add-on packages.

For a more convenient way to retrieve and construct terminator, see term()/terms().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

See also

Examples

term("evals", n_evals = 10)

#> <TerminatorEvals>
#> * Parameters: n_evals=10