Clock Time Terminator

Class to terminate the optimization after a fixed time point has been reached (as reported by Sys.time()).

Dictionary

This Terminator can be instantiated via the dictionary mlr_terminators or with the associated sugar function trm():

mlr_terminators$get("clock_time")
trm("clock_time")

Parameters

stop_time

POSIXct(1)
Terminator stops after this point in time.

See also

Super class

bbotk::Terminator -> TerminatorClockTime

Methods

Public methods

TerminatorClockTime$new()

TerminatorClockTime$is_terminated()

TerminatorClockTime$clone()

Inherited methods

bbotk::Terminator$format()

bbotk::Terminator$print()

bbotk::Terminator$remaining_time()

bbotk::Terminator$status()

Method new()

Creates a new instance of this R6 class.

Usage

TerminatorClockTime$new()

Method is_terminated()

Is TRUE iff the termination criterion is positive, and FALSE otherwise.

Usage

TerminatorClockTime$is_terminated(archive)

Arguments

archive

(Archive).

Returns

logical(1).

Method clone()

The objects of this class are cloneable with this method.

Usage

TerminatorClockTime$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

stop_time = as.POSIXct("2030-01-01 00:00:00")
trm("clock_time", stop_time = stop_time)
#> <TerminatorClockTime>: Clock Time
#> * Parameters: stop_time=<POSIXct>