Calculates the transformed x-values

Transforms a given data.table to a list with transformed x values. If no trafo is defined it will just convert the data.table to a list. Mainly for internal usage.

transform_xdt_to_xss(xdt, search_space)

Arguments

xdt
(data.table::data.table())
Set of untransformed points / points from the search space. One point per row, e.g. data.table(x1 = c(1, 3), x2 = c(2, 4)). Column names have to match ids of the search_space. However, xdt can contain additional columns.

search_space
(paradox::ParamSet)
Specifies the search space for the Optimizer. The paradox::ParamSet describes either a subset of the domain of the Objective or it describes a set of parameters together with a trafo function that transforms values from the search space to values of the domain. Depending on the context, this value defaults to the domain of the objective.

Value

list()