Objective Test Function

An ObjectiveRFun subclass for well-known optimization test functions. Adds optimum and optimum_x fields with the known global optimum.

Super classes

bbotk::Objective -> bbotk::ObjectiveRFun -> ObjectiveTestFunction

Public fields

optimum

(numeric(1))
Known global optimum value (f*).

optimum_x

(list())
List of known global optima, each a named list of input values.

Methods

Public methods

ObjectiveTestFunction$new()

ObjectiveTestFunction$clone()

Inherited methods

bbotk::Objective$eval_dt()

bbotk::Objective$eval_many()

bbotk::Objective$format()

bbotk::Objective$help()

bbotk::Objective$print()

bbotk::ObjectiveRFun$eval()

Method new()

Creates a new instance of this R6 class.

Usage

ObjectiveTestFunction$new(
  fun,
  domain,
  codomain = NULL,
  id,
  label,
  optimum,
  optimum_x,
  constants = ps()
)

Arguments

fun

(function)
Objective function function(xs).

domain

(paradox::ParamSet)
Specifies domain of function. The paradox::ParamSet should describe all possible input parameters of the objective function. This includes their id, their types and the possible range.

codomain

(paradox::ParamSet)
Specifies codomain of function. Most importantly the tags of each output "Parameter" define whether it should be minimized or maximized. The default is to minimize each component.

id

(character(1)).

label

(character(1)).

optimum

(numeric(1))
Known global optimum value.

optimum_x

(list())
List of known global optima.

constants

(paradox::ParamSet)
Changeable constants or parameters that are not subject to tuning can be stored and accessed here.

Method clone()

The objects of this class are cloneable with this method.

Usage

ObjectiveTestFunction$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.