Optimizer

Abstract Optimizer class that implements the base functionality each Optimizer subclass must provide. A Optimizer object describes the optimization strategy.

A Optimizer object must write its result to the $assign_result() method of the OptimInstance at the end in order to store the best point and its estimated performance vector.

Usage,NULL

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Public fields

param_classes

(character()).

properties

(character()).

packages

(character()).

Active bindings

param_set

(paradox::ParamSet).

Methods

Public methods

Optimizer$new()

Optimizer$format()

Optimizer$print()

Optimizer$optimize()

Optimizer$clone()

Method new()

Creates a new instance of this R6 class.

Usage

Optimizer$new(param_set, param_classes, properties, packages = character())

Arguments

param_set

(paradox::ParamSet).

param_classes

(character()).

properties

(character()).

packages

(character()).

Method format()

Helper for print outputs.

Usage

Optimizer$format()

Method print()

Print method.

Usage

Optimizer$print()

Returns

(character()).

Method optimize()

Performs the optimization and writes optimization result into OptimInstance. The optimization result is returned but the complete optimization path is stored in Archive of OptimInstance.

Usage

Optimizer$optimize(inst)

Arguments

inst

(OptimInstance).

Returns

data.table::data.table.

Method clone()

The objects of this class are cloneable with this method.

Usage

Optimizer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.