Optimization via Non-linear Optimization

OptimizerNLoptr class that implements non-linear optimization. Calls nloptr::nloptr() from package nloptr.

Source

Johnson SG (2020). “The NLopt nonlinear-optimization package.” https://github.com/stevengj/nlopt.

Parameters

algorithm

character(1)

x0

numeric()

eval_g_ineq

function()

xtol_rel

numeric(1)

xtol_abs

numeric(1)

ftol_rel

numeric(1)

ftol_abs

numeric(1)

For the meaning of the control parameters, see nloptr::nloptr() and nloptr::nloptr.print.options().

The termination conditions stopval, maxtime and maxeval of nloptr::nloptr() are deactivated and replaced by the Terminator subclasses. The x and function value tolerance termination conditions (xtol_rel = 10^-4, xtol_abs = rep(0.0, length(x0)), ftol_rel = 0.0 and ftol_abs = 0.0) are still available and implemented with their package defaults. To deactivate these conditions, set them to -1.

Super class

bbotk::Optimizer -> OptimizerNLoptr

Methods

Public methods

OptimizerNLoptr$new()

OptimizerNLoptr$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerNLoptr$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerNLoptr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
library(paradox)
library(data.table)

domain = ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))

search_space = ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))

codomain = ParamSet$new(list(ParamDbl$new("y", tags = "minimize")))

objective_function = function(xs) {
  list(y = as.numeric(xs)^2)
}

objective = ObjectiveRFun$new(fun = objective_function,
  domain = domain,
  codomain = codomain)

# We use the internal termination criterion xtol_rel
terminator = trm("none")
instance = OptimInstanceSingleCrit$new(objective = objective,
  search_space = search_space,
  terminator = terminator)

optimizer = opt("nloptr", x0 = 1, algorithm = "NLOPT_LN_BOBYQA")

# Modifies the instance by reference
optimizer$optimize(instance)

#>    x  x_domain y
#> 1: 0 <list[1]> 0

# Returns best scoring evaluation
instance$result

#>    x  x_domain y
#> 1: 0 <list[1]> 0

# Allows access of data.table of full path of all evaluations
instance$archive$data()

#>          x       y  x_domain           timestamp batch_nr
#>  1:  1e+00 1.0e+00 <list[1]> 2020-09-22 04:19:28        1
#>  2:  1e+00 1.0e+00 <list[1]> 2020-09-22 04:19:28        2
#>  3:  1e+00 1.0e+00 <list[1]> 2020-09-22 04:19:28        3
#>  4:  5e-01 2.5e-01 <list[1]> 2020-09-22 04:19:28        4
#>  5:  0e+00 0.0e+00 <list[1]> 2020-09-22 04:19:28        5
#>  6: -5e-02 2.5e-03 <list[1]> 2020-09-22 04:19:28        6
#>  7:  5e-03 2.5e-05 <list[1]> 2020-09-22 04:19:28        7
#>  8: -5e-03 2.5e-05 <list[1]> 2020-09-22 04:19:28        8
#>  9: -5e-04 2.5e-07 <list[1]> 2020-09-22 04:19:29        9
#> 10:  5e-04 2.5e-07 <list[1]> 2020-09-22 04:19:29       10
#> 11:  0e+00 0.0e+00 <list[1]> 2020-09-22 04:19:29       11

# }