Black-Box Optimization

This function optimizes a function or Objective with a given method.

Usage

bb_optimize(
  x,
  method = "random_search",
  max_evals = 1000,
  max_time = NULL,
  ...
)

# S3 method for `function`
bb_optimize(
  x,
  method = "random_search",
  max_evals = 1000,
  max_time = NULL,
  lower = NULL,
  upper = NULL,
  maximize = FALSE,
  ...
)

# S3 method for Objective
bb_optimize(
  x,
  method = "random_search",
  max_evals = 1000,
  max_time = NULL,
  search_space = NULL,
  ...
)

Arguments

x

(function | Objective).

method

(character(1) | Optimizer)
Key to retrieve optimizer from mlr_optimizers dictionary or Optimizer.

max_evals

(integer(1))
Number of allowed evaluations.

max_time

(integer(1))
Maximum allowed time in seconds.

...

(named list())
Named arguments passed to objective function. Ignored if Objective is optimized.

lower

(numeric())
Lower bounds on the parameters. If named, names are used to create the domain.

upper

(numeric())
Upper bounds on the parameters.

maximize

(logical())
Logical vector used to create the codomain e.g. c(TRUE, FALSE) -> ps(y1 = p_dbl(tags = "maximize"), y2 = pd_dbl(tags = "minimize")). If named, names are used to create the codomain.

search_space

(paradox::ParamSet).

Value

list of

"par" - Best found parameters

"value" - Optimal outcome

"instance" - OptimInstanceSingleCrit | OptimInstanceMultiCrit

Note

If both max_evals and max_time are NULL, TerminatorNone is used. This is useful if the Optimizer can terminate itself. If both are given, TerminatorCombo is created and the optimization stops if the time or evaluation budget is exhausted.

Examples

# function and bounds
fun = function(xs) {
  -(xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10
}

bb_optimize(fun, lower = c(-10, -5), upper = c(10, 5), max_evals = 10)
#> $par
#>           x1      x2
#> 1: -6.178272 3.26751
#> 
#> $value
#>        y1 
#> -96.16581 
#> 
#> $instance
#> <OptimInstanceSingleCrit>
#> * State:  Optimized
#> * Objective: <ObjectiveRFun:function>
#> * Search Space:
#>    id    class lower upper nlevels
#> 1: x1 ParamDbl   -10    10     Inf
#> 2: x2 ParamDbl    -5     5     Inf
#> * Terminator: <TerminatorEvals>
#> * Result:
#>           x1      x2        y1
#> 1: -6.178272 3.26751 -96.16581
#> * Archive:
#>            x1         x2         y1
#>  1:  2.285268 -3.4037107   9.755640
#>  2: -3.068733  0.7075302 -29.437837
#>  3:  4.668006  1.5650215 -17.957678
#>  4:  8.628076 -2.3963046 -34.295842
#>  5: -1.311474 -3.2999581  -1.055838
#>  6:  4.893156  1.0370215 -14.667896
#>  7: -4.574646 -3.0385073 -33.227458
#>  8:  7.962907  3.8502146 -72.481694
#>  9: -6.178272  3.2675097 -96.165809
#> 10: -1.085812  0.1954979  -9.733445
#> 

# function and constant
fun = function(xs, c) {
  -(xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + c
}

bb_optimize(fun, lower = c(-10, -5), upper = c(10, 5), max_evals = 10, c = 1)
#> $par
#>           x1         x2
#> 1: -9.455946 -0.4641841
#> 
#> $value
#>        y1 
#> -136.6691 
#> 
#> $instance
#> <OptimInstanceSingleCrit>
#> * State:  Optimized
#> * Objective: <ObjectiveRFun:function>
#> * Search Space:
#>    id    class lower upper nlevels
#> 1: x1 ParamDbl   -10    10     Inf
#> 2: x2 ParamDbl    -5     5     Inf
#> * Terminator: <TerminatorEvals>
#> * Result:
#>           x1         x2        y1
#> 1: -9.455946 -0.4641841 -136.6691
#> * Archive:
#>            x1         x2           y1
#>  1: -1.459599  2.7321632  -43.8265225
#>  2:  8.236493 -0.5511973  -43.8904791
#>  3:  8.859440  3.6819914  -90.7009225
#>  4:  9.201448  3.3342761  -90.9839052
#>  5: -9.455946 -0.4641841 -136.6690502
#>  6: -6.209353  1.0615726  -82.8898440
#>  7: -1.143101 -4.2967733  -10.5607066
#>  8:  2.331683 -1.7763687   -0.6072869
#>  9:  5.735986 -1.0702924  -16.6813655
#> 10:  5.118971 -0.1115151  -17.0713270
#> 

# objective
fun = function(xs) {
  c(z = -(xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10)
}

# define domain and codomain using a `ParamSet` from paradox
domain = ps(x1 = p_dbl(-10, 10), x2 = p_dbl(-5, 5))
codomain = ps(z = p_dbl(tags = "minimize"))
objective = ObjectiveRFun$new(fun, domain, codomain)

bb_optimize(objective, method = "random_search", max_evals = 10)
#> $par
#>           x1       x2
#> 1: -7.207211 3.293081
#> 
#> $value
#>         z 
#> -114.3756 
#> 
#> $instance
#> <OptimInstanceSingleCrit>
#> * State:  Optimized
#> * Objective: <ObjectiveRFun:function>
#> * Search Space:
#>    id    class lower upper nlevels
#> 1: x1 ParamDbl   -10    10     Inf
#> 2: x2 ParamDbl    -5     5     Inf
#> * Terminator: <TerminatorEvals>
#> * Result:
#>           x1       x2         z
#> 1: -7.207211 3.293081 -114.3756
#> * Archive:
#>             x1         x2           z
#>  1:  0.1257543 -3.8089892    5.832739
#>  2: -7.2072112  3.2930810 -114.375605
#>  3: -6.0428111 -0.6208643  -60.347097
#>  4:  8.9173393 -2.1185089  -38.626609
#>  5:  1.7682997  4.3189305  -43.620428
#>  6:  7.5627504  3.9499775  -69.246379
#>  7:  1.5303726  1.1926770   -7.799090
#>  8:  4.8942442  2.0955901  -24.341688
#>  9:  5.3661422  2.6798632  -33.591759
#> 10:  6.7937455 -1.4658321  -15.333667
#>