Asynchronous Optimization via Grid Search

OptimizerAsyncGridSearch class that implements a grid search. The grid is constructed as a Cartesian product over discretized values per parameter, see paradox::generate_design_grid(). The points of the grid are evaluated in a random order.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281–305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("async_grid_search")
opt("async_grid_search")

Parameters

batch_size

integer(1)
Maximum number of points to try in a batch.

Super classes

bbotk::Optimizer -> bbotk::OptimizerAsync -> OptimizerAsyncGridSearch

Methods

Public methods

OptimizerAsyncGridSearch$new()

OptimizerAsyncGridSearch$optimize()

OptimizerAsyncGridSearch$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAsyncGridSearch$new()

Method optimize()

Starts the asynchronous optimization.

Usage

OptimizerAsyncGridSearch$optimize(inst)

Arguments

inst

(OptimInstance).

Returns

data.table::data.table.

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAsyncGridSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# example only runs if a Redis server is available
if (mlr3misc::require_namespaces(c("rush", "redux", "mirai"), quietly = TRUE) &&
  redux::redis_available()) {
# define the objective function
fun = function(xs) {
  list(y = - (xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10)
}

# set domain
domain = ps(
  x1 = p_dbl(-10, 10),
  x2 = p_dbl(-5, 5)
)

# set codomain
codomain = ps(
  y = p_dbl(tags = "maximize")
)

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = codomain,
  properties = "deterministic"
)

# start workers
rush::rush_plan(worker_type = "remote")
mirai::daemons(1)

# initialize instance
instance = oi_async(
  objective = objective,
  terminator = trm("evals", n_evals = 20)
)

# load optimizer
optimizer = opt("async_grid_search", resolution = 10)

# trigger optimization
optimizer$optimize(instance)

# all evaluated configurations
instance$archive

# best performing configuration
instance$archive$best()

# covert to data.table
as.data.table(instance$archive)
}
#>         state         x1         x2          y        timestamp_xs   pid
#>        <char>      <num>      <num>      <num>              <POSc> <int>
#>   1: finished -10.000000 -5.0000000 -138.00000 2025-10-24 13:25:15  9655
#>   2: finished -10.000000 -3.8888889 -134.79012 2025-10-24 13:25:15  9655
#>   3: finished -10.000000 -2.7777778 -134.04938 2025-10-24 13:25:15  9655
#>   4: finished -10.000000 -1.6666667 -135.77778 2025-10-24 13:25:15  9655
#>   5: finished -10.000000 -0.5555556 -139.97531 2025-10-24 13:25:15  9655
#>   6: finished -10.000000  0.5555556 -146.64198 2025-10-24 13:25:15  9655
#>   7: finished -10.000000  1.6666667 -155.77778 2025-10-24 13:25:15  9655
#>   8: finished -10.000000  2.7777778 -167.38272 2025-10-24 13:25:15  9655
#>   9: finished -10.000000  3.8888889 -181.45679 2025-10-24 13:25:15  9655
#>  10: finished -10.000000  5.0000000 -198.00000 2025-10-24 13:25:15  9655
#>  11: finished  -7.777778 -5.0000000  -89.60494 2025-10-24 13:25:15  9655
#>  12: finished  -7.777778 -3.8888889  -86.39506 2025-10-24 13:25:15  9655
#>  13: finished  -7.777778 -2.7777778  -85.65432 2025-10-24 13:25:15  9655
#>  14: finished  -7.777778 -1.6666667  -87.38272 2025-10-24 13:25:15  9655
#>  15: finished  -7.777778 -0.5555556  -91.58025 2025-10-24 13:25:15  9655
#>  16: finished  -7.777778  0.5555556  -98.24691 2025-10-24 13:25:15  9655
#>  17: finished  -7.777778  1.6666667 -107.38272 2025-10-24 13:25:15  9655
#>  18: finished  -7.777778  2.7777778 -118.98765 2025-10-24 13:25:15  9655
#>  19: finished  -7.777778  3.8888889 -133.06173 2025-10-24 13:25:15  9655
#>  20: finished  -7.777778  5.0000000 -149.60494 2025-10-24 13:25:15  9655
#>  21:   failed  10.000000  5.0000000         NA 2025-10-24 13:25:15    NA
#>  22:   failed  10.000000  3.8888889         NA 2025-10-24 13:25:15    NA
#>  23:   failed  10.000000  2.7777778         NA 2025-10-24 13:25:15    NA
#>  24:   failed  10.000000  1.6666667         NA 2025-10-24 13:25:15    NA
#>  25:   failed  10.000000  0.5555556         NA 2025-10-24 13:25:15    NA
#>  26:   failed  10.000000 -0.5555556         NA 2025-10-24 13:25:15    NA
#>  27:   failed  10.000000 -1.6666667         NA 2025-10-24 13:25:15    NA
#>  28:   failed  10.000000 -2.7777778         NA 2025-10-24 13:25:15    NA
#>  29:   failed  10.000000 -3.8888889         NA 2025-10-24 13:25:15    NA
#>  30:   failed  10.000000 -5.0000000         NA 2025-10-24 13:25:15    NA
#>  31:   failed   7.777778  5.0000000         NA 2025-10-24 13:25:15    NA
#>  32:   failed   7.777778  3.8888889         NA 2025-10-24 13:25:15    NA
#>  33:   failed   7.777778  2.7777778         NA 2025-10-24 13:25:15    NA
#>  34:   failed   7.777778  1.6666667         NA 2025-10-24 13:25:15    NA
#>  35:   failed   7.777778  0.5555556         NA 2025-10-24 13:25:15    NA
#>  36:   failed   7.777778 -0.5555556         NA 2025-10-24 13:25:15    NA
#>  37:   failed   7.777778 -1.6666667         NA 2025-10-24 13:25:15    NA
#>  38:   failed   7.777778 -2.7777778         NA 2025-10-24 13:25:15    NA
#>  39:   failed   7.777778 -3.8888889         NA 2025-10-24 13:25:15    NA
#>  40:   failed   7.777778 -5.0000000         NA 2025-10-24 13:25:15    NA
#>  41:   failed   5.555556  5.0000000         NA 2025-10-24 13:25:15    NA
#>  42:   failed   5.555556  3.8888889         NA 2025-10-24 13:25:15    NA
#>  43:   failed   5.555556  2.7777778         NA 2025-10-24 13:25:15    NA
#>  44:   failed   5.555556  1.6666667         NA 2025-10-24 13:25:15    NA
#>  45:   failed   5.555556  0.5555556         NA 2025-10-24 13:25:15    NA
#>  46:   failed   5.555556 -0.5555556         NA 2025-10-24 13:25:15    NA
#>  47:   failed   5.555556 -1.6666667         NA 2025-10-24 13:25:15    NA
#>  48:   failed   5.555556 -2.7777778         NA 2025-10-24 13:25:15    NA
#>  49:   failed   5.555556 -3.8888889         NA 2025-10-24 13:25:15    NA
#>  50:   failed   5.555556 -5.0000000         NA 2025-10-24 13:25:15    NA
#>  51:   failed   3.333333  5.0000000         NA 2025-10-24 13:25:15    NA
#>  52:   failed   3.333333  3.8888889         NA 2025-10-24 13:25:15    NA
#>  53:   failed   3.333333  2.7777778         NA 2025-10-24 13:25:15    NA
#>  54:   failed   3.333333  1.6666667         NA 2025-10-24 13:25:15    NA
#>  55:   failed   3.333333  0.5555556         NA 2025-10-24 13:25:15    NA
#>  56:   failed   3.333333 -0.5555556         NA 2025-10-24 13:25:15    NA
#>  57:   failed   3.333333 -1.6666667         NA 2025-10-24 13:25:15    NA
#>  58:   failed   3.333333 -2.7777778         NA 2025-10-24 13:25:15    NA
#>  59:   failed   3.333333 -3.8888889         NA 2025-10-24 13:25:15    NA
#>  60:   failed   3.333333 -5.0000000         NA 2025-10-24 13:25:15    NA
#>  61:   failed   1.111111  5.0000000         NA 2025-10-24 13:25:15    NA
#>  62:   failed   1.111111  3.8888889         NA 2025-10-24 13:25:15    NA
#>  63:   failed   1.111111  2.7777778         NA 2025-10-24 13:25:15    NA
#>  64:   failed   1.111111  1.6666667         NA 2025-10-24 13:25:15    NA
#>  65:   failed   1.111111  0.5555556         NA 2025-10-24 13:25:15    NA
#>  66:   failed   1.111111 -0.5555556         NA 2025-10-24 13:25:15    NA
#>  67:   failed   1.111111 -1.6666667         NA 2025-10-24 13:25:15    NA
#>  68:   failed   1.111111 -2.7777778         NA 2025-10-24 13:25:15    NA
#>  69:   failed   1.111111 -3.8888889         NA 2025-10-24 13:25:15    NA
#>  70:   failed   1.111111 -5.0000000         NA 2025-10-24 13:25:15    NA
#>  71:   failed  -1.111111  5.0000000         NA 2025-10-24 13:25:15    NA
#>  72:   failed  -1.111111  3.8888889         NA 2025-10-24 13:25:15    NA
#>  73:   failed  -1.111111  2.7777778         NA 2025-10-24 13:25:15    NA
#>  74:   failed  -1.111111  1.6666667         NA 2025-10-24 13:25:15    NA
#>  75:   failed  -1.111111  0.5555556         NA 2025-10-24 13:25:15    NA
#>  76:   failed  -1.111111 -0.5555556         NA 2025-10-24 13:25:15    NA
#>  77:   failed  -1.111111 -1.6666667         NA 2025-10-24 13:25:15    NA
#>  78:   failed  -1.111111 -2.7777778         NA 2025-10-24 13:25:15    NA
#>  79:   failed  -1.111111 -3.8888889         NA 2025-10-24 13:25:15    NA
#>  80:   failed  -1.111111 -5.0000000         NA 2025-10-24 13:25:15    NA
#>  81:   failed  -3.333333  5.0000000         NA 2025-10-24 13:25:15    NA
#>  82:   failed  -3.333333  3.8888889         NA 2025-10-24 13:25:15    NA
#>  83:   failed  -3.333333  2.7777778         NA 2025-10-24 13:25:15    NA
#>  84:   failed  -3.333333  1.6666667         NA 2025-10-24 13:25:15    NA
#>  85:   failed  -3.333333  0.5555556         NA 2025-10-24 13:25:15    NA
#>  86:   failed  -3.333333 -0.5555556         NA 2025-10-24 13:25:15    NA
#>  87:   failed  -3.333333 -1.6666667         NA 2025-10-24 13:25:15    NA
#>  88:   failed  -3.333333 -2.7777778         NA 2025-10-24 13:25:15    NA
#>  89:   failed  -3.333333 -3.8888889         NA 2025-10-24 13:25:15    NA
#>  90:   failed  -3.333333 -5.0000000         NA 2025-10-24 13:25:15    NA
#>  91:   failed  -5.555556  5.0000000         NA 2025-10-24 13:25:15    NA
#>  92:   failed  -5.555556  3.8888889         NA 2025-10-24 13:25:15    NA
#>  93:   failed  -5.555556  2.7777778         NA 2025-10-24 13:25:15    NA
#>  94:   failed  -5.555556  1.6666667         NA 2025-10-24 13:25:15    NA
#>  95:   failed  -5.555556  0.5555556         NA 2025-10-24 13:25:15    NA
#>  96:   failed  -5.555556 -0.5555556         NA 2025-10-24 13:25:15    NA
#>  97:   failed  -5.555556 -1.6666667         NA 2025-10-24 13:25:15    NA
#>  98:   failed  -5.555556 -2.7777778         NA 2025-10-24 13:25:15    NA
#>  99:   failed  -5.555556 -3.8888889         NA 2025-10-24 13:25:15    NA
#> 100:   failed  -5.555556 -5.0000000         NA 2025-10-24 13:25:15    NA
#>         state         x1         x2          y        timestamp_xs   pid
#>                 worker_id        timestamp_ys
#>                    <char>              <POSc>
#>   1: academical_guineapig 2025-10-24 13:25:16
#>   2: academical_guineapig 2025-10-24 13:25:16
#>   3: academical_guineapig 2025-10-24 13:25:16
#>   4: academical_guineapig 2025-10-24 13:25:16
#>   5: academical_guineapig 2025-10-24 13:25:16
#>   6: academical_guineapig 2025-10-24 13:25:16
#>   7: academical_guineapig 2025-10-24 13:25:16
#>   8: academical_guineapig 2025-10-24 13:25:16
#>   9: academical_guineapig 2025-10-24 13:25:16
#>  10: academical_guineapig 2025-10-24 13:25:16
#>  11: academical_guineapig 2025-10-24 13:25:16
#>  12: academical_guineapig 2025-10-24 13:25:16
#>  13: academical_guineapig 2025-10-24 13:25:16
#>  14: academical_guineapig 2025-10-24 13:25:16
#>  15: academical_guineapig 2025-10-24 13:25:16
#>  16: academical_guineapig 2025-10-24 13:25:16
#>  17: academical_guineapig 2025-10-24 13:25:16
#>  18: academical_guineapig 2025-10-24 13:25:16
#>  19: academical_guineapig 2025-10-24 13:25:16
#>  20: academical_guineapig 2025-10-24 13:25:16
#>  21:                 <NA>                <NA>
#>  22:                 <NA>                <NA>
#>  23:                 <NA>                <NA>
#>  24:                 <NA>                <NA>
#>  25:                 <NA>                <NA>
#>  26:                 <NA>                <NA>
#>  27:                 <NA>                <NA>
#>  28:                 <NA>                <NA>
#>  29:                 <NA>                <NA>
#>  30:                 <NA>                <NA>
#>  31:                 <NA>                <NA>
#>  32:                 <NA>                <NA>
#>  33:                 <NA>                <NA>
#>  34:                 <NA>                <NA>
#>  35:                 <NA>                <NA>
#>  36:                 <NA>                <NA>
#>  37:                 <NA>                <NA>
#>  38:                 <NA>                <NA>
#>  39:                 <NA>                <NA>
#>  40:                 <NA>                <NA>
#>  41:                 <NA>                <NA>
#>  42:                 <NA>                <NA>
#>  43:                 <NA>                <NA>
#>  44:                 <NA>                <NA>
#>  45:                 <NA>                <NA>
#>  46:                 <NA>                <NA>
#>  47:                 <NA>                <NA>
#>  48:                 <NA>                <NA>
#>  49:                 <NA>                <NA>
#>  50:                 <NA>                <NA>
#>  51:                 <NA>                <NA>
#>  52:                 <NA>                <NA>
#>  53:                 <NA>                <NA>
#>  54:                 <NA>                <NA>
#>  55:                 <NA>                <NA>
#>  56:                 <NA>                <NA>
#>  57:                 <NA>                <NA>
#>  58:                 <NA>                <NA>
#>  59:                 <NA>                <NA>
#>  60:                 <NA>                <NA>
#>  61:                 <NA>                <NA>
#>  62:                 <NA>                <NA>
#>  63:                 <NA>                <NA>
#>  64:                 <NA>                <NA>
#>  65:                 <NA>                <NA>
#>  66:                 <NA>                <NA>
#>  67:                 <NA>                <NA>
#>  68:                 <NA>                <NA>
#>  69:                 <NA>                <NA>
#>  70:                 <NA>                <NA>
#>  71:                 <NA>                <NA>
#>  72:                 <NA>                <NA>
#>  73:                 <NA>                <NA>
#>  74:                 <NA>                <NA>
#>  75:                 <NA>                <NA>
#>  76:                 <NA>                <NA>
#>  77:                 <NA>                <NA>
#>  78:                 <NA>                <NA>
#>  79:                 <NA>                <NA>
#>  80:                 <NA>                <NA>
#>  81:                 <NA>                <NA>
#>  82:                 <NA>                <NA>
#>  83:                 <NA>                <NA>
#>  84:                 <NA>                <NA>
#>  85:                 <NA>                <NA>
#>  86:                 <NA>                <NA>
#>  87:                 <NA>                <NA>
#>  88:                 <NA>                <NA>
#>  89:                 <NA>                <NA>
#>  90:                 <NA>                <NA>
#>  91:                 <NA>                <NA>
#>  92:                 <NA>                <NA>
#>  93:                 <NA>                <NA>
#>  94:                 <NA>                <NA>
#>  95:                 <NA>                <NA>
#>  96:                 <NA>                <NA>
#>  97:                 <NA>                <NA>
#>  98:                 <NA>                <NA>
#>  99:                 <NA>                <NA>
#> 100:                 <NA>                <NA>
#>                 worker_id        timestamp_ys
#>                                      keys                 message x_domain_x1
#>                                    <char>                  <char>       <num>
#>   1: 5d2fca32-f24e-42e4-85e8-c5a8ecd92609                    <NA>  -10.000000
#>   2: 5e6de97c-9eba-467e-8567-6afe8e577275                    <NA>  -10.000000
#>   3: def8511d-551e-41df-a71b-6c4dca5dc63c                    <NA>  -10.000000
#>   4: 2f42140e-64ee-428a-9121-169890aada79                    <NA>  -10.000000
#>   5: f6342d27-4c38-4ab3-be20-bb99f9e595db                    <NA>  -10.000000
#>   6: 1dd89183-50ae-443a-9821-2efc17511184                    <NA>  -10.000000
#>   7: 19b4f8b4-a935-41fc-84e7-af9748fd240b                    <NA>  -10.000000
#>   8: 83f753bb-93d8-40c1-b979-888420f6942a                    <NA>  -10.000000
#>   9: 20529871-8f4b-403c-918b-6fd7dbeca83f                    <NA>  -10.000000
#>  10: e39960dc-4508-4db4-b05c-a78f5e3fcc2f                    <NA>  -10.000000
#>  11: 8e888e7c-5f53-4c4c-8054-460f99ed4fb5                    <NA>   -7.777778
#>  12: 338d83c7-256f-4947-a845-a7e492e08b81                    <NA>   -7.777778
#>  13: a44ecc4e-0ef1-47b8-a48d-9dcd6ef1163e                    <NA>   -7.777778
#>  14: 72a1cc04-ac48-4937-9d08-44c90002e830                    <NA>   -7.777778
#>  15: 30d23c81-3698-4567-ab5e-3da4a91c7315                    <NA>   -7.777778
#>  16: 48c9b90e-8bef-4cbd-9e9f-687f5bef8f05                    <NA>   -7.777778
#>  17: 1b8b4d40-96c6-4387-99c2-92a485517b47                    <NA>   -7.777778
#>  18: 12bb685f-fc12-4427-aa8c-fba3d2bc5aad                    <NA>   -7.777778
#>  19: 96688dd5-a5b9-4aa4-835a-60552026810d                    <NA>   -7.777778
#>  20: 087b6831-ed1b-4f1e-9d06-929118061864                    <NA>   -7.777778
#>  21: b77e37cd-8e4d-45bb-a412-583a959a1c19 Optimization terminated          NA
#>  22: 53b6583c-624d-4d2b-b2f0-cf54b3238ad9 Optimization terminated          NA
#>  23: cfd954cc-51aa-4a43-beeb-8f3f48a523fb Optimization terminated          NA
#>  24: bcdba95d-e5d0-4afe-9e7b-8a89e4c9183b Optimization terminated          NA
#>  25: 5a854060-9ff4-4f64-bfe4-f5e1340d977a Optimization terminated          NA
#>  26: 87eadda3-689e-4adb-90a3-3e2b920dd764 Optimization terminated          NA
#>  27: daabf1bc-d905-495f-aade-b6a9d5e3b1b6 Optimization terminated          NA
#>  28: 4e38dca2-15c3-48a3-9190-71ff05d56322 Optimization terminated          NA
#>  29: d0d11e6d-70f9-46f0-93aa-14b0c92e3bff Optimization terminated          NA
#>  30: ea86a4a6-2bbc-48b7-abfa-ea3b532bd370 Optimization terminated          NA
#>  31: 5e16fdc0-bec8-4a43-af79-01b717ea926e Optimization terminated          NA
#>  32: 6e3a8724-5ff1-4112-841f-b0e67caabea8 Optimization terminated          NA
#>  33: 4c679d7a-a2e4-465b-915b-51b56050595c Optimization terminated          NA
#>  34: 2b11b299-8c8b-406a-8c1f-5bc8066e32bd Optimization terminated          NA
#>  35: 24cc5cdf-bef6-4940-b7ea-9d037e71e406 Optimization terminated          NA
#>  36: d6cf190e-84b6-440b-8633-57f5502cd5f2 Optimization terminated          NA
#>  37: f9a79416-efa4-4e50-a61e-912ee74a5714 Optimization terminated          NA
#>  38: d53aea30-07bc-461d-b073-aeafeb9deb7e Optimization terminated          NA
#>  39: 1fc8e4c6-c5df-41bd-8519-76aa96b0c15c Optimization terminated          NA
#>  40: a73a8561-a9cb-4def-8242-88d6561029ca Optimization terminated          NA
#>  41: 5a8780c6-7f47-4068-a770-c8c0736601dd Optimization terminated          NA
#>  42: c42d747a-a9a3-4a8c-a82a-6c577aa80106 Optimization terminated          NA
#>  43: 5e37546e-0062-406d-8206-7625f5253bea Optimization terminated          NA
#>  44: 568d2330-a6fe-4714-8aa9-1a5e53d3b5ca Optimization terminated          NA
#>  45: 5ade81fe-f502-4ae9-9faf-90b63bc51cec Optimization terminated          NA
#>  46: 17237da8-5b18-42da-946d-cb7413a71582 Optimization terminated          NA
#>  47: 82705bcd-d543-4833-859e-e0df4a1ec18a Optimization terminated          NA
#>  48: 60b7ffc1-1f30-48b2-bf22-f46ba3f964a4 Optimization terminated          NA
#>  49: f295f34f-917f-4193-ab30-ac6764469fc9 Optimization terminated          NA
#>  50: f18052e1-5f53-40d8-aa64-99f26be55659 Optimization terminated          NA
#>  51: 56b64b4d-d07c-4f76-9c5e-cc5027ab1111 Optimization terminated          NA
#>  52: 847f5dec-1780-466d-91be-cb98d63139b3 Optimization terminated          NA
#>  53: c46cd13a-7b24-4dd5-b165-b71f8be63b3f Optimization terminated          NA
#>  54: 722165d3-0d52-49bf-a782-cc264ddf3c4e Optimization terminated          NA
#>  55: f07ac0a3-f24c-40d9-a5f5-cbb228431946 Optimization terminated          NA
#>  56: 7a3e9056-d8b8-4c72-a075-958cff94850d Optimization terminated          NA
#>  57: 31843c76-fac7-45e2-becf-39fdab44005e Optimization terminated          NA
#>  58: 14f49793-69ec-4de9-bd94-1498cea7e8b6 Optimization terminated          NA
#>  59: c46b931c-30d0-4630-992d-d00fdaaedd11 Optimization terminated          NA
#>  60: 6a7a3798-0539-44b1-ab38-aecde56f1c97 Optimization terminated          NA
#>  61: 59cf6ec0-8433-4842-a362-640bacd8570c Optimization terminated          NA
#>  62: f27db27f-1d6d-4aff-bb85-679dba4b0612 Optimization terminated          NA
#>  63: a58f8a26-5a19-4af1-b92d-07982fcd7015 Optimization terminated          NA
#>  64: cc7150e7-63e2-4fca-b2c4-a019a6eab91e Optimization terminated          NA
#>  65: d156efec-dd44-4820-b4c3-86176fbbbc2a Optimization terminated          NA
#>  66: afda6d81-d8ec-4226-b636-ed48b9faa993 Optimization terminated          NA
#>  67: bc850f59-83c2-4ddb-8891-cfea8f6cf3c6 Optimization terminated          NA
#>  68: 3a313cce-8c14-4a8e-83eb-eced8ce3479a Optimization terminated          NA
#>  69: 770f5030-904b-44fc-a75a-a10b04b95cbc Optimization terminated          NA
#>  70: 994b0a41-a1f2-4eb3-8d2f-38b0a9d157f8 Optimization terminated          NA
#>  71: e27bd82b-7edb-4e87-9a9b-a46c57b37ac6 Optimization terminated          NA
#>  72: 91092a8e-843b-41ab-b758-71f270ea5019 Optimization terminated          NA
#>  73: cc8c02ac-5c3b-4838-96c6-ec71e0b48aa0 Optimization terminated          NA
#>  74: 8147f0c4-722d-4e77-9017-6bf9226d45b5 Optimization terminated          NA
#>  75: d012f795-3039-4795-99d5-76ac49cbcce4 Optimization terminated          NA
#>  76: c246a656-1860-4578-93b4-fe3548e33af8 Optimization terminated          NA
#>  77: 3ec7d017-07b0-447b-9566-a13026bf1b8c Optimization terminated          NA
#>  78: bff1aa69-2a0b-40fe-bb6a-b47434f24ac9 Optimization terminated          NA
#>  79: f0d1d2dd-245c-4d0c-afe8-d86b8c40c37e Optimization terminated          NA
#>  80: edfa6f34-873e-470a-9fe6-bd055e72a100 Optimization terminated          NA
#>  81: b2811eda-c58a-4015-80e7-a4f81fdffa4b Optimization terminated          NA
#>  82: 660052b4-5758-4f9b-8e61-a6fbe618649d Optimization terminated          NA
#>  83: e11d1a56-39f4-4ac8-9432-d0c0884a4f3e Optimization terminated          NA
#>  84: 9bb8fd2d-5d3b-42a9-b8c9-8d45db6efb84 Optimization terminated          NA
#>  85: 6f9d9e21-71da-411d-b8ff-4c4f01ff6d35 Optimization terminated          NA
#>  86: 17dc3d54-b240-45fc-a5e1-45c07295837c Optimization terminated          NA
#>  87: 4543e20e-d5f9-4607-8fea-3e414103c929 Optimization terminated          NA
#>  88: dbc294e5-ae6d-4598-bd39-228e35fe7991 Optimization terminated          NA
#>  89: 9183ceb2-9c1a-452d-9c19-ba75f5054d05 Optimization terminated          NA
#>  90: 61c86208-4939-49ee-906e-b2ed29968087 Optimization terminated          NA
#>  91: 912f5dd5-7d82-4998-8d05-f91eb03ffc87 Optimization terminated          NA
#>  92: 4a908057-ac27-4100-ad7f-ac0d80f352de Optimization terminated          NA
#>  93: 2c7e7e11-bcdf-4630-bc06-2dddcd7dcfac Optimization terminated          NA
#>  94: 40218e96-fe01-4739-aceb-1974e4a290c7 Optimization terminated          NA
#>  95: 7de6e243-6c71-4f4c-b702-7c7ee7ff5037 Optimization terminated          NA
#>  96: c7b6f85c-61ff-483a-849d-43107ee00bf0 Optimization terminated          NA
#>  97: 9875bd35-72d9-4ba4-88fc-d655b296a3d1 Optimization terminated          NA
#>  98: 9482785c-bb18-4ac9-896a-4ec440c647fc Optimization terminated          NA
#>  99: db227702-0c70-4406-807f-158fe50e4b84 Optimization terminated          NA
#> 100: 1779313e-8366-43f5-b440-b5e53c07d5ee Optimization terminated          NA
#>                                      keys                 message x_domain_x1
#>      x_domain_x2
#>            <num>
#>   1:  -5.0000000
#>   2:  -3.8888889
#>   3:  -2.7777778
#>   4:  -1.6666667
#>   5:  -0.5555556
#>   6:   0.5555556
#>   7:   1.6666667
#>   8:   2.7777778
#>   9:   3.8888889
#>  10:   5.0000000
#>  11:  -5.0000000
#>  12:  -3.8888889
#>  13:  -2.7777778
#>  14:  -1.6666667
#>  15:  -0.5555556
#>  16:   0.5555556
#>  17:   1.6666667
#>  18:   2.7777778
#>  19:   3.8888889
#>  20:   5.0000000
#>  21:          NA
#>  22:          NA
#>  23:          NA
#>  24:          NA
#>  25:          NA
#>  26:          NA
#>  27:          NA
#>  28:          NA
#>  29:          NA
#>  30:          NA
#>  31:          NA
#>  32:          NA
#>  33:          NA
#>  34:          NA
#>  35:          NA
#>  36:          NA
#>  37:          NA
#>  38:          NA
#>  39:          NA
#>  40:          NA
#>  41:          NA
#>  42:          NA
#>  43:          NA
#>  44:          NA
#>  45:          NA
#>  46:          NA
#>  47:          NA
#>  48:          NA
#>  49:          NA
#>  50:          NA
#>  51:          NA
#>  52:          NA
#>  53:          NA
#>  54:          NA
#>  55:          NA
#>  56:          NA
#>  57:          NA
#>  58:          NA
#>  59:          NA
#>  60:          NA
#>  61:          NA
#>  62:          NA
#>  63:          NA
#>  64:          NA
#>  65:          NA
#>  66:          NA
#>  67:          NA
#>  68:          NA
#>  69:          NA
#>  70:          NA
#>  71:          NA
#>  72:          NA
#>  73:          NA
#>  74:          NA
#>  75:          NA
#>  76:          NA
#>  77:          NA
#>  78:          NA
#>  79:          NA
#>  80:          NA
#>  81:          NA
#>  82:          NA
#>  83:          NA
#>  84:          NA
#>  85:          NA
#>  86:          NA
#>  87:          NA
#>  88:          NA
#>  89:          NA
#>  90:          NA
#>  91:          NA
#>  92:          NA
#>  93:          NA
#>  94:          NA
#>  95:          NA
#>  96:          NA
#>  97:          NA
#>  98:          NA
#>  99:          NA
#> 100:          NA
#>      x_domain_x2