Asynchronous Optimization via Grid Search

OptimizerAsyncGridSearch class that implements a grid search. The grid is constructed as a Cartesian product over discretized values per parameter, see paradox::generate_design_grid(). The points of the grid are evaluated in a random order.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281–305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("async_grid_search")
opt("async_grid_search")

Parameters

batch_size

integer(1)
Maximum number of points to try in a batch.

Super classes

bbotk::Optimizer -> bbotk::OptimizerAsync -> OptimizerAsyncGridSearch

Methods

Public methods

OptimizerAsyncGridSearch$new()

OptimizerAsyncGridSearch$optimize()

OptimizerAsyncGridSearch$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAsyncGridSearch$new()

Method optimize()

Starts the asynchronous optimization.

Usage

OptimizerAsyncGridSearch$optimize(inst)

Arguments

inst

(OptimInstance).

Returns

data.table::data.table.

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAsyncGridSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# example only runs if a Redis server is available
if (mlr3misc::require_namespaces(c("rush", "redux", "mirai"), quietly = TRUE) &&
  redux::redis_available()) {
# define the objective function
fun = function(xs) {
  list(y = - (xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10)
}

# set domain
domain = ps(
  x1 = p_dbl(-10, 10),
  x2 = p_dbl(-5, 5)
)

# set codomain
codomain = ps(
  y = p_dbl(tags = "maximize")
)

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = codomain,
  properties = "deterministic"
)

# start workers
rush::rush_plan(worker_type = "remote")
mirai::daemons(1)

# initialize instance
instance = oi_async(
  objective = objective,
  terminator = trm("evals", n_evals = 20)
)

# load optimizer
optimizer = opt("async_grid_search", resolution = 10)

# trigger optimization
optimizer$optimize(instance)

# all evaluated configurations
instance$archive

# best performing configuration
instance$archive$best()

# covert to data.table
as.data.table(instance$archive)
}
#>         state         x1         x2          y        timestamp_xs   pid
#>        <char>      <num>      <num>      <num>              <POSc> <int>
#>   1: finished -10.000000 -5.0000000 -138.00000 2025-10-24 13:28:55  9576
#>   2: finished -10.000000 -3.8888889 -134.79012 2025-10-24 13:28:55  9576
#>   3: finished -10.000000 -2.7777778 -134.04938 2025-10-24 13:28:55  9576
#>   4: finished -10.000000 -1.6666667 -135.77778 2025-10-24 13:28:55  9576
#>   5: finished -10.000000 -0.5555556 -139.97531 2025-10-24 13:28:55  9576
#>   6: finished -10.000000  0.5555556 -146.64198 2025-10-24 13:28:55  9576
#>   7: finished -10.000000  1.6666667 -155.77778 2025-10-24 13:28:55  9576
#>   8: finished -10.000000  2.7777778 -167.38272 2025-10-24 13:28:55  9576
#>   9: finished -10.000000  3.8888889 -181.45679 2025-10-24 13:28:55  9576
#>  10: finished -10.000000  5.0000000 -198.00000 2025-10-24 13:28:55  9576
#>  11: finished  -7.777778 -5.0000000  -89.60494 2025-10-24 13:28:55  9576
#>  12: finished  -7.777778 -3.8888889  -86.39506 2025-10-24 13:28:55  9576
#>  13: finished  -7.777778 -2.7777778  -85.65432 2025-10-24 13:28:55  9576
#>  14: finished  -7.777778 -1.6666667  -87.38272 2025-10-24 13:28:55  9576
#>  15: finished  -7.777778 -0.5555556  -91.58025 2025-10-24 13:28:55  9576
#>  16: finished  -7.777778  0.5555556  -98.24691 2025-10-24 13:28:55  9576
#>  17: finished  -7.777778  1.6666667 -107.38272 2025-10-24 13:28:55  9576
#>  18: finished  -7.777778  2.7777778 -118.98765 2025-10-24 13:28:55  9576
#>  19: finished  -7.777778  3.8888889 -133.06173 2025-10-24 13:28:55  9576
#>  20: finished  -7.777778  5.0000000 -149.60494 2025-10-24 13:28:55  9576
#>  21:   failed  10.000000  5.0000000         NA 2025-10-24 13:28:55    NA
#>  22:   failed  10.000000  3.8888889         NA 2025-10-24 13:28:55    NA
#>  23:   failed  10.000000  2.7777778         NA 2025-10-24 13:28:55    NA
#>  24:   failed  10.000000  1.6666667         NA 2025-10-24 13:28:55    NA
#>  25:   failed  10.000000  0.5555556         NA 2025-10-24 13:28:55    NA
#>  26:   failed  10.000000 -0.5555556         NA 2025-10-24 13:28:55    NA
#>  27:   failed  10.000000 -1.6666667         NA 2025-10-24 13:28:55    NA
#>  28:   failed  10.000000 -2.7777778         NA 2025-10-24 13:28:55    NA
#>  29:   failed  10.000000 -3.8888889         NA 2025-10-24 13:28:55    NA
#>  30:   failed  10.000000 -5.0000000         NA 2025-10-24 13:28:55    NA
#>  31:   failed   7.777778  5.0000000         NA 2025-10-24 13:28:55    NA
#>  32:   failed   7.777778  3.8888889         NA 2025-10-24 13:28:55    NA
#>  33:   failed   7.777778  2.7777778         NA 2025-10-24 13:28:55    NA
#>  34:   failed   7.777778  1.6666667         NA 2025-10-24 13:28:55    NA
#>  35:   failed   7.777778  0.5555556         NA 2025-10-24 13:28:55    NA
#>  36:   failed   7.777778 -0.5555556         NA 2025-10-24 13:28:55    NA
#>  37:   failed   7.777778 -1.6666667         NA 2025-10-24 13:28:55    NA
#>  38:   failed   7.777778 -2.7777778         NA 2025-10-24 13:28:55    NA
#>  39:   failed   7.777778 -3.8888889         NA 2025-10-24 13:28:55    NA
#>  40:   failed   7.777778 -5.0000000         NA 2025-10-24 13:28:55    NA
#>  41:   failed   5.555556  5.0000000         NA 2025-10-24 13:28:55    NA
#>  42:   failed   5.555556  3.8888889         NA 2025-10-24 13:28:55    NA
#>  43:   failed   5.555556  2.7777778         NA 2025-10-24 13:28:55    NA
#>  44:   failed   5.555556  1.6666667         NA 2025-10-24 13:28:55    NA
#>  45:   failed   5.555556  0.5555556         NA 2025-10-24 13:28:55    NA
#>  46:   failed   5.555556 -0.5555556         NA 2025-10-24 13:28:55    NA
#>  47:   failed   5.555556 -1.6666667         NA 2025-10-24 13:28:55    NA
#>  48:   failed   5.555556 -2.7777778         NA 2025-10-24 13:28:55    NA
#>  49:   failed   5.555556 -3.8888889         NA 2025-10-24 13:28:55    NA
#>  50:   failed   5.555556 -5.0000000         NA 2025-10-24 13:28:55    NA
#>  51:   failed   3.333333  5.0000000         NA 2025-10-24 13:28:55    NA
#>  52:   failed   3.333333  3.8888889         NA 2025-10-24 13:28:55    NA
#>  53:   failed   3.333333  2.7777778         NA 2025-10-24 13:28:55    NA
#>  54:   failed   3.333333  1.6666667         NA 2025-10-24 13:28:55    NA
#>  55:   failed   3.333333  0.5555556         NA 2025-10-24 13:28:55    NA
#>  56:   failed   3.333333 -0.5555556         NA 2025-10-24 13:28:55    NA
#>  57:   failed   3.333333 -1.6666667         NA 2025-10-24 13:28:55    NA
#>  58:   failed   3.333333 -2.7777778         NA 2025-10-24 13:28:55    NA
#>  59:   failed   3.333333 -3.8888889         NA 2025-10-24 13:28:55    NA
#>  60:   failed   3.333333 -5.0000000         NA 2025-10-24 13:28:55    NA
#>  61:   failed   1.111111  5.0000000         NA 2025-10-24 13:28:55    NA
#>  62:   failed   1.111111  3.8888889         NA 2025-10-24 13:28:55    NA
#>  63:   failed   1.111111  2.7777778         NA 2025-10-24 13:28:55    NA
#>  64:   failed   1.111111  1.6666667         NA 2025-10-24 13:28:55    NA
#>  65:   failed   1.111111  0.5555556         NA 2025-10-24 13:28:55    NA
#>  66:   failed   1.111111 -0.5555556         NA 2025-10-24 13:28:55    NA
#>  67:   failed   1.111111 -1.6666667         NA 2025-10-24 13:28:55    NA
#>  68:   failed   1.111111 -2.7777778         NA 2025-10-24 13:28:55    NA
#>  69:   failed   1.111111 -3.8888889         NA 2025-10-24 13:28:55    NA
#>  70:   failed   1.111111 -5.0000000         NA 2025-10-24 13:28:55    NA
#>  71:   failed  -1.111111  5.0000000         NA 2025-10-24 13:28:55    NA
#>  72:   failed  -1.111111  3.8888889         NA 2025-10-24 13:28:55    NA
#>  73:   failed  -1.111111  2.7777778         NA 2025-10-24 13:28:55    NA
#>  74:   failed  -1.111111  1.6666667         NA 2025-10-24 13:28:55    NA
#>  75:   failed  -1.111111  0.5555556         NA 2025-10-24 13:28:55    NA
#>  76:   failed  -1.111111 -0.5555556         NA 2025-10-24 13:28:55    NA
#>  77:   failed  -1.111111 -1.6666667         NA 2025-10-24 13:28:55    NA
#>  78:   failed  -1.111111 -2.7777778         NA 2025-10-24 13:28:55    NA
#>  79:   failed  -1.111111 -3.8888889         NA 2025-10-24 13:28:55    NA
#>  80:   failed  -1.111111 -5.0000000         NA 2025-10-24 13:28:55    NA
#>  81:   failed  -3.333333  5.0000000         NA 2025-10-24 13:28:55    NA
#>  82:   failed  -3.333333  3.8888889         NA 2025-10-24 13:28:55    NA
#>  83:   failed  -3.333333  2.7777778         NA 2025-10-24 13:28:55    NA
#>  84:   failed  -3.333333  1.6666667         NA 2025-10-24 13:28:55    NA
#>  85:   failed  -3.333333  0.5555556         NA 2025-10-24 13:28:55    NA
#>  86:   failed  -3.333333 -0.5555556         NA 2025-10-24 13:28:55    NA
#>  87:   failed  -3.333333 -1.6666667         NA 2025-10-24 13:28:55    NA
#>  88:   failed  -3.333333 -2.7777778         NA 2025-10-24 13:28:55    NA
#>  89:   failed  -3.333333 -3.8888889         NA 2025-10-24 13:28:55    NA
#>  90:   failed  -3.333333 -5.0000000         NA 2025-10-24 13:28:55    NA
#>  91:   failed  -5.555556  5.0000000         NA 2025-10-24 13:28:55    NA
#>  92:   failed  -5.555556  3.8888889         NA 2025-10-24 13:28:55    NA
#>  93:   failed  -5.555556  2.7777778         NA 2025-10-24 13:28:55    NA
#>  94:   failed  -5.555556  1.6666667         NA 2025-10-24 13:28:55    NA
#>  95:   failed  -5.555556  0.5555556         NA 2025-10-24 13:28:55    NA
#>  96:   failed  -5.555556 -0.5555556         NA 2025-10-24 13:28:55    NA
#>  97:   failed  -5.555556 -1.6666667         NA 2025-10-24 13:28:55    NA
#>  98:   failed  -5.555556 -2.7777778         NA 2025-10-24 13:28:55    NA
#>  99:   failed  -5.555556 -3.8888889         NA 2025-10-24 13:28:55    NA
#> 100:   failed  -5.555556 -5.0000000         NA 2025-10-24 13:28:55    NA
#>         state         x1         x2          y        timestamp_xs   pid
#>                 worker_id        timestamp_ys
#>                    <char>              <POSc>
#>   1: academical_guineapig 2025-10-24 13:28:56
#>   2: academical_guineapig 2025-10-24 13:28:56
#>   3: academical_guineapig 2025-10-24 13:28:56
#>   4: academical_guineapig 2025-10-24 13:28:56
#>   5: academical_guineapig 2025-10-24 13:28:56
#>   6: academical_guineapig 2025-10-24 13:28:56
#>   7: academical_guineapig 2025-10-24 13:28:56
#>   8: academical_guineapig 2025-10-24 13:28:56
#>   9: academical_guineapig 2025-10-24 13:28:56
#>  10: academical_guineapig 2025-10-24 13:28:56
#>  11: academical_guineapig 2025-10-24 13:28:56
#>  12: academical_guineapig 2025-10-24 13:28:56
#>  13: academical_guineapig 2025-10-24 13:28:56
#>  14: academical_guineapig 2025-10-24 13:28:56
#>  15: academical_guineapig 2025-10-24 13:28:56
#>  16: academical_guineapig 2025-10-24 13:28:56
#>  17: academical_guineapig 2025-10-24 13:28:56
#>  18: academical_guineapig 2025-10-24 13:28:56
#>  19: academical_guineapig 2025-10-24 13:28:56
#>  20: academical_guineapig 2025-10-24 13:28:56
#>  21:                 <NA>                <NA>
#>  22:                 <NA>                <NA>
#>  23:                 <NA>                <NA>
#>  24:                 <NA>                <NA>
#>  25:                 <NA>                <NA>
#>  26:                 <NA>                <NA>
#>  27:                 <NA>                <NA>
#>  28:                 <NA>                <NA>
#>  29:                 <NA>                <NA>
#>  30:                 <NA>                <NA>
#>  31:                 <NA>                <NA>
#>  32:                 <NA>                <NA>
#>  33:                 <NA>                <NA>
#>  34:                 <NA>                <NA>
#>  35:                 <NA>                <NA>
#>  36:                 <NA>                <NA>
#>  37:                 <NA>                <NA>
#>  38:                 <NA>                <NA>
#>  39:                 <NA>                <NA>
#>  40:                 <NA>                <NA>
#>  41:                 <NA>                <NA>
#>  42:                 <NA>                <NA>
#>  43:                 <NA>                <NA>
#>  44:                 <NA>                <NA>
#>  45:                 <NA>                <NA>
#>  46:                 <NA>                <NA>
#>  47:                 <NA>                <NA>
#>  48:                 <NA>                <NA>
#>  49:                 <NA>                <NA>
#>  50:                 <NA>                <NA>
#>  51:                 <NA>                <NA>
#>  52:                 <NA>                <NA>
#>  53:                 <NA>                <NA>
#>  54:                 <NA>                <NA>
#>  55:                 <NA>                <NA>
#>  56:                 <NA>                <NA>
#>  57:                 <NA>                <NA>
#>  58:                 <NA>                <NA>
#>  59:                 <NA>                <NA>
#>  60:                 <NA>                <NA>
#>  61:                 <NA>                <NA>
#>  62:                 <NA>                <NA>
#>  63:                 <NA>                <NA>
#>  64:                 <NA>                <NA>
#>  65:                 <NA>                <NA>
#>  66:                 <NA>                <NA>
#>  67:                 <NA>                <NA>
#>  68:                 <NA>                <NA>
#>  69:                 <NA>                <NA>
#>  70:                 <NA>                <NA>
#>  71:                 <NA>                <NA>
#>  72:                 <NA>                <NA>
#>  73:                 <NA>                <NA>
#>  74:                 <NA>                <NA>
#>  75:                 <NA>                <NA>
#>  76:                 <NA>                <NA>
#>  77:                 <NA>                <NA>
#>  78:                 <NA>                <NA>
#>  79:                 <NA>                <NA>
#>  80:                 <NA>                <NA>
#>  81:                 <NA>                <NA>
#>  82:                 <NA>                <NA>
#>  83:                 <NA>                <NA>
#>  84:                 <NA>                <NA>
#>  85:                 <NA>                <NA>
#>  86:                 <NA>                <NA>
#>  87:                 <NA>                <NA>
#>  88:                 <NA>                <NA>
#>  89:                 <NA>                <NA>
#>  90:                 <NA>                <NA>
#>  91:                 <NA>                <NA>
#>  92:                 <NA>                <NA>
#>  93:                 <NA>                <NA>
#>  94:                 <NA>                <NA>
#>  95:                 <NA>                <NA>
#>  96:                 <NA>                <NA>
#>  97:                 <NA>                <NA>
#>  98:                 <NA>                <NA>
#>  99:                 <NA>                <NA>
#> 100:                 <NA>                <NA>
#>                 worker_id        timestamp_ys
#>                                      keys                 message x_domain_x1
#>                                    <char>                  <char>       <num>
#>   1: 9251da87-2df1-452c-8cad-f6c9f765e03c                    <NA>  -10.000000
#>   2: f4d85f1f-099f-49e2-a25a-6f267bd990b9                    <NA>  -10.000000
#>   3: cc7e0c93-92b2-4d14-831c-69da6efdc89f                    <NA>  -10.000000
#>   4: e9c1c44c-a722-4c76-9368-021e55a84a9d                    <NA>  -10.000000
#>   5: 8ef82077-49e6-4b47-ac1e-3e93e4b90e6a                    <NA>  -10.000000
#>   6: fb301780-7eb4-405a-b691-ce5bc32e3e23                    <NA>  -10.000000
#>   7: 98008f9e-ac7f-4b0a-a088-d3da861cceb2                    <NA>  -10.000000
#>   8: faaf319f-5912-4f09-af45-f5c091a78502                    <NA>  -10.000000
#>   9: 688e2703-ef1f-49bd-891d-75aafef2a4be                    <NA>  -10.000000
#>  10: d0bab3e9-7a43-439b-8111-b7b557df56f0                    <NA>  -10.000000
#>  11: 75f5b780-7bbf-482f-9dc1-d66814a82e5e                    <NA>   -7.777778
#>  12: d26b9609-02cf-4c77-9ed2-d7f6d09955ed                    <NA>   -7.777778
#>  13: beb87495-366f-4aac-8d92-a198a9e2b710                    <NA>   -7.777778
#>  14: 6f9a483d-e81f-4267-a481-117c03e719ef                    <NA>   -7.777778
#>  15: d649b960-6862-409f-8811-3cdd89470dbc                    <NA>   -7.777778
#>  16: 5b3ef85c-61e9-438d-a090-e771dd8fcbaf                    <NA>   -7.777778
#>  17: fcf7fda5-4eda-4f60-9a95-6c601964c6c0                    <NA>   -7.777778
#>  18: f815eeee-c915-4b7a-913f-f4f2d7e49e38                    <NA>   -7.777778
#>  19: 39d7a075-29e0-487a-ab37-0bbcf8e8825f                    <NA>   -7.777778
#>  20: 9c000168-0f21-4aa5-9551-d80f0beab73a                    <NA>   -7.777778
#>  21: 57f2447c-4c31-4abc-87ad-c42f0f6f1a77 Optimization terminated          NA
#>  22: 627c05fc-76fd-4a8a-9e51-23c87d382cc4 Optimization terminated          NA
#>  23: 62e6c4af-1b7e-43d3-8d05-a8819fa6a2d8 Optimization terminated          NA
#>  24: c686727b-6e05-4461-a21e-196593ca82ec Optimization terminated          NA
#>  25: 23d214c0-2048-4bf5-b306-490e1f8cab5a Optimization terminated          NA
#>  26: 4e80ef38-c59e-446c-b787-cb0c189ee460 Optimization terminated          NA
#>  27: 1040124e-e6c0-4f90-a967-02c3aa3b07b2 Optimization terminated          NA
#>  28: 056c000a-21a0-465f-9bb5-f1dbe5a1ad85 Optimization terminated          NA
#>  29: 22251f8d-2790-4958-85fc-6b8b70cf47a9 Optimization terminated          NA
#>  30: 48bcd339-8b88-4a41-8bf3-cbbd038c6d89 Optimization terminated          NA
#>  31: c2a2c293-fcb4-4c87-811e-36b7ce080140 Optimization terminated          NA
#>  32: c61807c5-4cd7-4473-bc76-fd29b00c03ff Optimization terminated          NA
#>  33: 57b16761-5750-4b6a-9156-c75124b6ce3c Optimization terminated          NA
#>  34: 65f4f601-ee7b-438b-b642-b447f9140ed6 Optimization terminated          NA
#>  35: 487056d2-ef64-4743-b2d9-0130bc448cae Optimization terminated          NA
#>  36: cb417be1-117d-4b37-8f81-da0ade3a43bb Optimization terminated          NA
#>  37: 7e465bbb-7d8c-483c-9050-486d6d97d5f2 Optimization terminated          NA
#>  38: 3262ea0c-9f7a-4f67-9145-cf97f9953c75 Optimization terminated          NA
#>  39: e5e43402-437e-478c-8d9a-1520873325b8 Optimization terminated          NA
#>  40: 075ba121-9ed0-4094-b648-3e782c6cf730 Optimization terminated          NA
#>  41: 1a6b5d10-1ed3-4881-9f1a-3eb1a55215d1 Optimization terminated          NA
#>  42: 46009dc8-9589-47ea-b116-a3c4f4ad6eb6 Optimization terminated          NA
#>  43: af35709b-9ac8-4038-8c1f-979e4875ecb3 Optimization terminated          NA
#>  44: 4d3df563-9f47-43d1-b822-87145efcf0f1 Optimization terminated          NA
#>  45: 1382af83-2a13-447f-97c4-5b61308d7a76 Optimization terminated          NA
#>  46: 0a99991c-f4e3-4739-8adb-05d734ab6c27 Optimization terminated          NA
#>  47: 319fe61b-cc9f-4149-91c7-95dd8fe8e722 Optimization terminated          NA
#>  48: 9973fb85-5f70-4e6b-a5c1-3d919b76ec84 Optimization terminated          NA
#>  49: 8759a2d5-ee5f-46bb-b0a0-b1fd3e48e36e Optimization terminated          NA
#>  50: d0e9fafe-d9b6-48b9-89ee-dd8914f64260 Optimization terminated          NA
#>  51: 23fb6d53-53bf-4607-a18e-29cd6ea8d95a Optimization terminated          NA
#>  52: 975a2d66-060a-4883-bbb9-562b21b5dd15 Optimization terminated          NA
#>  53: 86d94228-ebe7-4163-b2cd-c753237a5826 Optimization terminated          NA
#>  54: c5c8153a-526b-4420-8994-dd31ae551594 Optimization terminated          NA
#>  55: 242a71aa-44af-4b12-a19f-5ef0cbb9ff10 Optimization terminated          NA
#>  56: a8f13f6c-80af-46b5-9c4c-37397bdbad11 Optimization terminated          NA
#>  57: e735b796-2bd6-454f-8301-a784ef783693 Optimization terminated          NA
#>  58: e12bb162-b753-4eff-8c95-b08eda8f023f Optimization terminated          NA
#>  59: 4ef1c30b-0d24-41b2-8143-7d28719daf5f Optimization terminated          NA
#>  60: 708bf258-3952-4d39-8c97-06de7d615b2a Optimization terminated          NA
#>  61: 31328379-e564-4fd2-a869-3629a83819b9 Optimization terminated          NA
#>  62: 3ded7905-4d19-4723-836f-1b37899986a0 Optimization terminated          NA
#>  63: e72d1a9f-01bc-4474-910d-b05a620ed44e Optimization terminated          NA
#>  64: 447be313-1246-4d37-91e6-93cd5ba01d12 Optimization terminated          NA
#>  65: d191f296-e17a-4806-8992-7ec38afed747 Optimization terminated          NA
#>  66: c33fc35c-b3d3-46fd-825e-ad4b8c98ee1c Optimization terminated          NA
#>  67: 3e8274d7-bedd-4465-8331-6487125df9da Optimization terminated          NA
#>  68: 83827837-a281-4a1a-9c2b-8d289173bc6a Optimization terminated          NA
#>  69: 3f1cd019-6b30-42ad-99e5-95f3154fc396 Optimization terminated          NA
#>  70: f69fbe80-e3cf-4c66-94b1-561bd19310fa Optimization terminated          NA
#>  71: 282e4cfd-069a-4d5f-9635-70d192b3f1b9 Optimization terminated          NA
#>  72: 40148d73-cdef-4f0b-ad9f-8bbd6559e300 Optimization terminated          NA
#>  73: 0477aecf-1f45-466a-9472-eea3449b1b85 Optimization terminated          NA
#>  74: 2a6ec32e-82f2-4c21-80ef-17d895f2cbdb Optimization terminated          NA
#>  75: df9d23d3-efc9-48bf-b574-b060c21ce803 Optimization terminated          NA
#>  76: 0538a885-002b-4029-b270-27a9b4d899c9 Optimization terminated          NA
#>  77: dbea4f32-0a94-437d-be7c-01f967088442 Optimization terminated          NA
#>  78: c39c3565-e8f7-49db-87bd-06729f8dca76 Optimization terminated          NA
#>  79: c9ab03e1-4791-4f2d-b690-d7bd982ea1fa Optimization terminated          NA
#>  80: f4f33174-2d7e-43d0-b941-6fbadd1fd61d Optimization terminated          NA
#>  81: e4a9150b-4422-4266-b6f6-6716bc6d8d27 Optimization terminated          NA
#>  82: 81d647e1-c837-4331-9a5e-bc8c2c70948f Optimization terminated          NA
#>  83: 48b6f607-88a3-4ccf-9327-dff471f754dc Optimization terminated          NA
#>  84: 44bf7e5d-ec40-42c8-943a-ae3ba37b8277 Optimization terminated          NA
#>  85: 0be59f45-9f73-4215-9fa1-375bab92953e Optimization terminated          NA
#>  86: 3b718a44-828f-4e57-87a6-2cdba4b6ac9a Optimization terminated          NA
#>  87: 7083663a-ffab-4e77-9767-631c6fa1a710 Optimization terminated          NA
#>  88: 3e4e8e8b-a908-4493-9921-4813626819e1 Optimization terminated          NA
#>  89: 1d005297-734f-4096-b7ac-e9a95e403bf6 Optimization terminated          NA
#>  90: 04401881-a526-4574-9e3c-a1ee59031fa6 Optimization terminated          NA
#>  91: 25d0313d-8556-479e-bc0d-76750d2a94e4 Optimization terminated          NA
#>  92: a6671bf7-caf3-4f8e-95a7-99bb4fcd0f57 Optimization terminated          NA
#>  93: acccb712-f548-4746-ad3b-991d65f34e95 Optimization terminated          NA
#>  94: 7bc05381-63df-4c16-bdfe-40d5253e227b Optimization terminated          NA
#>  95: 99fa6c75-6500-40a0-92f9-2e8d33ea35fe Optimization terminated          NA
#>  96: 1697ca4e-2d43-4679-995a-cd4a0eba7f2a Optimization terminated          NA
#>  97: cd046fae-3d76-43c0-89fe-a0c0edb87b60 Optimization terminated          NA
#>  98: 3bbb2f60-3cc5-47fa-bbfb-b2756ad03cea Optimization terminated          NA
#>  99: 115e638f-ca55-43f7-8954-4b8f47bac742 Optimization terminated          NA
#> 100: 6d544193-8f9c-4dcc-934b-c419d7df7735 Optimization terminated          NA
#>                                      keys                 message x_domain_x1
#>      x_domain_x2
#>            <num>
#>   1:  -5.0000000
#>   2:  -3.8888889
#>   3:  -2.7777778
#>   4:  -1.6666667
#>   5:  -0.5555556
#>   6:   0.5555556
#>   7:   1.6666667
#>   8:   2.7777778
#>   9:   3.8888889
#>  10:   5.0000000
#>  11:  -5.0000000
#>  12:  -3.8888889
#>  13:  -2.7777778
#>  14:  -1.6666667
#>  15:  -0.5555556
#>  16:   0.5555556
#>  17:   1.6666667
#>  18:   2.7777778
#>  19:   3.8888889
#>  20:   5.0000000
#>  21:          NA
#>  22:          NA
#>  23:          NA
#>  24:          NA
#>  25:          NA
#>  26:          NA
#>  27:          NA
#>  28:          NA
#>  29:          NA
#>  30:          NA
#>  31:          NA
#>  32:          NA
#>  33:          NA
#>  34:          NA
#>  35:          NA
#>  36:          NA
#>  37:          NA
#>  38:          NA
#>  39:          NA
#>  40:          NA
#>  41:          NA
#>  42:          NA
#>  43:          NA
#>  44:          NA
#>  45:          NA
#>  46:          NA
#>  47:          NA
#>  48:          NA
#>  49:          NA
#>  50:          NA
#>  51:          NA
#>  52:          NA
#>  53:          NA
#>  54:          NA
#>  55:          NA
#>  56:          NA
#>  57:          NA
#>  58:          NA
#>  59:          NA
#>  60:          NA
#>  61:          NA
#>  62:          NA
#>  63:          NA
#>  64:          NA
#>  65:          NA
#>  66:          NA
#>  67:          NA
#>  68:          NA
#>  69:          NA
#>  70:          NA
#>  71:          NA
#>  72:          NA
#>  73:          NA
#>  74:          NA
#>  75:          NA
#>  76:          NA
#>  77:          NA
#>  78:          NA
#>  79:          NA
#>  80:          NA
#>  81:          NA
#>  82:          NA
#>  83:          NA
#>  84:          NA
#>  85:          NA
#>  86:          NA
#>  87:          NA
#>  88:          NA
#>  89:          NA
#>  90:          NA
#>  91:          NA
#>  92:          NA
#>  93:          NA
#>  94:          NA
#>  95:          NA
#>  96:          NA
#>  97:          NA
#>  98:          NA
#>  99:          NA
#> 100:          NA
#>      x_domain_x2