Asynchronous Optimization via Grid Search

OptimizerAsyncGridSearch class that implements a grid search. The grid is constructed as a Cartesian product over discretized values per parameter, see paradox::generate_design_grid(). The points of the grid are evaluated in a random order.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281–305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("async_grid_search")
opt("async_grid_search")

Parameters

batch_size

integer(1)
Maximum number of points to try in a batch.

Super classes

bbotk::Optimizer -> bbotk::OptimizerAsync -> OptimizerAsyncGridSearch

Methods

Public methods

OptimizerAsyncGridSearch$new()

OptimizerAsyncGridSearch$optimize()

OptimizerAsyncGridSearch$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAsyncGridSearch$new()

Method optimize()

Starts the asynchronous optimization.

Usage

OptimizerAsyncGridSearch$optimize(inst)

Arguments

inst

(OptimInstance).

Returns

data.table::data.table.

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAsyncGridSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# example only runs if a Redis server is available
if (mlr3misc::require_namespaces(c("rush", "redux", "mirai"), quietly = TRUE) &&
  redux::redis_available()) {
# define the objective function
fun = function(xs) {
  list(y = - (xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10)
}

# set domain
domain = ps(
  x1 = p_dbl(-10, 10),
  x2 = p_dbl(-5, 5)
)

# set codomain
codomain = ps(
  y = p_dbl(tags = "maximize")
)

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = codomain,
  properties = "deterministic"
)

# start workers
rush::rush_plan(worker_type = "remote")
mirai::daemons(1)

# initialize instance
instance = oi_async(
  objective = objective,
  terminator = trm("evals", n_evals = 20)
)

# load optimizer
optimizer = opt("async_grid_search", resolution = 10)

# trigger optimization
optimizer$optimize(instance)

# all evaluated configurations
instance$archive

# best performing configuration
instance$archive$best()

# covert to data.table
as.data.table(instance$archive)
}
#>         state         x1         x2          y        timestamp_xs   pid
#>        <char>      <num>      <num>      <num>              <POSc> <int>
#>   1: finished -10.000000 -5.0000000 -138.00000 2025-10-24 13:34:22  9671
#>   2: finished -10.000000 -3.8888889 -134.79012 2025-10-24 13:34:22  9671
#>   3: finished -10.000000 -2.7777778 -134.04938 2025-10-24 13:34:22  9671
#>   4: finished -10.000000 -1.6666667 -135.77778 2025-10-24 13:34:22  9671
#>   5: finished -10.000000 -0.5555556 -139.97531 2025-10-24 13:34:22  9671
#>   6: finished -10.000000  0.5555556 -146.64198 2025-10-24 13:34:22  9671
#>   7: finished -10.000000  1.6666667 -155.77778 2025-10-24 13:34:22  9671
#>   8: finished -10.000000  2.7777778 -167.38272 2025-10-24 13:34:22  9671
#>   9: finished -10.000000  3.8888889 -181.45679 2025-10-24 13:34:22  9671
#>  10: finished -10.000000  5.0000000 -198.00000 2025-10-24 13:34:22  9671
#>  11: finished  -7.777778 -5.0000000  -89.60494 2025-10-24 13:34:22  9671
#>  12: finished  -7.777778 -3.8888889  -86.39506 2025-10-24 13:34:22  9671
#>  13: finished  -7.777778 -2.7777778  -85.65432 2025-10-24 13:34:22  9671
#>  14: finished  -7.777778 -1.6666667  -87.38272 2025-10-24 13:34:22  9671
#>  15: finished  -7.777778 -0.5555556  -91.58025 2025-10-24 13:34:22  9671
#>  16: finished  -7.777778  0.5555556  -98.24691 2025-10-24 13:34:22  9671
#>  17: finished  -7.777778  1.6666667 -107.38272 2025-10-24 13:34:22  9671
#>  18: finished  -7.777778  2.7777778 -118.98765 2025-10-24 13:34:22  9671
#>  19: finished  -7.777778  3.8888889 -133.06173 2025-10-24 13:34:22  9671
#>  20: finished  -7.777778  5.0000000 -149.60494 2025-10-24 13:34:22  9671
#>  21:   failed  10.000000  5.0000000         NA 2025-10-24 13:34:22    NA
#>  22:   failed  10.000000  3.8888889         NA 2025-10-24 13:34:22    NA
#>  23:   failed  10.000000  2.7777778         NA 2025-10-24 13:34:22    NA
#>  24:   failed  10.000000  1.6666667         NA 2025-10-24 13:34:22    NA
#>  25:   failed  10.000000  0.5555556         NA 2025-10-24 13:34:22    NA
#>  26:   failed  10.000000 -0.5555556         NA 2025-10-24 13:34:22    NA
#>  27:   failed  10.000000 -1.6666667         NA 2025-10-24 13:34:22    NA
#>  28:   failed  10.000000 -2.7777778         NA 2025-10-24 13:34:22    NA
#>  29:   failed  10.000000 -3.8888889         NA 2025-10-24 13:34:22    NA
#>  30:   failed  10.000000 -5.0000000         NA 2025-10-24 13:34:22    NA
#>  31:   failed   7.777778  5.0000000         NA 2025-10-24 13:34:22    NA
#>  32:   failed   7.777778  3.8888889         NA 2025-10-24 13:34:22    NA
#>  33:   failed   7.777778  2.7777778         NA 2025-10-24 13:34:22    NA
#>  34:   failed   7.777778  1.6666667         NA 2025-10-24 13:34:22    NA
#>  35:   failed   7.777778  0.5555556         NA 2025-10-24 13:34:22    NA
#>  36:   failed   7.777778 -0.5555556         NA 2025-10-24 13:34:22    NA
#>  37:   failed   7.777778 -1.6666667         NA 2025-10-24 13:34:22    NA
#>  38:   failed   7.777778 -2.7777778         NA 2025-10-24 13:34:22    NA
#>  39:   failed   7.777778 -3.8888889         NA 2025-10-24 13:34:22    NA
#>  40:   failed   7.777778 -5.0000000         NA 2025-10-24 13:34:22    NA
#>  41:   failed   5.555556  5.0000000         NA 2025-10-24 13:34:22    NA
#>  42:   failed   5.555556  3.8888889         NA 2025-10-24 13:34:22    NA
#>  43:   failed   5.555556  2.7777778         NA 2025-10-24 13:34:22    NA
#>  44:   failed   5.555556  1.6666667         NA 2025-10-24 13:34:22    NA
#>  45:   failed   5.555556  0.5555556         NA 2025-10-24 13:34:22    NA
#>  46:   failed   5.555556 -0.5555556         NA 2025-10-24 13:34:22    NA
#>  47:   failed   5.555556 -1.6666667         NA 2025-10-24 13:34:22    NA
#>  48:   failed   5.555556 -2.7777778         NA 2025-10-24 13:34:22    NA
#>  49:   failed   5.555556 -3.8888889         NA 2025-10-24 13:34:22    NA
#>  50:   failed   5.555556 -5.0000000         NA 2025-10-24 13:34:22    NA
#>  51:   failed   3.333333  5.0000000         NA 2025-10-24 13:34:22    NA
#>  52:   failed   3.333333  3.8888889         NA 2025-10-24 13:34:22    NA
#>  53:   failed   3.333333  2.7777778         NA 2025-10-24 13:34:22    NA
#>  54:   failed   3.333333  1.6666667         NA 2025-10-24 13:34:22    NA
#>  55:   failed   3.333333  0.5555556         NA 2025-10-24 13:34:22    NA
#>  56:   failed   3.333333 -0.5555556         NA 2025-10-24 13:34:22    NA
#>  57:   failed   3.333333 -1.6666667         NA 2025-10-24 13:34:22    NA
#>  58:   failed   3.333333 -2.7777778         NA 2025-10-24 13:34:22    NA
#>  59:   failed   3.333333 -3.8888889         NA 2025-10-24 13:34:22    NA
#>  60:   failed   3.333333 -5.0000000         NA 2025-10-24 13:34:22    NA
#>  61:   failed   1.111111  5.0000000         NA 2025-10-24 13:34:22    NA
#>  62:   failed   1.111111  3.8888889         NA 2025-10-24 13:34:22    NA
#>  63:   failed   1.111111  2.7777778         NA 2025-10-24 13:34:22    NA
#>  64:   failed   1.111111  1.6666667         NA 2025-10-24 13:34:22    NA
#>  65:   failed   1.111111  0.5555556         NA 2025-10-24 13:34:22    NA
#>  66:   failed   1.111111 -0.5555556         NA 2025-10-24 13:34:22    NA
#>  67:   failed   1.111111 -1.6666667         NA 2025-10-24 13:34:22    NA
#>  68:   failed   1.111111 -2.7777778         NA 2025-10-24 13:34:22    NA
#>  69:   failed   1.111111 -3.8888889         NA 2025-10-24 13:34:22    NA
#>  70:   failed   1.111111 -5.0000000         NA 2025-10-24 13:34:22    NA
#>  71:   failed  -1.111111  5.0000000         NA 2025-10-24 13:34:22    NA
#>  72:   failed  -1.111111  3.8888889         NA 2025-10-24 13:34:22    NA
#>  73:   failed  -1.111111  2.7777778         NA 2025-10-24 13:34:22    NA
#>  74:   failed  -1.111111  1.6666667         NA 2025-10-24 13:34:22    NA
#>  75:   failed  -1.111111  0.5555556         NA 2025-10-24 13:34:22    NA
#>  76:   failed  -1.111111 -0.5555556         NA 2025-10-24 13:34:22    NA
#>  77:   failed  -1.111111 -1.6666667         NA 2025-10-24 13:34:22    NA
#>  78:   failed  -1.111111 -2.7777778         NA 2025-10-24 13:34:22    NA
#>  79:   failed  -1.111111 -3.8888889         NA 2025-10-24 13:34:22    NA
#>  80:   failed  -1.111111 -5.0000000         NA 2025-10-24 13:34:22    NA
#>  81:   failed  -3.333333  5.0000000         NA 2025-10-24 13:34:22    NA
#>  82:   failed  -3.333333  3.8888889         NA 2025-10-24 13:34:22    NA
#>  83:   failed  -3.333333  2.7777778         NA 2025-10-24 13:34:22    NA
#>  84:   failed  -3.333333  1.6666667         NA 2025-10-24 13:34:22    NA
#>  85:   failed  -3.333333  0.5555556         NA 2025-10-24 13:34:22    NA
#>  86:   failed  -3.333333 -0.5555556         NA 2025-10-24 13:34:22    NA
#>  87:   failed  -3.333333 -1.6666667         NA 2025-10-24 13:34:22    NA
#>  88:   failed  -3.333333 -2.7777778         NA 2025-10-24 13:34:22    NA
#>  89:   failed  -3.333333 -3.8888889         NA 2025-10-24 13:34:22    NA
#>  90:   failed  -3.333333 -5.0000000         NA 2025-10-24 13:34:22    NA
#>  91:   failed  -5.555556  5.0000000         NA 2025-10-24 13:34:22    NA
#>  92:   failed  -5.555556  3.8888889         NA 2025-10-24 13:34:22    NA
#>  93:   failed  -5.555556  2.7777778         NA 2025-10-24 13:34:22    NA
#>  94:   failed  -5.555556  1.6666667         NA 2025-10-24 13:34:22    NA
#>  95:   failed  -5.555556  0.5555556         NA 2025-10-24 13:34:22    NA
#>  96:   failed  -5.555556 -0.5555556         NA 2025-10-24 13:34:22    NA
#>  97:   failed  -5.555556 -1.6666667         NA 2025-10-24 13:34:22    NA
#>  98:   failed  -5.555556 -2.7777778         NA 2025-10-24 13:34:22    NA
#>  99:   failed  -5.555556 -3.8888889         NA 2025-10-24 13:34:22    NA
#> 100:   failed  -5.555556 -5.0000000         NA 2025-10-24 13:34:22    NA
#>         state         x1         x2          y        timestamp_xs   pid
#>                 worker_id        timestamp_ys
#>                    <char>              <POSc>
#>   1: academical_guineapig 2025-10-24 13:34:23
#>   2: academical_guineapig 2025-10-24 13:34:23
#>   3: academical_guineapig 2025-10-24 13:34:23
#>   4: academical_guineapig 2025-10-24 13:34:23
#>   5: academical_guineapig 2025-10-24 13:34:23
#>   6: academical_guineapig 2025-10-24 13:34:23
#>   7: academical_guineapig 2025-10-24 13:34:23
#>   8: academical_guineapig 2025-10-24 13:34:23
#>   9: academical_guineapig 2025-10-24 13:34:23
#>  10: academical_guineapig 2025-10-24 13:34:23
#>  11: academical_guineapig 2025-10-24 13:34:23
#>  12: academical_guineapig 2025-10-24 13:34:23
#>  13: academical_guineapig 2025-10-24 13:34:23
#>  14: academical_guineapig 2025-10-24 13:34:23
#>  15: academical_guineapig 2025-10-24 13:34:23
#>  16: academical_guineapig 2025-10-24 13:34:23
#>  17: academical_guineapig 2025-10-24 13:34:23
#>  18: academical_guineapig 2025-10-24 13:34:23
#>  19: academical_guineapig 2025-10-24 13:34:23
#>  20: academical_guineapig 2025-10-24 13:34:23
#>  21:                 <NA>                <NA>
#>  22:                 <NA>                <NA>
#>  23:                 <NA>                <NA>
#>  24:                 <NA>                <NA>
#>  25:                 <NA>                <NA>
#>  26:                 <NA>                <NA>
#>  27:                 <NA>                <NA>
#>  28:                 <NA>                <NA>
#>  29:                 <NA>                <NA>
#>  30:                 <NA>                <NA>
#>  31:                 <NA>                <NA>
#>  32:                 <NA>                <NA>
#>  33:                 <NA>                <NA>
#>  34:                 <NA>                <NA>
#>  35:                 <NA>                <NA>
#>  36:                 <NA>                <NA>
#>  37:                 <NA>                <NA>
#>  38:                 <NA>                <NA>
#>  39:                 <NA>                <NA>
#>  40:                 <NA>                <NA>
#>  41:                 <NA>                <NA>
#>  42:                 <NA>                <NA>
#>  43:                 <NA>                <NA>
#>  44:                 <NA>                <NA>
#>  45:                 <NA>                <NA>
#>  46:                 <NA>                <NA>
#>  47:                 <NA>                <NA>
#>  48:                 <NA>                <NA>
#>  49:                 <NA>                <NA>
#>  50:                 <NA>                <NA>
#>  51:                 <NA>                <NA>
#>  52:                 <NA>                <NA>
#>  53:                 <NA>                <NA>
#>  54:                 <NA>                <NA>
#>  55:                 <NA>                <NA>
#>  56:                 <NA>                <NA>
#>  57:                 <NA>                <NA>
#>  58:                 <NA>                <NA>
#>  59:                 <NA>                <NA>
#>  60:                 <NA>                <NA>
#>  61:                 <NA>                <NA>
#>  62:                 <NA>                <NA>
#>  63:                 <NA>                <NA>
#>  64:                 <NA>                <NA>
#>  65:                 <NA>                <NA>
#>  66:                 <NA>                <NA>
#>  67:                 <NA>                <NA>
#>  68:                 <NA>                <NA>
#>  69:                 <NA>                <NA>
#>  70:                 <NA>                <NA>
#>  71:                 <NA>                <NA>
#>  72:                 <NA>                <NA>
#>  73:                 <NA>                <NA>
#>  74:                 <NA>                <NA>
#>  75:                 <NA>                <NA>
#>  76:                 <NA>                <NA>
#>  77:                 <NA>                <NA>
#>  78:                 <NA>                <NA>
#>  79:                 <NA>                <NA>
#>  80:                 <NA>                <NA>
#>  81:                 <NA>                <NA>
#>  82:                 <NA>                <NA>
#>  83:                 <NA>                <NA>
#>  84:                 <NA>                <NA>
#>  85:                 <NA>                <NA>
#>  86:                 <NA>                <NA>
#>  87:                 <NA>                <NA>
#>  88:                 <NA>                <NA>
#>  89:                 <NA>                <NA>
#>  90:                 <NA>                <NA>
#>  91:                 <NA>                <NA>
#>  92:                 <NA>                <NA>
#>  93:                 <NA>                <NA>
#>  94:                 <NA>                <NA>
#>  95:                 <NA>                <NA>
#>  96:                 <NA>                <NA>
#>  97:                 <NA>                <NA>
#>  98:                 <NA>                <NA>
#>  99:                 <NA>                <NA>
#> 100:                 <NA>                <NA>
#>                 worker_id        timestamp_ys
#>                                      keys                 message x_domain_x1
#>                                    <char>                  <char>       <num>
#>   1: 057f50a2-cc46-4ded-ae87-dda5ebc911e9                    <NA>  -10.000000
#>   2: f4f089fc-d6b2-49f2-b729-1d68e6146927                    <NA>  -10.000000
#>   3: d39d0fbd-cbf1-4b70-a7a9-1ff6e915788c                    <NA>  -10.000000
#>   4: 5dbc61a2-ea7f-46ee-a2c1-8e44a13fef15                    <NA>  -10.000000
#>   5: aff9417a-37ed-471f-a0ba-1618bf9dea4b                    <NA>  -10.000000
#>   6: a8b1ee6f-d259-45de-83f2-c275977f9540                    <NA>  -10.000000
#>   7: 90f12182-bde1-4986-8cb5-f8e91534237d                    <NA>  -10.000000
#>   8: 5071d187-e1fa-465e-ad9b-7aa5de369c88                    <NA>  -10.000000
#>   9: 0034c838-05c6-4a45-aa10-4d7fd78e9716                    <NA>  -10.000000
#>  10: fc9c1cf3-5ef4-4c46-97f8-0090f6c104c2                    <NA>  -10.000000
#>  11: aaa86260-9d36-4d00-ab60-b1318647d792                    <NA>   -7.777778
#>  12: c908a85c-92c1-4dfc-9f7d-1a7fc14647dd                    <NA>   -7.777778
#>  13: e5475f85-e497-40d4-a207-3b8f438624e6                    <NA>   -7.777778
#>  14: e8e94d16-861c-4559-9abb-ba185e754c5b                    <NA>   -7.777778
#>  15: 003c57b6-a8e4-40d0-9cea-c5a463059117                    <NA>   -7.777778
#>  16: 0c0103df-5412-4730-8d16-ccbf25fcb900                    <NA>   -7.777778
#>  17: 25925f11-bb35-45be-b9b3-47f5fc4a5f89                    <NA>   -7.777778
#>  18: 878c1d02-7a01-4ab1-a76b-ed648d411124                    <NA>   -7.777778
#>  19: 1dc4a705-8977-4149-afdb-c43618fd43a6                    <NA>   -7.777778
#>  20: fa10f674-2748-489f-8a27-df476a0754f6                    <NA>   -7.777778
#>  21: 61e0605c-5fb3-4774-a231-1db5a69b7da4 Optimization terminated          NA
#>  22: 8e969a99-7764-4b9a-963f-f6d4c4fb76d8 Optimization terminated          NA
#>  23: 143e8a37-0857-41e0-ab4c-ccb51a7a098c Optimization terminated          NA
#>  24: 177954d5-73af-4cb1-9787-604311fd79a2 Optimization terminated          NA
#>  25: abd9e073-dd41-4c86-8d91-3a8ae05dc69a Optimization terminated          NA
#>  26: 3d008c0c-fb62-4bfa-b19c-b0dba9acc0c9 Optimization terminated          NA
#>  27: 7c471c04-6b7e-4092-b649-6adede584cc6 Optimization terminated          NA
#>  28: 22330d3b-024e-4aed-9d8a-0c1e5c5a02dd Optimization terminated          NA
#>  29: 3f6ddbd6-1198-49a6-b6df-a3eb6f6e3c4c Optimization terminated          NA
#>  30: 6d60d063-fc8c-4933-b88e-ab1f3f7c6b80 Optimization terminated          NA
#>  31: abb9ebf5-14c8-4d2f-8b56-5b9202314d58 Optimization terminated          NA
#>  32: 7f6c4587-a06f-4aa1-aee2-576a926b621a Optimization terminated          NA
#>  33: 02bc50a7-7242-43a0-a324-78d045264b16 Optimization terminated          NA
#>  34: 800906ba-26ea-46e3-b5ac-c999de84f8fa Optimization terminated          NA
#>  35: f6b49b3a-b9f5-4c63-8e32-74389adb7a76 Optimization terminated          NA
#>  36: 882ea94d-fe84-49c8-a1d0-195b11c73339 Optimization terminated          NA
#>  37: 9839e89d-6af4-4777-b5ce-5d50d8b193f6 Optimization terminated          NA
#>  38: 1e54e1b5-ff3a-4e1a-9474-a499d2b9ffda Optimization terminated          NA
#>  39: ded228e5-f8ff-4573-92c6-9910ea00ea86 Optimization terminated          NA
#>  40: ca0ec5bc-2ba0-4d19-9a0a-f566046bb5f9 Optimization terminated          NA
#>  41: 1073496d-4761-441c-a5fd-2e86a2dd853b Optimization terminated          NA
#>  42: 1792b68b-1114-4232-b37c-62d679b75180 Optimization terminated          NA
#>  43: 212d1fa1-f0e2-402b-9ec0-a154dbd180e5 Optimization terminated          NA
#>  44: 582ec5f7-49cd-4ea9-8791-3373e53042b5 Optimization terminated          NA
#>  45: 3ae43f43-264b-4442-a208-4e1f6bf9cc60 Optimization terminated          NA
#>  46: f33fdbc6-d938-4c5e-bcdf-16e977db2678 Optimization terminated          NA
#>  47: cf10b4bb-0d7a-4700-b87b-1994df42ab14 Optimization terminated          NA
#>  48: 46c22e40-4566-4603-accb-ce35f08d6160 Optimization terminated          NA
#>  49: 36b254f9-51c0-4a24-9e00-60d9dd7f4b7e Optimization terminated          NA
#>  50: 06981149-a79a-4e9d-8947-92826fc38125 Optimization terminated          NA
#>  51: 5dcee233-1ec4-43ce-897a-6e4fa5d3a5d5 Optimization terminated          NA
#>  52: 26291f08-873f-4ddc-a5cd-8c4a5cc5835c Optimization terminated          NA
#>  53: 8d1ba5db-a28e-4035-830b-c55b1d0f78f9 Optimization terminated          NA
#>  54: 795ea424-3f60-45d4-b508-0308ed8eb214 Optimization terminated          NA
#>  55: a79cb488-7696-4ce2-a26e-272f67558d58 Optimization terminated          NA
#>  56: a51efaef-071a-451c-869f-b81dccf90acb Optimization terminated          NA
#>  57: 7a121d6c-9ee4-458b-8d02-dce2bfcc9c65 Optimization terminated          NA
#>  58: 929c16bd-2966-4fa0-8c3c-512e1eb8ccda Optimization terminated          NA
#>  59: ab06953c-c84f-4f01-ba03-0ef848da26e7 Optimization terminated          NA
#>  60: 1bd271a0-ee6a-4f2b-91ad-15c89e81c8f0 Optimization terminated          NA
#>  61: 072d526b-7afe-44a8-bb63-fb79bd5c43f6 Optimization terminated          NA
#>  62: 3b958678-149d-4754-9a40-3f4027090d1d Optimization terminated          NA
#>  63: 10e83010-a66f-471f-b7ce-954b318164c3 Optimization terminated          NA
#>  64: 6a6c314e-1efa-48b5-abb2-c6ff53400c51 Optimization terminated          NA
#>  65: fda506bf-0990-4354-9065-2b10c94a632f Optimization terminated          NA
#>  66: 1bf8b702-679a-4a34-8979-1c40ca36156d Optimization terminated          NA
#>  67: d01e20d6-caec-47da-8cd8-60d745464e4f Optimization terminated          NA
#>  68: 89ad997e-c9d3-44f0-aa19-94d4e8a4bff0 Optimization terminated          NA
#>  69: f1b02bd0-9c6b-44b7-a7bf-a107c31f82d6 Optimization terminated          NA
#>  70: 72956092-96bb-4de0-b2fe-238250c1a98f Optimization terminated          NA
#>  71: 868977ae-3970-4c23-a042-ed056bb2681f Optimization terminated          NA
#>  72: 4afbd96f-8f1e-4655-8f9f-ed6e91a2e47e Optimization terminated          NA
#>  73: ea37c628-a076-4766-975a-6bfe0b36a30a Optimization terminated          NA
#>  74: 065f2fe4-d645-4558-94fa-991830586fd1 Optimization terminated          NA
#>  75: e979d995-896a-4a7c-b7ca-e621185d62ed Optimization terminated          NA
#>  76: 8ad3e28c-67a2-4fd1-a03b-72f67e3d13dd Optimization terminated          NA
#>  77: 7dc9d452-7949-4025-97cb-930109311e9f Optimization terminated          NA
#>  78: 07b37481-0eef-4bf7-96d2-39b9ad67a8f5 Optimization terminated          NA
#>  79: a936d5e2-16c9-4bdd-8c18-6fd0be8e6c2c Optimization terminated          NA
#>  80: 2de2b351-f285-434b-abd3-d0e118c4dcfa Optimization terminated          NA
#>  81: 18f0d66b-3309-4919-a477-b3e9fa65eeca Optimization terminated          NA
#>  82: 2a9323f0-8929-43eb-bd78-b2be8e8116b8 Optimization terminated          NA
#>  83: 66680581-d65e-421e-8f48-a93de8992ba7 Optimization terminated          NA
#>  84: 80c70622-9afe-4899-8ba7-511935bbe3a1 Optimization terminated          NA
#>  85: 7ebaf9d6-36be-4af6-803c-fb5061a1a887 Optimization terminated          NA
#>  86: 6045f598-8f76-4255-a184-4e0068e56e80 Optimization terminated          NA
#>  87: 7f863b20-0d29-45ff-b848-43a3e4259620 Optimization terminated          NA
#>  88: a6326ada-3949-4a0e-9c65-8b65123dd015 Optimization terminated          NA
#>  89: 3163ed27-b0bc-4a72-97dd-7c016c077ca7 Optimization terminated          NA
#>  90: a7cb5664-dbcd-41c2-8ab9-262f82b6d9e1 Optimization terminated          NA
#>  91: 11c06258-6edc-47b8-af83-8601a9a6eab6 Optimization terminated          NA
#>  92: b6380ec8-716a-4f87-b132-d301833df51c Optimization terminated          NA
#>  93: b9bbe983-678b-4f02-a4e8-6852d9c3938d Optimization terminated          NA
#>  94: 7a6e5cab-c058-4e19-a416-903298f1c583 Optimization terminated          NA
#>  95: 52082a09-9339-47f8-83af-d17942e54597 Optimization terminated          NA
#>  96: 99e68b87-db11-40e1-ae68-dd7db30af139 Optimization terminated          NA
#>  97: d969b93b-ab91-47b5-93e2-765da8912a67 Optimization terminated          NA
#>  98: abb8aa52-2c61-44c1-a2f9-de265fd54db0 Optimization terminated          NA
#>  99: 08c58923-2ad6-40d4-a514-ddbcae8fbec3 Optimization terminated          NA
#> 100: bdb11011-9d9c-4d88-8a54-14ddf2b0833c Optimization terminated          NA
#>                                      keys                 message x_domain_x1
#>      x_domain_x2
#>            <num>
#>   1:  -5.0000000
#>   2:  -3.8888889
#>   3:  -2.7777778
#>   4:  -1.6666667
#>   5:  -0.5555556
#>   6:   0.5555556
#>   7:   1.6666667
#>   8:   2.7777778
#>   9:   3.8888889
#>  10:   5.0000000
#>  11:  -5.0000000
#>  12:  -3.8888889
#>  13:  -2.7777778
#>  14:  -1.6666667
#>  15:  -0.5555556
#>  16:   0.5555556
#>  17:   1.6666667
#>  18:   2.7777778
#>  19:   3.8888889
#>  20:   5.0000000
#>  21:          NA
#>  22:          NA
#>  23:          NA
#>  24:          NA
#>  25:          NA
#>  26:          NA
#>  27:          NA
#>  28:          NA
#>  29:          NA
#>  30:          NA
#>  31:          NA
#>  32:          NA
#>  33:          NA
#>  34:          NA
#>  35:          NA
#>  36:          NA
#>  37:          NA
#>  38:          NA
#>  39:          NA
#>  40:          NA
#>  41:          NA
#>  42:          NA
#>  43:          NA
#>  44:          NA
#>  45:          NA
#>  46:          NA
#>  47:          NA
#>  48:          NA
#>  49:          NA
#>  50:          NA
#>  51:          NA
#>  52:          NA
#>  53:          NA
#>  54:          NA
#>  55:          NA
#>  56:          NA
#>  57:          NA
#>  58:          NA
#>  59:          NA
#>  60:          NA
#>  61:          NA
#>  62:          NA
#>  63:          NA
#>  64:          NA
#>  65:          NA
#>  66:          NA
#>  67:          NA
#>  68:          NA
#>  69:          NA
#>  70:          NA
#>  71:          NA
#>  72:          NA
#>  73:          NA
#>  74:          NA
#>  75:          NA
#>  76:          NA
#>  77:          NA
#>  78:          NA
#>  79:          NA
#>  80:          NA
#>  81:          NA
#>  82:          NA
#>  83:          NA
#>  84:          NA
#>  85:          NA
#>  86:          NA
#>  87:          NA
#>  88:          NA
#>  89:          NA
#>  90:          NA
#>  91:          NA
#>  92:          NA
#>  93:          NA
#>  94:          NA
#>  95:          NA
#>  96:          NA
#>  97:          NA
#>  98:          NA
#>  99:          NA
#> 100:          NA
#>      x_domain_x2