Asynchronous Optimization via Grid Search

OptimizerAsyncGridSearch class that implements a grid search. The grid is constructed as a Cartesian product over discretized values per parameter, see paradox::generate_design_grid(). The points of the grid are evaluated in a random order.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281–305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("async_grid_search")
opt("async_grid_search")

Parameters

batch_size

integer(1)
Maximum number of points to try in a batch.

Super classes

bbotk::Optimizer -> bbotk::OptimizerAsync -> OptimizerAsyncGridSearch

Methods

Public methods

OptimizerAsyncGridSearch$new()

OptimizerAsyncGridSearch$optimize()

OptimizerAsyncGridSearch$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAsyncGridSearch$new()

Method optimize()

Starts the asynchronous optimization.

Usage

OptimizerAsyncGridSearch$optimize(inst)

Arguments

inst

(OptimInstance).

Returns

data.table::data.table.

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAsyncGridSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# example only runs if a Redis server is available
if (mlr3misc::require_namespaces(c("rush", "redux", "mirai"), quietly = TRUE) &&
  redux::redis_available()) {
# define the objective function
fun = function(xs) {
  list(y = - (xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10)
}

# set domain
domain = ps(
  x1 = p_dbl(-10, 10),
  x2 = p_dbl(-5, 5)
)

# set codomain
codomain = ps(
  y = p_dbl(tags = "maximize")
)

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = codomain,
  properties = "deterministic"
)

# start workers
rush::rush_plan(worker_type = "remote")
mirai::daemons(1)

# initialize instance
instance = oi_async(
  objective = objective,
  terminator = trm("evals", n_evals = 20)
)

# load optimizer
optimizer = opt("async_grid_search", resolution = 10)

# trigger optimization
optimizer$optimize(instance)

# all evaluated configurations
instance$archive

# best performing configuration
instance$archive$best()

# covert to data.table
as.data.table(instance$archive)
}
#>         state         x1         x2          y        timestamp_xs   pid
#>        <char>      <num>      <num>      <num>              <POSc> <int>
#>   1: finished -10.000000 -5.0000000 -138.00000 2025-11-01 09:40:56 29828
#>   2: finished -10.000000 -3.8888889 -134.79012 2025-11-01 09:40:56 29828
#>   3: finished -10.000000 -2.7777778 -134.04938 2025-11-01 09:40:56 29828
#>   4: finished -10.000000 -1.6666667 -135.77778 2025-11-01 09:40:56 29828
#>   5: finished -10.000000 -0.5555556 -139.97531 2025-11-01 09:40:56 29828
#>   6: finished -10.000000  0.5555556 -146.64198 2025-11-01 09:40:56 29828
#>   7: finished -10.000000  1.6666667 -155.77778 2025-11-01 09:40:56 29828
#>   8: finished -10.000000  2.7777778 -167.38272 2025-11-01 09:40:56 29828
#>   9: finished -10.000000  3.8888889 -181.45679 2025-11-01 09:40:56 29828
#>  10: finished -10.000000  5.0000000 -198.00000 2025-11-01 09:40:56 29828
#>  11: finished  -7.777778 -5.0000000  -89.60494 2025-11-01 09:40:56 29828
#>  12: finished  -7.777778 -3.8888889  -86.39506 2025-11-01 09:40:56 29828
#>  13: finished  -7.777778 -2.7777778  -85.65432 2025-11-01 09:40:56 29828
#>  14: finished  -7.777778 -1.6666667  -87.38272 2025-11-01 09:40:56 29828
#>  15: finished  -7.777778 -0.5555556  -91.58025 2025-11-01 09:40:56 29828
#>  16: finished  -7.777778  0.5555556  -98.24691 2025-11-01 09:40:56 29828
#>  17: finished  -7.777778  1.6666667 -107.38272 2025-11-01 09:40:56 29828
#>  18: finished  -7.777778  2.7777778 -118.98765 2025-11-01 09:40:56 29828
#>  19: finished  -7.777778  3.8888889 -133.06173 2025-11-01 09:40:56 29828
#>  20: finished  -7.777778  5.0000000 -149.60494 2025-11-01 09:40:56 29828
#>  21:   failed  10.000000  5.0000000         NA 2025-11-01 09:40:56    NA
#>  22:   failed  10.000000  3.8888889         NA 2025-11-01 09:40:56    NA
#>  23:   failed  10.000000  2.7777778         NA 2025-11-01 09:40:56    NA
#>  24:   failed  10.000000  1.6666667         NA 2025-11-01 09:40:56    NA
#>  25:   failed  10.000000  0.5555556         NA 2025-11-01 09:40:56    NA
#>  26:   failed  10.000000 -0.5555556         NA 2025-11-01 09:40:56    NA
#>  27:   failed  10.000000 -1.6666667         NA 2025-11-01 09:40:56    NA
#>  28:   failed  10.000000 -2.7777778         NA 2025-11-01 09:40:56    NA
#>  29:   failed  10.000000 -3.8888889         NA 2025-11-01 09:40:56    NA
#>  30:   failed  10.000000 -5.0000000         NA 2025-11-01 09:40:56    NA
#>  31:   failed   7.777778  5.0000000         NA 2025-11-01 09:40:56    NA
#>  32:   failed   7.777778  3.8888889         NA 2025-11-01 09:40:56    NA
#>  33:   failed   7.777778  2.7777778         NA 2025-11-01 09:40:56    NA
#>  34:   failed   7.777778  1.6666667         NA 2025-11-01 09:40:56    NA
#>  35:   failed   7.777778  0.5555556         NA 2025-11-01 09:40:56    NA
#>  36:   failed   7.777778 -0.5555556         NA 2025-11-01 09:40:56    NA
#>  37:   failed   7.777778 -1.6666667         NA 2025-11-01 09:40:56    NA
#>  38:   failed   7.777778 -2.7777778         NA 2025-11-01 09:40:56    NA
#>  39:   failed   7.777778 -3.8888889         NA 2025-11-01 09:40:56    NA
#>  40:   failed   7.777778 -5.0000000         NA 2025-11-01 09:40:56    NA
#>  41:   failed   5.555556  5.0000000         NA 2025-11-01 09:40:56    NA
#>  42:   failed   5.555556  3.8888889         NA 2025-11-01 09:40:56    NA
#>  43:   failed   5.555556  2.7777778         NA 2025-11-01 09:40:56    NA
#>  44:   failed   5.555556  1.6666667         NA 2025-11-01 09:40:56    NA
#>  45:   failed   5.555556  0.5555556         NA 2025-11-01 09:40:56    NA
#>  46:   failed   5.555556 -0.5555556         NA 2025-11-01 09:40:56    NA
#>  47:   failed   5.555556 -1.6666667         NA 2025-11-01 09:40:56    NA
#>  48:   failed   5.555556 -2.7777778         NA 2025-11-01 09:40:56    NA
#>  49:   failed   5.555556 -3.8888889         NA 2025-11-01 09:40:56    NA
#>  50:   failed   5.555556 -5.0000000         NA 2025-11-01 09:40:56    NA
#>  51:   failed   3.333333  5.0000000         NA 2025-11-01 09:40:56    NA
#>  52:   failed   3.333333  3.8888889         NA 2025-11-01 09:40:56    NA
#>  53:   failed   3.333333  2.7777778         NA 2025-11-01 09:40:56    NA
#>  54:   failed   3.333333  1.6666667         NA 2025-11-01 09:40:56    NA
#>  55:   failed   3.333333  0.5555556         NA 2025-11-01 09:40:56    NA
#>  56:   failed   3.333333 -0.5555556         NA 2025-11-01 09:40:56    NA
#>  57:   failed   3.333333 -1.6666667         NA 2025-11-01 09:40:56    NA
#>  58:   failed   3.333333 -2.7777778         NA 2025-11-01 09:40:56    NA
#>  59:   failed   3.333333 -3.8888889         NA 2025-11-01 09:40:56    NA
#>  60:   failed   3.333333 -5.0000000         NA 2025-11-01 09:40:56    NA
#>  61:   failed   1.111111  5.0000000         NA 2025-11-01 09:40:56    NA
#>  62:   failed   1.111111  3.8888889         NA 2025-11-01 09:40:56    NA
#>  63:   failed   1.111111  2.7777778         NA 2025-11-01 09:40:56    NA
#>  64:   failed   1.111111  1.6666667         NA 2025-11-01 09:40:56    NA
#>  65:   failed   1.111111  0.5555556         NA 2025-11-01 09:40:56    NA
#>  66:   failed   1.111111 -0.5555556         NA 2025-11-01 09:40:56    NA
#>  67:   failed   1.111111 -1.6666667         NA 2025-11-01 09:40:56    NA
#>  68:   failed   1.111111 -2.7777778         NA 2025-11-01 09:40:56    NA
#>  69:   failed   1.111111 -3.8888889         NA 2025-11-01 09:40:56    NA
#>  70:   failed   1.111111 -5.0000000         NA 2025-11-01 09:40:56    NA
#>  71:   failed  -1.111111  5.0000000         NA 2025-11-01 09:40:56    NA
#>  72:   failed  -1.111111  3.8888889         NA 2025-11-01 09:40:56    NA
#>  73:   failed  -1.111111  2.7777778         NA 2025-11-01 09:40:56    NA
#>  74:   failed  -1.111111  1.6666667         NA 2025-11-01 09:40:56    NA
#>  75:   failed  -1.111111  0.5555556         NA 2025-11-01 09:40:56    NA
#>  76:   failed  -1.111111 -0.5555556         NA 2025-11-01 09:40:56    NA
#>  77:   failed  -1.111111 -1.6666667         NA 2025-11-01 09:40:56    NA
#>  78:   failed  -1.111111 -2.7777778         NA 2025-11-01 09:40:56    NA
#>  79:   failed  -1.111111 -3.8888889         NA 2025-11-01 09:40:56    NA
#>  80:   failed  -1.111111 -5.0000000         NA 2025-11-01 09:40:56    NA
#>  81:   failed  -3.333333  5.0000000         NA 2025-11-01 09:40:56    NA
#>  82:   failed  -3.333333  3.8888889         NA 2025-11-01 09:40:56    NA
#>  83:   failed  -3.333333  2.7777778         NA 2025-11-01 09:40:56    NA
#>  84:   failed  -3.333333  1.6666667         NA 2025-11-01 09:40:56    NA
#>  85:   failed  -3.333333  0.5555556         NA 2025-11-01 09:40:56    NA
#>  86:   failed  -3.333333 -0.5555556         NA 2025-11-01 09:40:56    NA
#>  87:   failed  -3.333333 -1.6666667         NA 2025-11-01 09:40:56    NA
#>  88:   failed  -3.333333 -2.7777778         NA 2025-11-01 09:40:56    NA
#>  89:   failed  -3.333333 -3.8888889         NA 2025-11-01 09:40:56    NA
#>  90:   failed  -3.333333 -5.0000000         NA 2025-11-01 09:40:56    NA
#>  91:   failed  -5.555556  5.0000000         NA 2025-11-01 09:40:56    NA
#>  92:   failed  -5.555556  3.8888889         NA 2025-11-01 09:40:56    NA
#>  93:   failed  -5.555556  2.7777778         NA 2025-11-01 09:40:56    NA
#>  94:   failed  -5.555556  1.6666667         NA 2025-11-01 09:40:56    NA
#>  95:   failed  -5.555556  0.5555556         NA 2025-11-01 09:40:56    NA
#>  96:   failed  -5.555556 -0.5555556         NA 2025-11-01 09:40:56    NA
#>  97:   failed  -5.555556 -1.6666667         NA 2025-11-01 09:40:56    NA
#>  98:   failed  -5.555556 -2.7777778         NA 2025-11-01 09:40:56    NA
#>  99:   failed  -5.555556 -3.8888889         NA 2025-11-01 09:40:56    NA
#> 100:   failed  -5.555556 -5.0000000         NA 2025-11-01 09:40:56    NA
#>         state         x1         x2          y        timestamp_xs   pid
#>                 worker_id        timestamp_ys
#>                    <char>              <POSc>
#>   1: academical_guineapig 2025-11-01 09:40:56
#>   2: academical_guineapig 2025-11-01 09:40:57
#>   3: academical_guineapig 2025-11-01 09:40:57
#>   4: academical_guineapig 2025-11-01 09:40:57
#>   5: academical_guineapig 2025-11-01 09:40:57
#>   6: academical_guineapig 2025-11-01 09:40:57
#>   7: academical_guineapig 2025-11-01 09:40:57
#>   8: academical_guineapig 2025-11-01 09:40:57
#>   9: academical_guineapig 2025-11-01 09:40:57
#>  10: academical_guineapig 2025-11-01 09:40:57
#>  11: academical_guineapig 2025-11-01 09:40:57
#>  12: academical_guineapig 2025-11-01 09:40:57
#>  13: academical_guineapig 2025-11-01 09:40:57
#>  14: academical_guineapig 2025-11-01 09:40:57
#>  15: academical_guineapig 2025-11-01 09:40:57
#>  16: academical_guineapig 2025-11-01 09:40:57
#>  17: academical_guineapig 2025-11-01 09:40:57
#>  18: academical_guineapig 2025-11-01 09:40:57
#>  19: academical_guineapig 2025-11-01 09:40:57
#>  20: academical_guineapig 2025-11-01 09:40:57
#>  21:                 <NA>                <NA>
#>  22:                 <NA>                <NA>
#>  23:                 <NA>                <NA>
#>  24:                 <NA>                <NA>
#>  25:                 <NA>                <NA>
#>  26:                 <NA>                <NA>
#>  27:                 <NA>                <NA>
#>  28:                 <NA>                <NA>
#>  29:                 <NA>                <NA>
#>  30:                 <NA>                <NA>
#>  31:                 <NA>                <NA>
#>  32:                 <NA>                <NA>
#>  33:                 <NA>                <NA>
#>  34:                 <NA>                <NA>
#>  35:                 <NA>                <NA>
#>  36:                 <NA>                <NA>
#>  37:                 <NA>                <NA>
#>  38:                 <NA>                <NA>
#>  39:                 <NA>                <NA>
#>  40:                 <NA>                <NA>
#>  41:                 <NA>                <NA>
#>  42:                 <NA>                <NA>
#>  43:                 <NA>                <NA>
#>  44:                 <NA>                <NA>
#>  45:                 <NA>                <NA>
#>  46:                 <NA>                <NA>
#>  47:                 <NA>                <NA>
#>  48:                 <NA>                <NA>
#>  49:                 <NA>                <NA>
#>  50:                 <NA>                <NA>
#>  51:                 <NA>                <NA>
#>  52:                 <NA>                <NA>
#>  53:                 <NA>                <NA>
#>  54:                 <NA>                <NA>
#>  55:                 <NA>                <NA>
#>  56:                 <NA>                <NA>
#>  57:                 <NA>                <NA>
#>  58:                 <NA>                <NA>
#>  59:                 <NA>                <NA>
#>  60:                 <NA>                <NA>
#>  61:                 <NA>                <NA>
#>  62:                 <NA>                <NA>
#>  63:                 <NA>                <NA>
#>  64:                 <NA>                <NA>
#>  65:                 <NA>                <NA>
#>  66:                 <NA>                <NA>
#>  67:                 <NA>                <NA>
#>  68:                 <NA>                <NA>
#>  69:                 <NA>                <NA>
#>  70:                 <NA>                <NA>
#>  71:                 <NA>                <NA>
#>  72:                 <NA>                <NA>
#>  73:                 <NA>                <NA>
#>  74:                 <NA>                <NA>
#>  75:                 <NA>                <NA>
#>  76:                 <NA>                <NA>
#>  77:                 <NA>                <NA>
#>  78:                 <NA>                <NA>
#>  79:                 <NA>                <NA>
#>  80:                 <NA>                <NA>
#>  81:                 <NA>                <NA>
#>  82:                 <NA>                <NA>
#>  83:                 <NA>                <NA>
#>  84:                 <NA>                <NA>
#>  85:                 <NA>                <NA>
#>  86:                 <NA>                <NA>
#>  87:                 <NA>                <NA>
#>  88:                 <NA>                <NA>
#>  89:                 <NA>                <NA>
#>  90:                 <NA>                <NA>
#>  91:                 <NA>                <NA>
#>  92:                 <NA>                <NA>
#>  93:                 <NA>                <NA>
#>  94:                 <NA>                <NA>
#>  95:                 <NA>                <NA>
#>  96:                 <NA>                <NA>
#>  97:                 <NA>                <NA>
#>  98:                 <NA>                <NA>
#>  99:                 <NA>                <NA>
#> 100:                 <NA>                <NA>
#>                 worker_id        timestamp_ys
#>                                      keys                 message x_domain_x1
#>                                    <char>                  <char>       <num>
#>   1: 24337b53-36d5-4ff2-9ebb-ee64c565d899                    <NA>  -10.000000
#>   2: d16fbcd2-2d52-40f9-9a64-0ac97cb0b69b                    <NA>  -10.000000
#>   3: d57cd642-d164-4237-964c-978d861cb368                    <NA>  -10.000000
#>   4: bb304f6b-8ad2-425f-b934-a3333efe186d                    <NA>  -10.000000
#>   5: 4657de36-a381-487b-95ae-84a373aed609                    <NA>  -10.000000
#>   6: f2e02b83-2d88-4b8b-a6d0-2c96f19a0110                    <NA>  -10.000000
#>   7: aa066066-1094-45e6-91b4-6dece4edc416                    <NA>  -10.000000
#>   8: 2d6736ed-61e8-4d73-b6ff-49fa6a9b600a                    <NA>  -10.000000
#>   9: 9490ed33-aacb-4506-b903-4179f194b7c0                    <NA>  -10.000000
#>  10: 6ef8e2c7-e024-43c1-b954-4786ef2b9d54                    <NA>  -10.000000
#>  11: 0c77d0a9-4427-49bd-acfa-1b9e3f5d97c1                    <NA>   -7.777778
#>  12: fb825290-1a79-4a52-8b1d-e7ad3b21e9f3                    <NA>   -7.777778
#>  13: d57f0157-9d7a-45dc-8917-db98b57ff5ef                    <NA>   -7.777778
#>  14: a95ad81c-559b-4a7f-85e5-2cfdfc8fb426                    <NA>   -7.777778
#>  15: bcf2b16c-1f98-4789-97e8-754b934f4e31                    <NA>   -7.777778
#>  16: 2f790807-4776-405e-a136-304585ea2738                    <NA>   -7.777778
#>  17: 447ff324-7a1a-4071-b144-2f231f309b8e                    <NA>   -7.777778
#>  18: 4c6295ae-8fcf-4ef8-9789-042f98be224c                    <NA>   -7.777778
#>  19: 7225f0b2-69fb-4486-859c-0ef4535f5f94                    <NA>   -7.777778
#>  20: 35012d02-5004-4450-97bb-10598875adb8                    <NA>   -7.777778
#>  21: 48b57492-b466-43e2-bbbe-d24b04f97b86 Optimization terminated          NA
#>  22: 424a5a50-e372-46d4-8eda-e55f262249b7 Optimization terminated          NA
#>  23: cfc30406-3fb0-43b0-ab17-bb4433beb3c4 Optimization terminated          NA
#>  24: b9acef94-fcc9-4d84-b910-2e436f9c7e49 Optimization terminated          NA
#>  25: 27bc7d6e-9647-4a98-8777-6cefb75560bd Optimization terminated          NA
#>  26: 03512b32-86bb-41d2-9919-19b7856dd192 Optimization terminated          NA
#>  27: 571f4eb8-7fa5-4f77-8ab9-a39883208e4b Optimization terminated          NA
#>  28: cd1fd264-b442-4164-aee4-2a8dcde5a41f Optimization terminated          NA
#>  29: 93086f3b-553e-4ef3-abd1-7057b9da7472 Optimization terminated          NA
#>  30: f00213ba-6f03-4f37-8004-64806e8edb08 Optimization terminated          NA
#>  31: 708b8f9d-5dc0-4213-bc10-fcd73e6e396f Optimization terminated          NA
#>  32: 26bed61b-27e5-4737-8c75-1d2f7c0d1060 Optimization terminated          NA
#>  33: aebfa2e5-b8c6-4374-8fe8-65224774e1b2 Optimization terminated          NA
#>  34: c349a3df-67eb-40e1-b2c2-ae3ff6614eab Optimization terminated          NA
#>  35: 05e830e6-b3fb-4d57-85a5-0fd975ea03dc Optimization terminated          NA
#>  36: 1ef3cf9b-f842-4766-b73e-fe4bd96ce096 Optimization terminated          NA
#>  37: 7257442e-3200-476f-8dee-f48069b0ea0e Optimization terminated          NA
#>  38: c2bde311-bcc0-4c3c-bd4b-3a545309c1e8 Optimization terminated          NA
#>  39: 1df58d96-2daf-488a-906f-9deff4f627d6 Optimization terminated          NA
#>  40: eb1eed4f-3c99-4f87-a00a-48aaefde0439 Optimization terminated          NA
#>  41: 259c084a-40c3-4fd1-9988-8b8ba5bcef57 Optimization terminated          NA
#>  42: bd705f1e-44e5-4add-895f-4a970c267e4b Optimization terminated          NA
#>  43: d245c812-7eb2-48f6-a289-2e70ed16198b Optimization terminated          NA
#>  44: d99fb1eb-618d-4105-89ba-ab54765a4969 Optimization terminated          NA
#>  45: 060713ab-cee7-4e8f-98da-5fa7cd971803 Optimization terminated          NA
#>  46: 4b355eb1-bc4b-475c-bb56-8c5d71a884ee Optimization terminated          NA
#>  47: dc68a706-584f-4671-b6c3-f561ce82baae Optimization terminated          NA
#>  48: 9fc55d49-9eda-4d08-ada2-ef459d1e7214 Optimization terminated          NA
#>  49: d0505952-377d-48d0-8c18-02b06991df29 Optimization terminated          NA
#>  50: dd9a3d66-2125-4ca9-94eb-293739817df8 Optimization terminated          NA
#>  51: de419f7c-c69b-4be8-99d1-eac646d2fb14 Optimization terminated          NA
#>  52: 4c37b5b2-30d1-42c1-a2d4-257e7c312dca Optimization terminated          NA
#>  53: 289d5652-52b3-4220-b877-08d1b435efc4 Optimization terminated          NA
#>  54: bfa7dbea-97e8-42aa-ab87-a92ef86da8e8 Optimization terminated          NA
#>  55: a98a0d7c-645f-48d4-96aa-ed3b7e5a61fd Optimization terminated          NA
#>  56: 28ccf29c-93b5-4d7d-9a6b-8024c04091c3 Optimization terminated          NA
#>  57: c60b101f-89a2-4361-b35f-b3d064e5f6c1 Optimization terminated          NA
#>  58: bc181638-f045-4fff-bdbe-b3560510d9ab Optimization terminated          NA
#>  59: 1039a2a4-56d8-437b-a1e5-cea6059bf627 Optimization terminated          NA
#>  60: e525090f-0191-46fd-b297-0b7e0a26846c Optimization terminated          NA
#>  61: 9348ab42-9714-44a4-bc01-dcc99fc1090e Optimization terminated          NA
#>  62: 6e1abd9c-ee15-42b6-bf87-ba4fc73b60a9 Optimization terminated          NA
#>  63: 819804f7-9321-4eff-a128-3c70d8a3135f Optimization terminated          NA
#>  64: 5252bd58-8e51-4226-8eee-003696df64a9 Optimization terminated          NA
#>  65: 71f4a56f-11b1-4226-acc5-3aa6ff63d714 Optimization terminated          NA
#>  66: 92bdfcf3-af0e-4c36-97fa-cb487804cd8f Optimization terminated          NA
#>  67: f1ea5633-a83f-4517-84ac-d0b8eed59de1 Optimization terminated          NA
#>  68: 9ec015b1-67fe-448a-a2f9-4eaa7cc382f5 Optimization terminated          NA
#>  69: db450932-4a7f-48a7-80f4-3ccfeee2b7ac Optimization terminated          NA
#>  70: cb29b1b3-e2d1-4ac8-8946-c2f097b27cfa Optimization terminated          NA
#>  71: 4d178bb9-1327-4a1b-9743-45eedcaec8b2 Optimization terminated          NA
#>  72: 97461736-27b7-4704-aae8-50bb83cb1998 Optimization terminated          NA
#>  73: 22b15948-487b-4a0c-af22-90d12ae4ac26 Optimization terminated          NA
#>  74: d1528eb8-36ec-4377-9cbc-67f46e599774 Optimization terminated          NA
#>  75: 98149ec8-fc81-4f5d-950a-8869fd300c6e Optimization terminated          NA
#>  76: 66a36906-a449-4c6a-a5e5-7c4aa9828e38 Optimization terminated          NA
#>  77: c682472f-e9aa-4c83-8d8a-0da068238bac Optimization terminated          NA
#>  78: 4d00f8ae-17a9-4a9f-b4f9-fcdc952d2fc2 Optimization terminated          NA
#>  79: 850ef6ba-3c37-4606-8401-c6d3c51034c2 Optimization terminated          NA
#>  80: f9362325-1f68-4099-ad12-3d805bcaaef9 Optimization terminated          NA
#>  81: d1ca571a-e2e1-44f5-89c0-e3317eb3c316 Optimization terminated          NA
#>  82: 88e36813-4840-4063-bc50-3a02cede57f5 Optimization terminated          NA
#>  83: eb5dc894-5873-4ec1-935b-5e98435cb55e Optimization terminated          NA
#>  84: bf5f4c6e-8a6f-45b6-8dc0-571593fd7378 Optimization terminated          NA
#>  85: ffd91029-6d74-4968-9f75-346216b11b3f Optimization terminated          NA
#>  86: 7efd0405-266f-443a-87d8-edf4853c8916 Optimization terminated          NA
#>  87: 19a1e80b-f0d1-400d-a17d-10aefc01bad7 Optimization terminated          NA
#>  88: 41397e1a-d74b-40cd-aa3d-bf01d49d7635 Optimization terminated          NA
#>  89: 5edb140f-fb70-43af-8683-c2326ed04bd0 Optimization terminated          NA
#>  90: 47914d7f-547c-4f1b-9401-2db5efbcd519 Optimization terminated          NA
#>  91: ff14eb71-1a57-4217-bca1-199d876b61e4 Optimization terminated          NA
#>  92: 7b28610b-30a6-4d2d-8980-322581d9d2af Optimization terminated          NA
#>  93: 561ce4ed-307e-47c2-8526-807b68bfdf68 Optimization terminated          NA
#>  94: a6bc0745-1cc2-4551-bb57-b7c992f6f4f2 Optimization terminated          NA
#>  95: 2cacf2da-96db-45c5-ad46-8772265a6a9a Optimization terminated          NA
#>  96: d2cea274-5328-40f3-aae4-2ac3a26f1212 Optimization terminated          NA
#>  97: ba2d0c4e-eb0d-47dd-94ea-8c255ca57ca7 Optimization terminated          NA
#>  98: b196efa3-24b2-4719-907d-73dac6f590ae Optimization terminated          NA
#>  99: c493b628-d059-411b-bf98-3ebf7a8c5803 Optimization terminated          NA
#> 100: 40fcd091-6b69-4fab-bb22-3d2bb3511d95 Optimization terminated          NA
#>                                      keys                 message x_domain_x1
#>      x_domain_x2
#>            <num>
#>   1:  -5.0000000
#>   2:  -3.8888889
#>   3:  -2.7777778
#>   4:  -1.6666667
#>   5:  -0.5555556
#>   6:   0.5555556
#>   7:   1.6666667
#>   8:   2.7777778
#>   9:   3.8888889
#>  10:   5.0000000
#>  11:  -5.0000000
#>  12:  -3.8888889
#>  13:  -2.7777778
#>  14:  -1.6666667
#>  15:  -0.5555556
#>  16:   0.5555556
#>  17:   1.6666667
#>  18:   2.7777778
#>  19:   3.8888889
#>  20:   5.0000000
#>  21:          NA
#>  22:          NA
#>  23:          NA
#>  24:          NA
#>  25:          NA
#>  26:          NA
#>  27:          NA
#>  28:          NA
#>  29:          NA
#>  30:          NA
#>  31:          NA
#>  32:          NA
#>  33:          NA
#>  34:          NA
#>  35:          NA
#>  36:          NA
#>  37:          NA
#>  38:          NA
#>  39:          NA
#>  40:          NA
#>  41:          NA
#>  42:          NA
#>  43:          NA
#>  44:          NA
#>  45:          NA
#>  46:          NA
#>  47:          NA
#>  48:          NA
#>  49:          NA
#>  50:          NA
#>  51:          NA
#>  52:          NA
#>  53:          NA
#>  54:          NA
#>  55:          NA
#>  56:          NA
#>  57:          NA
#>  58:          NA
#>  59:          NA
#>  60:          NA
#>  61:          NA
#>  62:          NA
#>  63:          NA
#>  64:          NA
#>  65:          NA
#>  66:          NA
#>  67:          NA
#>  68:          NA
#>  69:          NA
#>  70:          NA
#>  71:          NA
#>  72:          NA
#>  73:          NA
#>  74:          NA
#>  75:          NA
#>  76:          NA
#>  77:          NA
#>  78:          NA
#>  79:          NA
#>  80:          NA
#>  81:          NA
#>  82:          NA
#>  83:          NA
#>  84:          NA
#>  85:          NA
#>  86:          NA
#>  87:          NA
#>  88:          NA
#>  89:          NA
#>  90:          NA
#>  91:          NA
#>  92:          NA
#>  93:          NA
#>  94:          NA
#>  95:          NA
#>  96:          NA
#>  97:          NA
#>  98:          NA
#>  99:          NA
#> 100:          NA
#>      x_domain_x2