Optimization via Covariance Matrix Adaptation Evolution Strategy

OptimizerCmaes class that implements CMA-ES. Calls adagio::pureCMAES() from package adagio. The algorithm is typically applied to search space dimensions between three and fifty. Lower search space dimensions might crash.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("cmaes")
opt("cmaes")

Parameters

sigma

numeric(1)

start_values

character(1)
Create random start values or based on center of search space? In the latter case, it is the center of the parameters before a trafo is applied.

For the meaning of the control parameters, see adagio::pureCMAES(). Note that we have removed all control parameters which refer to the termination of the algorithm and where our terminators allow to obtain the same behavior.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerCmaes

Methods

Public methods

OptimizerCmaes$new()

OptimizerCmaes$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerCmaes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerCmaes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("adagio")) {
  search_space = domain = ps(
    x1 = p_dbl(-10, 10),
    x2 = p_dbl(-5, 5)
  )

  codomain = ps(y = p_dbl(tags = "maximize"))

  objective_function = function(xs) {
    c(y = -(xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10)
  }

  objective = ObjectiveRFun$new(
    fun = objective_function,
    domain = domain,
    codomain = codomain)

  instance = OptimInstanceSingleCrit$new(
    objective = objective,
    search_space = search_space,
    terminator = trm("evals", n_evals = 10))

  optimizer = opt("cmaes")

  # modifies the instance by reference
  optimizer$optimize(instance)

  # returns best scoring evaluation
  instance$result

  # allows access of data.table of full path of all evaluations
  as.data.table(instance$archive$data)
}
#>              x1         x2           y  x_domain           timestamp batch_nr
#>  1:  10.0000000  5.0000000 -118.000000 <list[2]> 2022-08-30 04:40:35        1
#>  2:  10.0000000  5.0000000 -118.000000 <list[2]> 2022-08-30 04:40:35        2
#>  3: -10.0000000 -4.5223573 -136.317572 <list[2]> 2022-08-30 04:40:35        3
#>  4:  -1.6685656 -1.7753813   -4.958064 <list[2]> 2022-08-30 04:40:35        4
#>  5:  -2.8399838  5.0000000  -77.425444 <list[2]> 2022-08-30 04:40:35        5
#>  6:   0.3916698  0.4802594   -4.698931 <list[2]> 2022-08-30 04:40:35        6
#>  7:  -1.4419089  4.3076930  -55.249114 <list[2]> 2022-08-30 04:40:35        7
#>  8:   7.6011385  3.8690383  -68.556441 <list[2]> 2022-08-30 04:40:35        8
#>  9:   8.6213707 -5.0000000  -37.842550 <list[2]> 2022-08-30 04:40:35        9
#> 10:   1.3461954 -0.2951200    2.256164 <list[2]> 2022-08-30 04:40:35       10