Optimization via Focus Search

OptimizerFocusSearch class that implements a Focus Search.

Focus Search starts with evaluating n_points drawn uniformly at random. For 1 to maxit batches, n_points are then drawn uniformly at random and if the best value of a batch outperforms the previous best value over all batches evaluated so far, the search space is shrinked around this new best point prior to the next batch being sampled and evaluated.

For details on the shrinking, see shrink_ps.

Depending on the Terminator this procedure simply restarts after maxit is reached.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("focus_search")
opt("focus_search")

Parameters

n_points

integer(1)
Number of points to evaluate in each random search batch.

maxit

integer(1)
Number of random search batches to run.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerFocusSearch

Methods

Public methods

OptimizerFocusSearch$new()

OptimizerFocusSearch$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerFocusSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerFocusSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

search_space = domain = ps(x = p_dbl(lower = -1, upper = 1))

codomain = ps(y = p_dbl(tags = "minimize"))

objective_function = function(xs) {
  list(y = as.numeric(xs)^2)
}

objective = ObjectiveRFun$new(
 fun = objective_function,
 domain = domain,
 codomain = codomain)

instance = OptimInstanceSingleCrit$new(
 objective = objective,
 search_space = search_space,
 terminator = trm("evals", n_evals = 10))

optimizer = opt("focus_search")

# modifies the instance by reference
optimizer$optimize(instance)
#>               x  x_domain            y
#> 1: -0.005347974 <list[1]> 2.860082e-05

# returns best scoring evaluation
instance$result
#>               x  x_domain            y
#> 1: -0.005347974 <list[1]> 2.860082e-05

# allows access of data.table of full path of all evaluations
as.data.table(instance$archive$data)
#>                 x            y  x_domain           timestamp batch_nr
#>   1:  0.988794926 9.777154e-01 <list[1]> 2022-06-27 04:34:44        1
#>   2: -0.643205152 4.137129e-01 <list[1]> 2022-06-27 04:34:44        1
#>   3: -0.319427189 1.020337e-01 <list[1]> 2022-06-27 04:34:44        1
#>   4: -0.582575131 3.393938e-01 <list[1]> 2022-06-27 04:34:44        1
#>   5:  0.249045049 6.202344e-02 <list[1]> 2022-06-27 04:34:44        1
#>   6: -0.689162817 4.749454e-01 <list[1]> 2022-06-27 04:34:44        1
#>   7: -0.266462585 7.100231e-02 <list[1]> 2022-06-27 04:34:44        1
#>   8:  0.732198047 5.361140e-01 <list[1]> 2022-06-27 04:34:44        1
#>   9:  0.743261985 5.524384e-01 <list[1]> 2022-06-27 04:34:44        1
#>  10: -0.544540997 2.965249e-01 <list[1]> 2022-06-27 04:34:44        1
#>  11:  0.462666893 2.140607e-01 <list[1]> 2022-06-27 04:34:44        1
#>  12: -0.256596710 6.584187e-02 <list[1]> 2022-06-27 04:34:44        1
#>  13: -0.216467205 4.685805e-02 <list[1]> 2022-06-27 04:34:44        1
#>  14:  0.426875492 1.822227e-01 <list[1]> 2022-06-27 04:34:44        1
#>  15:  0.642189450 4.124073e-01 <list[1]> 2022-06-27 04:34:44        1
#>  16:  0.858130443 7.363879e-01 <list[1]> 2022-06-27 04:34:44        1
#>  17:  0.487109843 2.372760e-01 <list[1]> 2022-06-27 04:34:44        1
#>  18:  0.778723501 6.064103e-01 <list[1]> 2022-06-27 04:34:44        1
#>  19:  0.731063136 5.344533e-01 <list[1]> 2022-06-27 04:34:44        1
#>  20: -0.909454389 8.271073e-01 <list[1]> 2022-06-27 04:34:44        1
#>  21:  0.942973488 8.891990e-01 <list[1]> 2022-06-27 04:34:44        1
#>  22:  0.684349730 4.683346e-01 <list[1]> 2022-06-27 04:34:44        1
#>  23: -0.924017855 8.538090e-01 <list[1]> 2022-06-27 04:34:44        1
#>  24: -0.059514483 3.541974e-03 <list[1]> 2022-06-27 04:34:44        1
#>  25: -0.651948042 4.250362e-01 <list[1]> 2022-06-27 04:34:44        1
#>  26: -0.898940090 8.080933e-01 <list[1]> 2022-06-27 04:34:44        1
#>  27:  0.046134758 2.128416e-03 <list[1]> 2022-06-27 04:34:44        1
#>  28: -0.900437398 8.107875e-01 <list[1]> 2022-06-27 04:34:44        1
#>  29: -0.524446325 2.750439e-01 <list[1]> 2022-06-27 04:34:44        1
#>  30:  0.591945304 3.503992e-01 <list[1]> 2022-06-27 04:34:44        1
#>  31: -0.519194471 2.695629e-01 <list[1]> 2022-06-27 04:34:44        1
#>  32:  0.308883809 9.540921e-02 <list[1]> 2022-06-27 04:34:44        1
#>  33: -0.508389049 2.584594e-01 <list[1]> 2022-06-27 04:34:44        1
#>  34: -0.887805967 7.881994e-01 <list[1]> 2022-06-27 04:34:44        1
#>  35:  0.568625764 3.233353e-01 <list[1]> 2022-06-27 04:34:44        1
#>  36:  0.049287209 2.429229e-03 <list[1]> 2022-06-27 04:34:44        1
#>  37:  0.814040336 6.626617e-01 <list[1]> 2022-06-27 04:34:44        1
#>  38: -0.937461446 8.788340e-01 <list[1]> 2022-06-27 04:34:44        1
#>  39:  0.849844622 7.222359e-01 <list[1]> 2022-06-27 04:34:44        1
#>  40: -0.104786258 1.098016e-02 <list[1]> 2022-06-27 04:34:44        1
#>  41: -0.666008191 4.435669e-01 <list[1]> 2022-06-27 04:34:44        1
#>  42: -0.397938826 1.583553e-01 <list[1]> 2022-06-27 04:34:44        1
#>  43: -0.634107999 4.020930e-01 <list[1]> 2022-06-27 04:34:44        1
#>  44:  0.439988685 1.935900e-01 <list[1]> 2022-06-27 04:34:44        1
#>  45:  0.383013293 1.466992e-01 <list[1]> 2022-06-27 04:34:44        1
#>  46:  0.916632806 8.402157e-01 <list[1]> 2022-06-27 04:34:44        1
#>  47:  0.556437775 3.096230e-01 <list[1]> 2022-06-27 04:34:44        1
#>  48:  0.694487559 4.823130e-01 <list[1]> 2022-06-27 04:34:44        1
#>  49:  0.305580721 9.337958e-02 <list[1]> 2022-06-27 04:34:44        1
#>  50: -0.436217753 1.902859e-01 <list[1]> 2022-06-27 04:34:44        1
#>  51:  0.788724451 6.220863e-01 <list[1]> 2022-06-27 04:34:44        1
#>  52:  0.246241644 6.063495e-02 <list[1]> 2022-06-27 04:34:44        1
#>  53:  0.999255748 9.985120e-01 <list[1]> 2022-06-27 04:34:44        1
#>  54:  0.575017921 3.306456e-01 <list[1]> 2022-06-27 04:34:44        1
#>  55: -0.027979133 7.828319e-04 <list[1]> 2022-06-27 04:34:44        1
#>  56: -0.309296291 9.566420e-02 <list[1]> 2022-06-27 04:34:44        1
#>  57: -0.589286114 3.472581e-01 <list[1]> 2022-06-27 04:34:44        1
#>  58:  0.998966004 9.979331e-01 <list[1]> 2022-06-27 04:34:44        1
#>  59: -0.484945762 2.351724e-01 <list[1]> 2022-06-27 04:34:44        1
#>  60:  0.955127982 9.122695e-01 <list[1]> 2022-06-27 04:34:44        1
#>  61: -0.067859215 4.604873e-03 <list[1]> 2022-06-27 04:34:44        1
#>  62:  0.538433058 2.899102e-01 <list[1]> 2022-06-27 04:34:44        1
#>  63: -0.183570156 3.369800e-02 <list[1]> 2022-06-27 04:34:44        1
#>  64:  0.928798341 8.626664e-01 <list[1]> 2022-06-27 04:34:44        1
#>  65:  0.335395574 1.124902e-01 <list[1]> 2022-06-27 04:34:44        1
#>  66: -0.419000329 1.755613e-01 <list[1]> 2022-06-27 04:34:44        1
#>  67:  0.742379311 5.511270e-01 <list[1]> 2022-06-27 04:34:44        1
#>  68:  0.505936020 2.559713e-01 <list[1]> 2022-06-27 04:34:44        1
#>  69: -0.864093757 7.466580e-01 <list[1]> 2022-06-27 04:34:44        1
#>  70:  0.833150215 6.941393e-01 <list[1]> 2022-06-27 04:34:44        1
#>  71: -0.750437466 5.631564e-01 <list[1]> 2022-06-27 04:34:44        1
#>  72: -0.906219652 8.212341e-01 <list[1]> 2022-06-27 04:34:44        1
#>  73: -0.361936919 1.309983e-01 <list[1]> 2022-06-27 04:34:44        1
#>  74: -0.414612410 1.719035e-01 <list[1]> 2022-06-27 04:34:44        1
#>  75: -0.061817850 3.821447e-03 <list[1]> 2022-06-27 04:34:44        1
#>  76: -0.274514193 7.535804e-02 <list[1]> 2022-06-27 04:34:44        1
#>  77: -0.765339155 5.857440e-01 <list[1]> 2022-06-27 04:34:44        1
#>  78: -0.382955718 1.466551e-01 <list[1]> 2022-06-27 04:34:44        1
#>  79:  0.343538433 1.180187e-01 <list[1]> 2022-06-27 04:34:44        1
#>  80:  0.249109353 6.205547e-02 <list[1]> 2022-06-27 04:34:44        1
#>  81:  0.680605243 4.632235e-01 <list[1]> 2022-06-27 04:34:44        1
#>  82:  0.666973344 4.448534e-01 <list[1]> 2022-06-27 04:34:44        1
#>  83: -0.055262392 3.053932e-03 <list[1]> 2022-06-27 04:34:44        1
#>  84:  0.522696081 2.732112e-01 <list[1]> 2022-06-27 04:34:44        1
#>  85: -0.586716135 3.442358e-01 <list[1]> 2022-06-27 04:34:44        1
#>  86: -0.572567174 3.278332e-01 <list[1]> 2022-06-27 04:34:44        1
#>  87: -0.370970604 1.376192e-01 <list[1]> 2022-06-27 04:34:44        1
#>  88:  0.175855352 3.092510e-02 <list[1]> 2022-06-27 04:34:44        1
#>  89:  0.247455242 6.123410e-02 <list[1]> 2022-06-27 04:34:44        1
#>  90: -0.541469201 2.931889e-01 <list[1]> 2022-06-27 04:34:44        1
#>  91: -0.349033887 1.218247e-01 <list[1]> 2022-06-27 04:34:44        1
#>  92: -0.005347974 2.860082e-05 <list[1]> 2022-06-27 04:34:44        1
#>  93:  0.456425790 2.083245e-01 <list[1]> 2022-06-27 04:34:44        1
#>  94: -0.914447893 8.362149e-01 <list[1]> 2022-06-27 04:34:44        1
#>  95: -0.961359676 9.242124e-01 <list[1]> 2022-06-27 04:34:44        1
#>  96: -0.687402259 4.725219e-01 <list[1]> 2022-06-27 04:34:44        1
#>  97:  0.368699505 1.359393e-01 <list[1]> 2022-06-27 04:34:44        1
#>  98: -0.364322165 1.327306e-01 <list[1]> 2022-06-27 04:34:44        1
#>  99:  0.226093679 5.111835e-02 <list[1]> 2022-06-27 04:34:44        1
#> 100:  0.490184051 2.402804e-01 <list[1]> 2022-06-27 04:34:44        1
#>                 x            y  x_domain           timestamp batch_nr