Optimization via Generalized Simulated Annealing

OptimizerGenSA class that implements generalized simulated annealing. Calls GenSA::GenSA() from package GenSA.

Source

Tsallis C, Stariolo DA (1996). “Generalized simulated annealing.” Physica A: Statistical Mechanics and its Applications, 233(1-2), 395--406. doi: 10.1016/s0378-4371(96)00271-3 .

Xiang Y, Gubian S, Suomela B, Hoeng J (2013). “Generalized Simulated Annealing for Global Optimization: The GenSA Package.” The R Journal, 5(1), 13. doi: 10.32614/rj-2013-002 .

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("gensa")
opt("gensa")

Parameters

smooth

logical(1)

temperature

numeric(1)

acceptance.param

numeric(1)

verbose

logical(1)

trace.mat

logical(1)

For the meaning of the control parameters, see GenSA::GenSA(). Note that we have removed all control parameters which refer to the termination of the algorithm and where our terminators allow to obtain the same behavior.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerGenSA

Methods

Public methods

OptimizerGenSA$new()

OptimizerGenSA$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerGenSA$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerGenSA$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("GenSA")) {

  search_space = domain = ps(x = p_dbl(lower = -1, upper = 1))

  codomain = ps(y = p_dbl(tags = "minimize"))

  objective_function = function(xs) {
    list(y = as.numeric(xs)^2)
  }

  objective = ObjectiveRFun$new(
    fun = objective_function,
    domain = domain,
    codomain = codomain)

  instance = OptimInstanceSingleCrit$new(
    objective = objective,
    search_space = search_space,
    terminator = trm("evals", n_evals = 10))

  optimizer = opt("cmaes")

  # Modifies the instance by reference
  optimizer$optimize(instance)

  # Returns best scoring evaluation
  instance$result

  # Allows access of data.table of full path of all evaluations
  as.data.table(instance$archive$data)
}
#> Loading required namespace: GenSA
#> Warning: CMA-ES is typically applied to search space dimensions between three and fifty. A lower search space dimension might crash.
#>                x            y  x_domain           timestamp batch_nr
#>  1: -0.334652425 1.119922e-01 <list[1]> 2021-12-07 04:13:26        1
#>  2:  0.075190920 5.653674e-03 <list[1]> 2021-12-07 04:13:27        2
#>  3: -0.867123477 7.519031e-01 <list[1]> 2021-12-07 04:13:27        3
#>  4:  0.747398911 5.586051e-01 <list[1]> 2021-12-07 04:13:27        4
#>  5:  0.881847150 7.776544e-01 <list[1]> 2021-12-07 04:13:27        5
#>  6:  0.287015016 8.237762e-02 <list[1]> 2021-12-07 04:13:27        6
#>  7: -1.000000000 1.000000e+00 <list[1]> 2021-12-07 04:13:27        7
#>  8:  0.670012283 4.489165e-01 <list[1]> 2021-12-07 04:13:27        8
#>  9: -0.552093111 3.048068e-01 <list[1]> 2021-12-07 04:13:27        9
#> 10: -0.005578496 3.111962e-05 <list[1]> 2021-12-07 04:13:27       10