Optimization via Iterated Racing

OptimizerBatchIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43–58. doi:10.1016/j.orp.2016.09.002 .

Parameters

instances

list()
A list of instances where the configurations executed on.

targetRunnerParallel

function()
A function that executes the objective function with a specific parameter configuration and instance. A default function is provided, see section "Target Runner and Instances".

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with OptimizerBatchIrace.

In contrast to irace::defaultScenario(), we set digits = 15. This represents double parameters with a higher precision and avoids rounding errors.

Target Runner and Instances

The irace package uses a targetRunner script or R function to evaluate a configuration on a particular instance. Usually it is not necessary to specify a targetRunner function when using OptimizerBatchIrace. A default function is used that forwards several configurations and instances to the user defined objective function. As usually, the user defined function has a xs, xss or xdt parameter depending on the used Objective class. For irace, the function needs an additional instances parameter.

fun = function(xs, instances) {
 # function to evaluate configuration in `xs` on instance `instances`
}

Archive

The Archive holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The optimization result (instance$result) is the best performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("irace")
opt("irace")

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super classes

bbotk::Optimizer -> bbotk::OptimizerBatch -> OptimizerBatchIrace

Methods

Public methods

OptimizerBatchIrace$new()

OptimizerBatchIrace$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$print()

bbotk::OptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerBatchIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerBatchIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15)
)

codomain = ps(y = p_dbl(tags = "minimize"))

# branin function with noise
# the noise generates different instances of the branin function
# the noise values are passed via the `instances` parameter
fun = function(xdt, instances) {
  ys = branin(xdt[["x1"]], xdt[["x2"]], noise = as.numeric(instances))
  data.table(y = ys)
}

# define objective with instances as a constant
objective = ObjectiveRFunDt$new(
 fun = fun,
 domain = domain,
 codomain = codomain,
 constants = ps(instances = p_uty()))

instance = OptimInstanceBatchSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 1000))

# create instances of branin function
instances = rnorm(10, mean = 0, sd = 0.1)

# load optimizer irace and set branin instances
optimizer = opt("irace", instances = instances)

# modifies the instance by reference
optimizer$optimize(instance)
#> # 2024-11-26 15:22:50 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 3
#> # minNbSurvival: 3
#> # nbParameters: 2
#> # seed: 2011091433
#> # confidence level: 0.95
#> # budget: 1000
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2024-11-26 15:22:50 UTC: Iteration 1 of 3
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 1000
#> # currentBudget: 333
#> # nbConfigurations: 55
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          1|         55|         31|     1.539374348|         55|00:00:00|   NA|  NA|    NA|
#> |x|          2|         55|         31|     1.539374348|        110|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         55|         31|     1.539374348|        165|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         55|         31|     1.539374348|        220|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          1|         31|     1.539374348|        275|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          31    mean value:      1.539374348
#> Description of the best-so-far configuration:
#>    .ID.               x1               x2 .PARENT.
#> 31   31 9.50708507443778 3.59838451375253       NA
#> 
#> # 2024-11-26 15:22:51 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  x1               x2
#> 31 9.50708507443778 3.59838451375253
#> # 2024-11-26 15:22:52 UTC: Iteration 2 of 3
#> # experimentsUsedSoFar: 275
#> # remainingBudget: 725
#> # currentBudget: 362
#> # nbConfigurations: 52
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          6|         52|         93|    0.4198923247|         52|00:00:00|   NA|  NA|    NA|
#> |x|          1|         52|         93|    0.4198923247|        103|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|         52|         93|    0.4198923247|        154|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         52|         93|    0.4198923247|        205|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          2|         93|    0.4198923247|        256|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          31    mean value:      1.539374348
#> Description of the best-so-far configuration:
#>    .ID.               x1               x2 .PARENT.
#> 31   31 9.50708507443778 3.59838451375253       NA
#> 
#> # 2024-11-26 15:22:53 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  x1               x2
#> 31 9.50708507443778 3.59838451375253
#> 93 9.39726519810422 2.58742994785361
#> # 2024-11-26 15:22:54 UTC: Iteration 3 of 3
#> # experimentsUsedSoFar: 531
#> # remainingBudget: 469
#> # currentBudget: 469
#> # nbConfigurations: 60
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          7|         60|        161|    0.4012148257|         60|00:00:00|   NA|  NA|    NA|
#> |x|          6|         60|        161|    0.4012148257|        118|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         60|        161|    0.4012148257|        177|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         60|        161|    0.4012148257|        235|00:00:00|+1.00|1.00|0.0000|
#> |-|          1|          3|        161|    0.4012148257|        293|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          93    mean value:     0.4198923247
#> Description of the best-so-far configuration:
#>    .ID.               x1               x2 .PARENT.
#> 93   93 9.39726519810422 2.58742994785361       31
#> 
#> # 2024-11-26 15:22:55 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                   x1               x2
#> 93  9.39726519810422 2.58742994785361
#> 31  9.50708507443778 3.59838451375253
#> 161 9.44086545535470 2.53426478010450
#> # 2024-11-26 15:22:55 UTC: Iteration 4 of 4
#> # experimentsUsedSoFar: 824
#> # remainingBudget: 176
#> # currentBudget: 176
#> # nbConfigurations: 21
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          8|         21|        165|    0.3979001066|         21|00:00:00|   NA|  NA|    NA|
#> |x|          6|         21|        165|    0.3979001066|         39|00:00:00|+1.00|1.00|0.0000|
#> |x|          7|         21|        165|    0.3979001066|         57|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         21|        165|    0.3979001066|         75|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          4|        165|    0.3979001066|         94|00:00:00|+1.00|1.00|0.0000|
#> |!|          2|          4|        165|    0.3979001066|         96|00:00:00|+1.00|1.00|0.0000|
#> |!|          1|          4|        165|    0.3979001066|         97|00:00:00|+1.00|1.00|0.0000|
#> |-|          3|          1|        165|    0.3979001066|         98|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         165    mean value:     0.3979001066
#> Description of the best-so-far configuration:
#>     .ID.               x1               x2 .PARENT.
#> 165  165 9.42640309724388 2.47663375212708      161
#> 
#> # 2024-11-26 15:22:56 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                   x1               x2
#> 165 9.42640309724388 2.47663375212708
#> # 2024-11-26 15:22:57 UTC: Iteration 5 of 5
#> # experimentsUsedSoFar: 922
#> # remainingBudget: 78
#> # currentBudget: 78
#> # nbConfigurations: 8
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          9|          8|        165|    0.3979001066|          8|00:00:00|   NA|  NA|    NA|
#> |x|          8|          8|        165|    0.3979001066|         15|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          8|        165|    0.3979001066|         22|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|          8|        165|    0.3979001066|         29|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          1|        165|    0.3979001066|         36|00:00:00|   NA|  NA|    NA|
#> |.|          7|          1|        165|    0.3979001066|         36|00:00:00|   NA|  NA|    NA|
#> |.|          3|          1|        165|    0.3979001066|         36|00:00:00|   NA|  NA|    NA|
#> |.|          1|          1|        165|    0.3979001066|         36|00:00:00|   NA|  NA|    NA|
#> |.|          6|          1|        165|    0.3979001066|         36|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         165    mean value:     0.3979001066
#> Description of the best-so-far configuration:
#>     .ID.               x1               x2 .PARENT.
#> 165  165 9.42640309724388 2.47663375212708      161
#> 
#> # 2024-11-26 15:22:57 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                   x1               x2
#> 165 9.42640309724388 2.47663375212708
#> # 2024-11-26 15:22:57 UTC: Iteration 6 of 6
#> # experimentsUsedSoFar: 958
#> # remainingBudget: 42
#> # currentBudget: 42
#> # nbConfigurations: 5
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|         10|          5|        165|    0.3979001066|          5|00:00:00|   NA|  NA|    NA|
#> |x|          6|          5|        165|    0.3979001066|          9|00:00:00|+1.00|1.00|0.0000|
#> |x|          8|          5|        165|    0.3979001066|         13|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          5|        165|    0.3979001066|         17|00:00:00|+1.00|1.00|0.0000|
#> |-|          7|          1|        165|    0.3979001066|         21|00:00:00|   NA|  NA|    NA|
#> |.|          2|          1|        165|    0.3979001066|         21|00:00:00|   NA|  NA|    NA|
#> |.|          5|          1|        165|    0.3979001066|         21|00:00:00|   NA|  NA|    NA|
#> |.|          3|          1|        165|    0.3979001066|         21|00:00:00|   NA|  NA|    NA|
#> |.|          1|          1|        165|    0.3979001066|         21|00:00:00|   NA|  NA|    NA|
#> |.|          9|          1|        165|    0.3979001066|         21|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         165    mean value:     0.3979001066
#> Description of the best-so-far configuration:
#>     .ID.               x1               x2 .PARENT.
#> 165  165 9.42640309724388 2.47663375212708      161
#> 
#> # 2024-11-26 15:22:58 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                   x1               x2
#> 165 9.42640309724388 2.47663375212708
#> # 2024-11-26 15:22:58 UTC: Stopped because there is not enough budget left to race more than the minimum (3)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 7
#> # nbIterations: 7
#> # experimentsUsedSoFar: 979
#> # timeUsed: 0
#> # remainingBudget: 21
#> # currentBudget: 21
#> # number of elites: 1
#> # nbConfigurations: 2
#> # Total CPU user time: 8.627, CPU sys time: 0.048, Wall-clock time: 8.676
#>          x1       x2 configuration  x_domain         y
#>       <num>    <num>         <num>    <list>     <num>
#> 1: 9.426403 2.476634           165 <list[2]> 0.3979001

# best scoring configuration
instance$result
#>          x1       x2 configuration  x_domain         y
#>       <num>    <num>         <num>    <list>     <num>
#> 1: 9.426403 2.476634           165 <list[2]> 0.3979001

# all evaluations
as.data.table(instance$archive)
#>             x1        x2           y  race  step instance configuration
#>          <num>     <num>       <num> <num> <int>    <int>         <num>
#>   1:  4.828025  2.003804  11.5656272     1     1        2             1
#>   2: -2.300947  5.169679  30.3900663     1     1        2             2
#>   3:  2.256502 14.607673 137.1196519     1     1        2             3
#>   4:  1.762303 13.011215  96.8107637     1     1        2             4
#>   5:  3.613208  2.504441   1.7692893     1     1        2             5
#>  ---                                                                   
#> 975:  9.423526  2.479131   0.3979218     6     1        9           193
#> 976:  9.434351  2.482994   0.3983273     6     1        6           190
#> 977:  9.424986  2.470257   0.3979118     6     1        6           191
#> 978:  9.436607  2.480188   0.3985823     6     1        6           192
#> 979:  9.423526  2.479131   0.3979218     6     1        6           193
#>                timestamp batch_nr
#>                   <POSc>    <int>
#>   1: 2024-11-26 15:22:50        1
#>   2: 2024-11-26 15:22:50        1
#>   3: 2024-11-26 15:22:50        1
#>   4: 2024-11-26 15:22:50        1
#>   5: 2024-11-26 15:22:50        1
#>  ---                             
#> 975: 2024-11-26 15:22:58       32
#> 976: 2024-11-26 15:22:58       33
#> 977: 2024-11-26 15:22:58       33
#> 978: 2024-11-26 15:22:58       33
#> 979: 2024-11-26 15:22:58       33
# }