Optimization via Iterated Racing

OptimizerIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi:10.1016/j.orp.2016.09.002 .

Parameters

instances

list()
A list of instances where the configurations executed on.

targetRunnerParallel

function()
A function that executes the objective function with a specific parameter configuration and instance. A default function is provided, see section "Target Runner and Instances".

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with OptimizerIrace.

Target Runner and Instances

The irace package uses a targetRunner script or R function to evaluate a configuration on a particular instance. Usually it is not necessary to specify a targetRunner function when using OptimizerIrace. A default function is used that forwards several configurations and instances to the user defined objective function. As usually, the user defined function has a xs, xss or xdt parameter depending on the used Objective class. For irace, the function needs an additional instances parameter.

fun = function(xs, instances) {
 # function to evaluate configuration in `xs` on instance `instances`
}

Archive

The Archive holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The optimization result (instance$result) is the best performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("irace")
opt("irace")

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerIrace

Methods

Public methods

OptimizerIrace$new()

OptimizerIrace$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15)
)

codomain = ps(y = p_dbl(tags = "minimize"))

# branin function with noise
# the noise generates different instances of the branin function
# the noise values are passed via the `instances` parameter
fun = function(xdt, instances) {
  a = 1
  b = 5.1 / (4 * (pi^2))
  c = 5 / pi
  r = 6
  s = 10
  t = 1 / (8 * pi)

  data.table(y = (
    a * ((xdt[["x2"]] -
      b * (xdt[["x1"]]^2L) +
      c * xdt[["x1"]] - r)^2) +
      ((s * (1 - t)) * cos(xdt[["x1"]])) +
      unlist(instances)))
}

objective = ObjectiveRFunDt$new(fun = fun, domain = domain, codomain = codomain)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 1000))

# create instances of branin function
instances = rnorm(10, mean = 0, sd = 0.1)

# load optimizer irace and set branin instances
optimizer = opt("irace", instances = instances)

# modifies the instance by reference
optimizer$optimize(instance)
#> # 2022-08-27 04:35:50 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 3
#> # minNbSurvival: 3
#> # nbParameters: 2
#> # seed: 1140600263
#> # confidence level: 0.95
#> # budget: 1000
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2022-08-27 04:35:50 UTC: Iteration 1 of 3
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 1000
#> # currentBudget: 333
#> # nbConfigurations: 55
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          1|         55|         27|   -8.811997825|         55|00:00:00|   NA|  NA|    NA|
#> |x|          2|         55|         27|   -8.747708166|        110|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         55|         27|   -8.775657410|        165|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         55|         27|   -8.785434276|        220|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          1|         27|   -8.792197170|        275|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          27    mean value:    -8.792197170
#> Description of the best-so-far configuration:
#>    .ID.     x1     x2 .PARENT.
#> 27   27 9.1215 2.8026       NA
#> 
#> # 2022-08-27 04:35:51 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>        x1     x2
#> 27 9.1215 2.8026
#> # 2022-08-27 04:35:51 UTC: Iteration 2 of 3
#> # experimentsUsedSoFar: 275
#> # remainingBudget: 725
#> # currentBudget: 362
#> # nbConfigurations: 52
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          6|         52|         94|   -9.681249931|         52|00:00:00|   NA|  NA|    NA|
#> |x|          5|         52|         94|   -9.609821777|        103|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|         52|         94|   -9.583595419|        154|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         52|         94|   -9.575371758|        205|00:00:00|+1.00|1.00|0.0000|
#> |-|          2|          2|         94|   -9.540810083|        256|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          27    mean value:    -8.820515150
#> Description of the best-so-far configuration:
#>    .ID.     x1     x2 .PARENT.
#> 27   27 9.1215 2.8026       NA
#> 
#> # 2022-08-27 04:35:51 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>        x1     x2
#> 27 9.1215 2.8026
#> 94 9.5168 2.4823
#> # 2022-08-27 04:35:51 UTC: Iteration 3 of 3
#> # experimentsUsedSoFar: 531
#> # remainingBudget: 469
#> # currentBudget: 469
#> # nbConfigurations: 60
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          7|         60|        131|   -9.592628287|         60|00:00:00|   NA|  NA|    NA|
#> |x|          2|         60|        131|   -9.519567294|        118|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         60|        131|   -9.538995751|        177|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|         60|        131|   -9.585545025|        235|00:00:00|+1.00|1.00|0.0000|
#> |-|          3|          3|        131|   -9.587364758|        293|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          94    mean value:    -9.540949363
#> Description of the best-so-far configuration:
#>    .ID.     x1     x2 .PARENT.
#> 94   94 9.5168 2.4823       27
#> 
#> # 2022-08-27 04:35:52 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1     x2
#> 94  9.5168 2.4823
#> 27  9.1215 2.8026
#> 131 9.4341 2.4459
#> # 2022-08-27 04:35:52 UTC: Iteration 4 of 4
#> # experimentsUsedSoFar: 824
#> # remainingBudget: 176
#> # currentBudget: 176
#> # nbConfigurations: 21
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          8|         21|        167|   -9.526946138|         21|00:00:00|   NA|  NA|    NA|
#> |x|          7|         21|        167|   -9.559825690|         39|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         21|        167|   -9.565860333|         57|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|         21|        167|   -9.569998623|         76|00:00:00|+1.00|1.00|0.0000|
#> |-|          1|          4|        167|   -9.571031413|         95|00:00:00|+1.00|1.00|0.0000|
#> |!|          2|          4|        167|   -9.550290053|         96|00:00:00|+1.00|1.00|0.0000|
#> |!|          6|          4|        167|   -9.575287160|         97|00:00:00|+1.00|1.00|0.0000|
#> |-|          3|          1|        167|   -9.577716345|         98|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         167    mean value:    -9.577716345
#> Description of the best-so-far configuration:
#>     .ID.     x1     x2 .PARENT.
#> 167  167 9.4214 2.4315      131
#> 
#> # 2022-08-27 04:35:52 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1     x2
#> 167 9.4214 2.4315
#> # 2022-08-27 04:35:52 UTC: Iteration 5 of 5
#> # experimentsUsedSoFar: 922
#> # remainingBudget: 78
#> # currentBudget: 78
#> # nbConfigurations: 8
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          9|          8|        183|   -9.584124314|          8|00:00:00|   NA|  NA|    NA|
#> |x|          7|          8|        183|   -9.588600956|         15|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|          8|        183|   -9.634281356|         22|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|          8|        183|   -9.621407479|         29|00:00:00|+1.00|1.00|0.0000|
#> |-|          2|          2|        183|   -9.586517105|         36|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         167    mean value:    -9.578386969
#> Description of the best-so-far configuration:
#>     .ID.     x1     x2 .PARENT.
#> 167  167 9.4214 2.4315      131
#> 
#> # 2022-08-27 04:35:52 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1     x2
#> 167 9.4214 2.4315
#> 183 9.4256 2.4392
#> # 2022-08-27 04:35:52 UTC: Iteration 6 of 6
#> # experimentsUsedSoFar: 958
#> # remainingBudget: 42
#> # currentBudget: 42
#> # nbConfigurations: 6
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|         10|          6|        183|   -9.540959155|          6|00:00:00|   NA|  NA|    NA|
#> |x|          7|          6|        183|   -9.567018376|         10|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          6|        183|   -9.570779576|         15|00:00:00|+1.00|1.00|0.0000|
#> |x|          8|          6|        183|   -9.559914305|         20|00:00:00|+1.00|1.00|0.0000|
#> |-|          9|          2|        183|   -9.564756307|         24|00:00:00|+1.00|1.00|0.0000|
#> |.|          5|          2|        183|   -9.567761231|         24|00:00:00|+1.00|1.00|0.0000|
#> |.|          6|          2|        183|   -9.590315649|         24|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         167    mean value:    -9.574606952
#> Description of the best-so-far configuration:
#>     .ID.     x1     x2 .PARENT.
#> 167  167 9.4214 2.4315      131
#> 
#> # 2022-08-27 04:35:52 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1     x2
#> 167 9.4214 2.4315
#> 183 9.4256 2.4392
#> # 2022-08-27 04:35:52 UTC: Stopped because there is not enough budget left to race more than the minimum (3)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 7
#> # nbIterations: 7
#> # experimentsUsedSoFar: 982
#> # timeUsed: 0
#> # remainingBudget: 18
#> # currentBudget: 18
#> # number of elites: 2
#> # nbConfigurations: 3
#>        x1     x2 configuration  x_domain         y
#> 1: 9.4214 2.4315           167 <list[2]> -9.574607

# best scoring configuration
instance$result
#>        x1     x2 configuration  x_domain         y
#> 1: 9.4214 2.4315           167 <list[2]> -9.574607

# all evaluations
as.data.table(instance$archive)
#>          x1      x2           y race step instance configuration
#>   1: 4.2972 11.0393 86.26904956    1    1        4             1
#>   2: 4.3731  0.0322 -0.98507004    1    1        4             2
#>   3: 7.4349  9.2148 66.44977493    1    1        4             3
#>   4: 4.1960  3.7837  0.06883147    1    1        4             4
#>   5: 5.5502  5.3881 25.15611962    1    1        4             5
#>  ---                                                            
#> 978: 9.4229  2.4281 -9.52658301    6    1        1           193
#> 979: 9.4151  2.4266 -9.58338971    6    1        7           190
#> 980: 9.4259  2.4241 -9.58276522    6    1        7           191
#> 981: 9.4092  2.4294 -9.58323861    6    1        7           192
#> 982: 9.4229  2.4281 -9.58338883    6    1        7           193
#>                timestamp batch_nr
#>   1: 2022-08-27 04:35:50        1
#>   2: 2022-08-27 04:35:50        1
#>   3: 2022-08-27 04:35:50        1
#>   4: 2022-08-27 04:35:50        1
#>   5: 2022-08-27 04:35:50        1
#>  ---                             
#> 978: 2022-08-27 04:35:52       32
#> 979: 2022-08-27 04:35:52       33
#> 980: 2022-08-27 04:35:52       33
#> 981: 2022-08-27 04:35:52       33
#> 982: 2022-08-27 04:35:52       33