Optimization via Iterated Racing

OptimizerIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi: 10.1016/j.orp.2016.09.002 .

Parameters

instances

list()
A list of instances where the configurations executed on.

targetRunnerParallel

function()
A function that executes the objective function with a specific parameter configuration and instance. A default function is provided, see section "Target Runner and Instances".

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with OptimizerIrace.

Target Runner and Instances

The irace package uses a targetRunner script or R function to evaluate a configuration on a particular instance. Usually it is not necessary to specify a targetRunner function when using OptimizerIrace. A default function is used that forwards several configurations and instances to the user defined objective function. As usually, the user defined function has a xs, xss or xdt parameter depending on the used Objective class. For irace, the function needs an additional instances parameter.

fun = function(xs, instances) {
 # function to evaluate configuration in `xs` on instance `instances`
}

Archive

The Archive holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The optimization result (instance$result) is the best performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("irace")
opt("irace")

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerIrace

Methods

Public methods

OptimizerIrace$new()

OptimizerIrace$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15)
)

codomain = ps(y = p_dbl(tags = "minimize"))

# branin function with noise
# the noise generates different instances of the branin function
# the noise values are passed via the `instances` parameter
fun = function(xdt, instances) {
  a = 1
  b = 5.1 / (4 * (pi^2))
  c = 5 / pi
  r = 6
  s = 10
  t = 1 / (8 * pi)

  data.table(y = (
    a * ((xdt[["x2"]] -
      b * (xdt[["x1"]]^2L) +
      c * xdt[["x1"]] - r)^2) +
      ((s * (1 - t)) * cos(xdt[["x1"]])) +
      unlist(instances)))
}

objective = ObjectiveRFunDt$new(fun = fun, domain = domain, codomain = codomain)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 1000))

# create instances of branin function
instances = rnorm(10, mean = 0, sd = 0.1)

# load optimizer irace and set branin instances
optimizer = opt("irace", instances = instances)

# modifies the instance by reference
optimizer$optimize(instance)
#> # 2021-12-08 04:13:31 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 3
#> # minNbSurvival: 3
#> # nbParameters: 2
#> # seed: 1556805966
#> # confidence level: 0.95
#> # budget: 1000
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2021-12-08 04:13:31 UTC: Iteration 1 of 3
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 1000
#> # currentBudget: 333
#> # nbConfigurations: 55
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          1|         55|          2|   -9.383209699|         55|00:00:00|   NA|  NA|    NA|
#> |x|          2|         55|          2|   -9.382340014|        110|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         55|          2|   -9.372366548|        165|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         55|          2|   -9.354131537|        220|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          1|          2|   -9.364029070|        275|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           2    mean value:    -9.364029070
#> Description of the best-so-far configuration:
#>   .ID.      x1      x2 .PARENT.
#> 2    2 -3.0688 12.5115       NA
#> 
#> # 2021-12-08 04:13:32 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>        x1      x2
#> 2 -3.0688 12.5115
#> # 2021-12-08 04:13:32 UTC: Iteration 2 of 3
#> # experimentsUsedSoFar: 275
#> # remainingBudget: 725
#> # currentBudget: 362
#> # nbConfigurations: 52
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          6|         52|         75|   -9.560211021|         52|00:00:00|   NA|  NA|    NA|
#> |x|          4|         52|         75|   -9.502238582|        103|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|         52|         75|   -9.517645333|        154|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|         52|         75|   -9.520246334|        205|00:00:00|+1.00|1.00|0.0000|
#> |-|          3|          2|         75|   -9.515648918|        256|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           2    mean value:    -9.372586122
#> Description of the best-so-far configuration:
#>   .ID.      x1      x2 .PARENT.
#> 2    2 -3.0688 12.5115       NA
#> 
#> # 2021-12-08 04:13:32 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1      x2
#> 2  -3.0688 12.5115
#> 75 -3.2350 12.4143
#> # 2021-12-08 04:13:33 UTC: Iteration 3 of 3
#> # experimentsUsedSoFar: 531
#> # remainingBudget: 469
#> # currentBudget: 469
#> # nbConfigurations: 60
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          7|         60|        140|   -9.667776920|         60|00:00:00|   NA|  NA|    NA|
#> |x|          4|         60|        140|   -9.575370375|        118|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|         60|        140|   -9.583216485|        176|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         60|        140|   -9.571401599|        234|00:00:00|+1.00|1.00|0.0000|
#> |-|          2|          3|        140|   -9.570122810|        293|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          75    mean value:    -9.533376256
#> Description of the best-so-far configuration:
#>    .ID.     x1      x2 .PARENT.
#> 75   75 -3.235 12.4143        2
#> 
#> # 2021-12-08 04:13:33 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>          x1      x2
#> 75  -3.2350 12.4143
#> 2   -3.0688 12.5115
#> 140 -3.1846 12.4202
#> # 2021-12-08 04:13:33 UTC: Iteration 4 of 4
#> # experimentsUsedSoFar: 824
#> # remainingBudget: 176
#> # currentBudget: 176
#> # nbConfigurations: 21
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          8|         21|        181|   -9.691425999|         21|00:00:00|   NA|  NA|    NA|
#> |x|          2|         21|        181|   -9.632336269|         39|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|         21|        181|   -9.623940044|         57|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|         21|        181|   -9.611701510|         76|00:00:00|+1.00|1.00|0.0000|
#> |-|          3|          4|        181|   -9.598200373|         94|00:00:00|+1.00|1.00|0.0000|
#> |!|          4|          4|        181|   -9.580367430|         95|00:00:00|+1.00|1.00|0.0000|
#> |!|          5|          4|        181|   -9.582514285|         97|00:00:00|+1.00|1.00|0.0000|
#> |-|          7|          1|        181|   -9.594201975|         98|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         181    mean value:    -9.594201975
#> Description of the best-so-far configuration:
#>     .ID.      x1      x2 .PARENT.
#> 181  181 -3.1582 12.3473      140
#> 
#> # 2021-12-08 04:13:33 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>          x1      x2
#> 181 -3.1582 12.3473
#> # 2021-12-08 04:13:33 UTC: Iteration 5 of 5
#> # experimentsUsedSoFar: 922
#> # remainingBudget: 78
#> # currentBudget: 78
#> # nbConfigurations: 8
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          9|          8|        188|   -9.511941391|          8|00:00:00|   NA|  NA|    NA|
#> |x|          7|          8|        188|   -9.594068846|         15|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|          8|        188|   -9.587188242|         22|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|          8|        188|   -9.584182782|         29|00:00:00|+1.00|1.00|0.0000|
#> |-|          4|          2|        188|   -9.565622868|         36|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         181    mean value:    -9.585041855
#> Description of the best-so-far configuration:
#>     .ID.      x1      x2 .PARENT.
#> 181  181 -3.1582 12.3473      140
#> 
#> # 2021-12-08 04:13:34 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>          x1      x2
#> 181 -3.1582 12.3473
#> 188 -3.1532 12.3422
#> # 2021-12-08 04:13:34 UTC: Iteration 6 of 6
#> # experimentsUsedSoFar: 958
#> # remainingBudget: 42
#> # currentBudget: 42
#> # nbConfigurations: 6
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|         10|          6|        190|   -9.370847668|          6|00:00:00|   NA|  NA|    NA|
#> |x|          3|          6|        190|   -9.457841061|         11|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|          6|        190|   -9.497102219|         15|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|          6|        190|   -9.516297956|         19|00:00:00|+1.00|1.00|0.0000|
#> |-|          7|          3|        190|   -9.548369252|         23|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         181    mean value:    -9.563558574
#> Description of the best-so-far configuration:
#>     .ID.      x1      x2 .PARENT.
#> 181  181 -3.1582 12.3473      140
#> 
#> # 2021-12-08 04:13:34 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>          x1      x2
#> 181 -3.1582 12.3473
#> 188 -3.1532 12.3422
#> 190 -3.1570 12.3364
#> # 2021-12-08 04:13:34 UTC: Iteration 7 of 7
#> # experimentsUsedSoFar: 981
#> # remainingBudget: 19
#> # currentBudget: 19
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|         11|          4|        194|   -9.676876092|          4|00:00:00|   NA|  NA|    NA|
#> |x|          2|          4|        194|   -9.625491458|          5|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|          4|        194|   -9.619663599|          8|00:00:00|+1.00|1.00|0.0000|
#> |x|          8|          4|        194|   -9.637819270|         11|00:00:00|+1.00|1.00|0.0000|
#> |!|         10|          4|        194|   -9.584469281|         12|00:00:00|+1.00|1.00|0.0000|
#> |!|          1|          4|        194|   -9.583032100|         13|00:00:00|+1.00|1.00|0.0000|
#> |!|          3|          4|        194|   -9.577606959|         14|00:00:00|+1.00|1.00|0.0000|
#> |!|          5|          4|        194|   -9.579938051|         17|00:00:00|+1.00|1.00|0.0000|
#> |!|          4|          4|        194|   -9.570174157|         19|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         188    mean value:    -9.573962453
#> Description of the best-so-far configuration:
#>     .ID.      x1      x2 .PARENT.
#> 188  188 -3.1532 12.3422      181
#> 
#> # 2021-12-08 04:13:34 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>          x1      x2
#> 188 -3.1532 12.3422
#> 181 -3.1582 12.3473
#> 190 -3.1570 12.3364
#> # 2021-12-08 04:13:34 UTC: Stopped because budget is exhausted
#> # Iteration: 8
#> # nbIterations: 7
#> # experimentsUsedSoFar: 1000
#> # timeUsed: 0
#> # remainingBudget: 0
#> # currentBudget: 19
#> # number of elites: 3
#> # nbConfigurations: 4
#>         x1      x2 configuration  x_domain         y
#> 1: -3.1532 12.3422           188 <list[2]> -9.573962

# best scoring configuration
instance$result
#>         x1      x2 configuration  x_domain         y
#> 1: -3.1532 12.3422           188 <list[2]> -9.573962

# all evaluations
as.data.table(instance$archive)
#>            x1      x2         y race step instance configuration
#>    1: -0.8277 10.5450 16.375669    1    1        3             1
#>    2: -3.0688 12.5115 -9.383210    1    1        3             2
#>    3:  6.1177  1.6092  9.756739    1    1        3             3
#>    4:  2.8288  9.9990 46.651015    1    1        3             4
#>    5:  7.2457  6.0782 28.820399    1    1        3             5
#>   ---                                                           
#>  996: -3.1532 12.3422 -9.595576    7    1        7           188
#>  997: -3.1570 12.3364 -9.596034    7    1        7           190
#>  998: -3.1545 12.3327 -9.596256    7    1        7           194
#>  999: -3.1570 12.3364 -9.491841    7    1        4           190
#> 1000: -3.1545 12.3327 -9.492063    7    1        4           194
#>                 timestamp batch_nr
#>    1: 2021-12-08 04:13:32        1
#>    2: 2021-12-08 04:13:32        1
#>    3: 2021-12-08 04:13:32        1
#>    4: 2021-12-08 04:13:32        1
#>    5: 2021-12-08 04:13:32        1
#>   ---                             
#>  996: 2021-12-08 04:13:34       41
#>  997: 2021-12-08 04:13:34       41
#>  998: 2021-12-08 04:13:34       41
#>  999: 2021-12-08 04:13:34       42
#> 1000: 2021-12-08 04:13:34       42