Optimization via Iterated Racing

OptimizerIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi: 10.1016/j.orp.2016.09.002 .

Parameters

instances

list()
A list of instances where the configurations executed on.

targetRunnerParallel

function()
A function that executes the objective function with a specific parameter configuration and instance. A default function is provided, see section "Target Runner and Instances".

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with OptimizerIrace.

Target Runner and Instances

The irace package uses a targetRunner script or R function to evaluate a configuration on a particular instance. Usually it is not necessary to specify a targetRunner function when using OptimizerIrace. A default function is used that forwards several configurations and instances to the user defined objective function. As usually, the user defined function has a xs, xss or xdt parameter depending on the used Objective class. For irace, the function needs an additional instances parameter.

fun = function(xs, instances) {
 # function to evaluate configuration in `xs` on instance `instances`
}

Archive

The Archive holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The optimization result (instance$result) is the best performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("irace")
opt("irace")

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerIrace

Methods

Public methods

OptimizerIrace$new()

OptimizerIrace$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15)
)

codomain = ps(y = p_dbl(tags = "minimize"))

# branin function with noise
# the noise generates different instances of the branin function
# the noise values are passed via the `instances` parameter
fun = function(xdt, instances) {
  a = 1
  b = 5.1 / (4 * (pi^2))
  c = 5 / pi
  r = 6
  s = 10
  t = 1 / (8 * pi)

  data.table(y = (
    a * ((xdt[["x2"]] -
      b * (xdt[["x1"]]^2L) +
      c * xdt[["x1"]] - r)^2) +
      ((s * (1 - t)) * cos(xdt[["x1"]])) +
      unlist(instances)))
}

objective = ObjectiveRFunDt$new(fun = fun, domain = domain, codomain = codomain)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 1000))

# create instances of branin function
instances = rnorm(10, mean = 0, sd = 0.1)

# load optimizer irace and set branin instances
optimizer = opt("irace", instances = instances)

# modifies the instance by reference
optimizer$optimize(instance)
#> # 2022-02-25 10:16:37 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 3
#> # minNbSurvival: 3
#> # nbParameters: 2
#> # seed: 1030232769
#> # confidence level: 0.95
#> # budget: 1000
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2022-02-25 10:16:37 UTC: Iteration 1 of 3
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 1000
#> # currentBudget: 333
#> # nbConfigurations: 55
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          1|         55|         40|   -9.524350126|         55|00:00:00|   NA|  NA|    NA|
#> |x|          2|         55|         40|   -9.520351165|        110|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         55|         40|   -9.514304162|        165|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         55|         40|   -9.509133226|        220|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          1|         40|   -9.538471350|        275|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          40    mean value:    -9.538471350
#> Description of the best-so-far configuration:
#>    .ID.      x1      x2 .PARENT.
#> 40   40 -3.0405 11.8327       NA
#> 
#> # 2022-02-25 10:16:38 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1      x2
#> 40 -3.0405 11.8327
#> # 2022-02-25 10:16:38 UTC: Iteration 2 of 3
#> # experimentsUsedSoFar: 275
#> # remainingBudget: 725
#> # currentBudget: 362
#> # nbConfigurations: 52
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          6|         52|        105|   -9.474625950|         52|00:00:00|   NA|  NA|    NA|
#> |x|          2|         52|        105|   -9.501402176|        103|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         52|        105|   -9.505613569|        154|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         52|        105|   -9.505571831|        205|00:00:00|+1.00|1.00|0.0000|
#> |-|          1|          2|        105|   -9.511692729|        256|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          40    mean value:    -9.525859417
#> Description of the best-so-far configuration:
#>    .ID.      x1      x2 .PARENT.
#> 40   40 -3.0405 11.8327       NA
#> 
#> # 2022-02-25 10:16:38 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>          x1      x2
#> 40  -3.0405 11.8327
#> 105 -3.1680 12.6108
#> # 2022-02-25 10:16:38 UTC: Iteration 3 of 3
#> # experimentsUsedSoFar: 531
#> # remainingBudget: 469
#> # currentBudget: 469
#> # nbConfigurations: 60
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          7|         60|        110|   -9.499738367|         60|00:00:00|   NA|  NA|    NA|
#> |x|          5|         60|        110|   -9.617667470|        119|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|         60|        110|   -9.592635806|        177|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|         60|        110|   -9.595507568|        235|00:00:00|+1.00|1.00|0.0000|
#> |-|          2|          3|        110|   -9.595631040|        293|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         105    mean value:    -9.522557933
#> Description of the best-so-far configuration:
#>     .ID.     x1      x2 .PARENT.
#> 105  105 -3.168 12.6108       40
#> 
#> # 2022-02-25 10:16:39 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>          x1      x2
#> 105 -3.1680 12.6108
#> 40  -3.0405 11.8327
#> 110 -3.1819 12.3306
#> # 2022-02-25 10:16:39 UTC: Iteration 4 of 4
#> # experimentsUsedSoFar: 824
#> # remainingBudget: 176
#> # currentBudget: 176
#> # nbConfigurations: 21
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          8|         21|        171|   -9.465561512|         21|00:00:00|   NA|  NA|    NA|
#> |x|          2|         21|        171|   -9.534725532|         39|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|         21|        171|   -9.604270753|         57|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         21|        171|   -9.600639941|         76|00:00:00|+1.00|1.00|0.0000|
#> |-|          6|          4|        171|   -9.590579373|         94|00:00:00|+1.00|1.00|0.0000|
#> |!|          7|          4|        171|   -9.576733309|         95|00:00:00|+1.00|1.00|0.0000|
#> |-|          1|          3|        171|   -9.581755333|         96|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         105    mean value:    -9.505969487
#> Description of the best-so-far configuration:
#>     .ID.     x1      x2 .PARENT.
#> 105  105 -3.168 12.6108       40
#> 
#> # 2022-02-25 10:16:39 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>          x1      x2
#> 105 -3.1680 12.6108
#> 40  -3.0405 11.8327
#> 171 -3.1595 12.3034
#> # 2022-02-25 10:16:39 UTC: Iteration 5 of 5
#> # experimentsUsedSoFar: 920
#> # remainingBudget: 80
#> # currentBudget: 80
#> # nbConfigurations: 10
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          9|         10|        189|   -9.709440773|         10|00:00:00|   NA|  NA|    NA|
#> |x|          2|         10|        189|   -9.657047659|         17|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|         10|        189|   -9.686073835|         24|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|         10|        189|   -9.667718493|         31|00:00:00|+1.00|1.00|0.0000|
#> |-|          7|          4|        189|   -9.635828390|         38|00:00:00|+1.00|1.00|0.0000|
#> |!|          8|          4|        189|   -9.607578076|         39|00:00:00|+1.00|1.00|0.0000|
#> |!|          3|          4|        189|   -9.605140136|         40|00:00:00|+1.00|1.00|0.0000|
#> |!|          4|          4|        189|   -9.602237964|         42|00:00:00|+1.00|1.00|0.0000|
#> |-|          6|          1|        189|   -9.596556200|         43|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         189    mean value:    -9.596556200
#> Description of the best-so-far configuration:
#>     .ID.      x1      x2 .PARENT.
#> 189  189 -3.1558 12.3046      171
#> 
#> # 2022-02-25 10:16:39 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>          x1      x2
#> 189 -3.1558 12.3046
#> # 2022-02-25 10:16:39 UTC: Iteration 6 of 6
#> # experimentsUsedSoFar: 963
#> # remainingBudget: 37
#> # currentBudget: 37
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|         10|          4|        189|   -9.730094846|          4|00:00:00|   NA|  NA|    NA|
#> |x|          4|          4|        189|   -9.656008802|          7|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|          4|        189|   -9.685381264|         10|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|          4|        189|   -9.665199584|         13|00:00:00|+1.00|1.00|0.0000|
#> |-|          7|          1|        189|   -9.633813263|         16|00:00:00|   NA|  NA|    NA|
#> |.|          1|          1|        189|   -9.630286464|         16|00:00:00|   NA|  NA|    NA|
#> |.|          9|          1|        189|   -9.641594222|         16|00:00:00|   NA|  NA|    NA|
#> |.|          6|          1|        189|   -9.630282706|         16|00:00:00|   NA|  NA|    NA|
#> |.|          3|          1|        189|   -9.625863794|         16|00:00:00|   NA|  NA|    NA|
#> |.|          8|          1|        189|   -9.609910065|         16|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         189    mean value:    -9.609910065
#> Description of the best-so-far configuration:
#>     .ID.      x1      x2 .PARENT.
#> 189  189 -3.1558 12.3046      171
#> 
#> # 2022-02-25 10:16:40 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>          x1      x2
#> 189 -3.1558 12.3046
#> # 2022-02-25 10:16:40 UTC: Stopped because there is not enough budget left to race more than the minimum (3)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 7
#> # nbIterations: 7
#> # experimentsUsedSoFar: 979
#> # timeUsed: 0
#> # remainingBudget: 21
#> # currentBudget: 21
#> # number of elites: 1
#> # nbConfigurations: 2
#>         x1      x2 configuration  x_domain        y
#> 1: -3.1558 12.3046           189 <list[2]> -9.60991

# best scoring configuration
instance$result
#>         x1      x2 configuration  x_domain        y
#> 1: -3.1558 12.3046           189 <list[2]> -9.60991

# all evaluations
as.data.table(instance$archive)
#>           x1      x2          y race step instance configuration
#>   1:  7.8578 13.2478 138.658413    1    1        4             1
#>   2:  3.0348 13.1495 106.859624    1    1        4             2
#>   3: -3.2536  1.9039 103.696588    1    1        4             3
#>   4: -4.6299  1.8094 204.503513    1    1        4             4
#>   5:  5.8096  4.8141  22.225249    1    1        4             5
#>  ---                                                            
#> 975: -3.1640 12.2897  -9.601696    6    1        7           191
#> 976: -3.1624 12.2947  -9.602644    6    1        7           192
#> 977: -3.1598 12.3056  -9.507492    6    1        2           190
#> 978: -3.1640 12.2897  -9.505310    6    1        2           191
#> 979: -3.1624 12.2947  -9.506258    6    1        2           192
#>                timestamp batch_nr
#>   1: 2022-02-25 10:16:37        1
#>   2: 2022-02-25 10:16:37        1
#>   3: 2022-02-25 10:16:37        1
#>   4: 2022-02-25 10:16:37        1
#>   5: 2022-02-25 10:16:37        1
#>  ---                             
#> 975: 2022-02-25 10:16:40       35
#> 976: 2022-02-25 10:16:40       35
#> 977: 2022-02-25 10:16:40       36
#> 978: 2022-02-25 10:16:40       36
#> 979: 2022-02-25 10:16:40       36