Optimization via Iterated Racing

OptimizerIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi: 10.1016/j.orp.2016.09.002 .

Parameters

instances

list()
A list of instances where the target algorithm is executed on.

targetRunner

function()
A function that executes the target algorithm with a specific parameter configuration and instance.

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with OptimizerIrace.

Target Runner and Instances

The irace package uses a targetRunner script or R function to evaluate a configuration on a particular instance. Usually it is not necessary to specify a targetRunner function when using OptimizerIrace. A default function is used that forwards several configurations and an instance to the user defined objective function. As usual, the configurations are passed in xss to the objective. The instance passed via the $irace_instance public field.

ObjectiveIrace = R6Class("ObjectiveIrace", inherit = Objective,
  public = list(
    irace_instance = NULL
  ),

  private = list(
    .eval_many = function(xss) {
      # objective function to evaluate configurations in xss
      # access irace instance via self$irace_instance
    }
  )
)

Alternatively, $.eval(xs) can be implemented which only passes one configuration at once. The use of $.eval_many(xss) allows us to implement a parallelization e.g. with the future package.

Archive

The Archive holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The optimization result (instance$result) is the best performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("irace")
opt("irace")

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerIrace

Methods

Public methods

OptimizerIrace$new()

OptimizerIrace$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(data.table)
library(R6)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15)
)

ObjectiveIrace = R6Class("ObjectiveIrace", inherit = Objective,
  public = list(
    irace_instance = NULL
  ),

  private = list(
    .eval_many = function(xss) {
      # branin function with `tau` noise 
      a = 1
      b = 5.1 / (4 * (pi ^ 2))
      c = 5 / pi
      r = 6
      s = 10
      t = 1 / (8 * pi)
 
      data.table(y = mapply(function(xs, tau) {
       a * ((xs$x2 -
       b * (xs$x1 ^ 2) +
       c * xs$x1 - r) ^ 2) +
       ((s * (1 - t)) * cos(xs$x1)) +
       tau
      }, xss, self$irace_instance))
    }
  )
)

codomain = ps(y = p_dbl(tags = "minimize"))
objective = ObjectiveIrace$new(domain = domain,  codomain = codomain)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 1000))

# create instances of branin function
irace_instance = rnorm(10, mean = 0, sd = 0.1)
optimizer = opt("irace", instances = irace_instance)

# modifies the instance by reference
optimizer$optimize(instance)

#> # 2021-07-02 21:18:33 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 3
#> # minNbSurvival: 3
#> # nbParameters: 2
#> # seed: 1133693201
#> # confidence level: 0.95
#> # budget: 1000
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2021-07-02 21:18:33 UTC: Iteration 1 of 3
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 1000
#> # currentBudget: 333
#> # nbConfigurations: 55
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          1|         55|         40|   -9.115799370|         55|00:00:00|   NA|  NA|    NA|
#> |x|          2|         55|         40|   -9.046761677|        110|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         55|         40|   -9.051203193|        165|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         55|         40|   -9.066497019|        220|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          1|         40|   -9.051381306|        275|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          40    mean value:    -9.051381306
#> Description of the best-so-far configuration:
#>    .ID.     x1     x2 .PARENT.
#> 40   40 9.5078 1.8498       NA
#> 
#> # 2021-07-02 21:18:33 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>        x1     x2
#> 40 9.5078 1.8498
#> # 2021-07-02 21:18:33 UTC: Iteration 2 of 3
#> # experimentsUsedSoFar: 275
#> # remainingBudget: 725
#> # currentBudget: 362
#> # nbConfigurations: 52
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          6|         52|         67|   -9.564838955|         52|00:00:00|   NA|  NA|    NA|
#> |x|          1|         52|         67|   -9.586113158|        103|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|         52|         67|   -9.547179430|        154|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         52|         67|   -9.548303127|        205|00:00:00|+1.00|1.00|0.0000|
#> |-|          4|          2|         67|   -9.559435799|        256|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          40    mean value:    -9.055026249
#> Description of the best-so-far configuration:
#>    .ID.     x1     x2 .PARENT.
#> 40   40 9.5078 1.8498       NA
#> 
#> # 2021-07-02 21:18:34 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>        x1     x2
#> 40 9.5078 1.8498
#> 67 9.4981 2.5211
#> # 2021-07-02 21:18:34 UTC: Iteration 3 of 3
#> # experimentsUsedSoFar: 531
#> # remainingBudget: 469
#> # currentBudget: 469
#> # nbConfigurations: 60
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          7|         60|        157|   -9.579757933|         60|00:00:00|   NA|  NA|    NA|
#> |x|          1|         60|        157|   -9.604254662|        118|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|         60|        157|   -9.566395110|        176|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|         60|        157|   -9.571347079|        234|00:00:00|+1.00|1.00|0.0000|
#> |-|          3|          3|        157|   -9.571685313|        292|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          40    mean value:    -9.056709058
#> Description of the best-so-far configuration:
#>    .ID.     x1     x2 .PARENT.
#> 40   40 9.5078 1.8498       NA
#> 
#> # 2021-07-02 21:18:34 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1     x2
#> 40  9.5078 1.8498
#> 67  9.4981 2.5211
#> 157 9.4500 2.5370
#> # 2021-07-02 21:18:35 UTC: Iteration 4 of 4
#> # experimentsUsedSoFar: 823
#> # remainingBudget: 177
#> # currentBudget: 177
#> # nbConfigurations: 22
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          8|         22|        183|   -9.439273024|         22|00:00:00|   NA|  NA|    NA|
#> |x|          7|         22|        183|   -9.511626037|         41|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|         22|        183|   -9.510447889|         62|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|         22|        183|   -9.506560198|         81|00:00:00|+1.00|1.00|0.0000|
#> |-|          3|          4|        183|   -9.520700031|        100|00:00:00|+1.00|1.00|0.0000|
#> |!|          1|          4|        183|   -9.539412111|        101|00:00:00|+1.00|1.00|0.0000|
#> |-|          6|          3|        183|   -9.546699539|        102|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          67    mean value:    -9.531470902
#> Description of the best-so-far configuration:
#>    .ID.     x1     x2 .PARENT.
#> 67   67 9.4981 2.5211       40
#> 
#> # 2021-07-02 21:18:35 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1     x2
#> 67  9.4981 2.5211
#> 40  9.5078 1.8498
#> 183 9.4182 2.4861
#> # 2021-07-02 21:18:35 UTC: Iteration 5 of 5
#> # experimentsUsedSoFar: 925
#> # remainingBudget: 75
#> # currentBudget: 75
#> # nbConfigurations: 9
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          9|          9|        183|   -9.615027613|          9|00:00:00|   NA|  NA|    NA|
#> |x|          3|          9|        183|   -9.596143489|         15|00:00:00|+1.00|1.00|0.0000|
#> |x|          8|          9|        183|   -9.543853334|         21|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|          9|        183|   -9.531614282|         27|00:00:00|+1.00|1.00|0.0000|
#> |-|          7|          3|        183|   -9.542087235|         33|00:00:00|+1.00|1.00|0.0000|
#> |.|          1|          3|        183|   -9.557234781|         33|00:00:00|+1.00|1.00|0.0000|
#> |.|          5|          3|        183|   -9.550214326|         33|00:00:00|+1.00|1.00|0.0000|
#> |.|          6|          3|        183|   -9.555240548|         33|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          67    mean value:    -9.537912187
#> Description of the best-so-far configuration:
#>    .ID.     x1     x2 .PARENT.
#> 67   67 9.4981 2.5211       40
#> 
#> # 2021-07-02 21:18:35 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1     x2
#> 67  9.4981 2.5211
#> 40  9.5078 1.8498
#> 183 9.4182 2.4861
#> # 2021-07-02 21:18:35 UTC: Iteration 6 of 6
#> # experimentsUsedSoFar: 958
#> # remainingBudget: 42
#> # currentBudget: 42
#> # nbConfigurations: 6
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|         10|          6|        183|   -9.792636659|          6|00:00:00|   NA|  NA|    NA|
#> |x|          2|          6|        183|   -9.643766891|          9|00:00:00|+1.00|1.00|0.0000|
#> |x|          8|          6|        183|   -9.575602269|         12|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          6|        183|   -9.589089611|         16|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          3|        183|   -9.572890007|         19|00:00:00|+1.00|1.00|0.0000|
#> |.|          1|          3|        183|   -9.582903758|         19|00:00:00|+1.00|1.00|0.0000|
#> |.|          9|          3|        183|   -9.587492880|         19|00:00:00|+1.00|1.00|0.0000|
#> |.|          3|          3|        183|   -9.586213691|         19|00:00:00|+1.00|1.00|0.0000|
#> |.|          6|          3|        183|   -9.586681514|         19|00:00:00|+1.00|1.00|0.0000|
#> |.|          7|          3|        183|   -9.586411268|         19|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         183    mean value:    -9.586411268
#> Description of the best-so-far configuration:
#>     .ID.     x1     x2 .PARENT.
#> 183  183 9.4182 2.4861      157
#> 
#> # 2021-07-02 21:18:35 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1     x2
#> 183 9.4182 2.4861
#> 67  9.4981 2.5211
#> 40  9.5078 1.8498
#> # 2021-07-02 21:18:35 UTC: Iteration 7 of 7
#> # experimentsUsedSoFar: 977
#> # remainingBudget: 23
#> # currentBudget: 23
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|         11|          4|        183|   -9.577259365|          4|00:00:00|   NA|  NA|    NA|
#> |x|          5|          4|        183|   -9.542675479|          5|00:00:00|+1.00|1.00|0.0000|
#> |x|          8|          4|        183|   -9.508207994|          6|00:00:00|+1.00|1.00|0.0000|
#> |x|          7|          4|        183|   -9.527150758|          7|00:00:00|+1.00|1.00|0.0000|
#> |-|          9|          3|        183|   -9.544726129|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          2|          3|        183|   -9.536421295|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          1|          3|        183|   -9.550214326|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          6|          3|        183|   -9.555240548|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          4|          3|        183|   -9.563497336|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|         10|          3|        183|   -9.586411268|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          3|          3|        183|   -9.585579277|          8|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         183    mean value:    -9.585579277
#> Description of the best-so-far configuration:
#>     .ID.     x1     x2 .PARENT.
#> 183  183 9.4182 2.4861      157
#> 
#> # 2021-07-02 21:18:35 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1     x2
#> 183 9.4182 2.4861
#> 67  9.4981 2.5211
#> 40  9.5078 1.8498
#> # 2021-07-02 21:18:35 UTC: Iteration 8 of 8
#> # experimentsUsedSoFar: 985
#> # remainingBudget: 15
#> # currentBudget: 15
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|         12|          4|        183|   -9.632972510|          4|00:00:00|   NA|  NA|    NA|
#> |x|         10|          4|        183|   -9.712804584|          5|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|          4|        183|   -9.672011091|          6|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|          4|        183|   -9.631031216|          7|00:00:00|+1.00|1.00|0.0000|
#> |-|         11|          3|        183|   -9.620276846|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          1|          3|        183|   -9.622392790|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          2|          3|        183|   -9.604179123|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          4|          3|        183|   -9.607350688|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          3|          3|        183|   -9.604007207|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          9|          3|        183|   -9.605109248|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          7|          3|        183|   -9.603188321|          8|00:00:00|+1.00|1.00|0.0000|
#> |.|          8|          3|        183|   -9.589528713|          8|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         183    mean value:    -9.589528713
#> Description of the best-so-far configuration:
#>     .ID.     x1     x2 .PARENT.
#> 183  183 9.4182 2.4861      157
#> 
#> # 2021-07-02 21:18:35 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>         x1     x2
#> 183 9.4182 2.4861
#> 67  9.4981 2.5211
#> 40  9.5078 1.8498
#> # 2021-07-02 21:18:35 UTC: Stopped because there is not enough budget left to race more than the minimum (3)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 9
#> # nbIterations: 9
#> # experimentsUsedSoFar: 993
#> # timeUsed: 0
#> # remainingBudget: 7
#> # currentBudget: 7
#> # number of elites: 3
#> # nbConfigurations: 3

#>        x1     x2 configuration  x_domain         y
#> 1: 9.4182 2.4861           183 <list[2]> -9.589529

# best scoring evaluation
instance$result

#>        x1     x2 configuration  x_domain         y
#> 1: 9.4182 2.4861           183 <list[2]> -9.589529

# all evaluations
as.data.table(instance$archive)

#>           x1      x2          y race step instance configuration
#>   1: -4.8222 11.5573  27.250474    1    1        9             1
#>   2: -1.5696  4.8659  15.586282    1    1        9             2
#>   3:  5.6902 12.2173 130.936192    1    1        9             3
#>   4:  8.2027  2.3163  -2.850992    1    1        9             4
#>   5:  8.4679  7.7880  30.459487    1    1        9             5
#>  ---                                                            
#> 989:  9.5007  1.8534  -9.134670    8    1        9           194
#> 990:  9.5007  1.8534  -9.294334    8    1        7           194
#> 991:  9.5007  1.8534  -9.092122    8    1        5           194
#> 992:  9.5007  1.8534  -9.009789    8    1        1           194
#> 993:  9.5007  1.8534  -9.078957    8    1        3           194
#>                timestamp batch_nr x_domain_x1 x_domain_x2
#>   1: 2021-07-02 21:18:33        1     -4.8222     11.5573
#>   2: 2021-07-02 21:18:33        1     -1.5696      4.8659
#>   3: 2021-07-02 21:18:33        1      5.6902     12.2173
#>   4: 2021-07-02 21:18:33        1      8.2027      2.3163
#>   5: 2021-07-02 21:18:33        1      8.4679      7.7880
#>  ---                                                     
#> 989: 2021-07-02 21:18:35       38      9.5007      1.8534
#> 990: 2021-07-02 21:18:35       39      9.5007      1.8534
#> 991: 2021-07-02 21:18:35       40      9.5007      1.8534
#> 992: 2021-07-02 21:18:35       41      9.5007      1.8534
#> 993: 2021-07-02 21:18:35       42      9.5007      1.8534