Optimization via Iterated Racing

OptimizerBatchIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43–58. doi:10.1016/j.orp.2016.09.002 .

Parameters

instances

list()
A list of instances where the configurations executed on.

targetRunnerParallel

function()
A function that executes the objective function with a specific parameter configuration and instance. A default function is provided, see section "Target Runner and Instances".

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with OptimizerBatchIrace.

In contrast to irace::defaultScenario(), we set digits = 15. This represents double parameters with a higher precision and avoids rounding errors.

Target Runner and Instances

The irace package uses a targetRunner script or R function to evaluate a configuration on a particular instance. Usually it is not necessary to specify a targetRunner function when using OptimizerBatchIrace. A default function is used that forwards several configurations and instances to the user defined objective function. As usually, the user defined function has a xs, xss or xdt parameter depending on the used Objective class. For irace, the function needs an additional instances parameter.

fun = function(xs, instances) {
 # function to evaluate configuration in `xs` on instance `instances`
}

Archive

The Archive holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The optimization result (instance$result) is the best performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("irace")
opt("irace")

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super classes

bbotk::Optimizer -> bbotk::OptimizerBatch -> OptimizerBatchIrace

Methods

Public methods

OptimizerBatchIrace$new()

OptimizerBatchIrace$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$print()

bbotk::OptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerBatchIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerBatchIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# runtime of the example is too long
# \donttest{

library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15)
)

codomain = ps(y = p_dbl(tags = "minimize"))

# branin function with noise
# the noise generates different instances of the branin function
# the noise values are passed via the `instances` parameter
fun = function(xdt, instances) {
  ys = branin(xdt[["x1"]], xdt[["x2"]], noise = as.numeric(instances))
  data.table(y = ys)
}

# define objective with instances as a constant
objective = ObjectiveRFunDt$new(
 fun = fun,
 domain = domain,
 codomain = codomain,
 constants = ps(instances = p_uty()))

instance = OptimInstanceBatchSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 96))

# create instances of branin function
instances = rnorm(10, mean = 0, sd = 0.1)

# load optimizer irace and set branin instances
optimizer = opt("irace", instances = instances)

# modifies the instance by reference
optimizer$optimize(instance)
#> # 2025-09-20 18:39:09 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 3
#> # minNbSurvival: 3
#> # nbParameters: 2
#> # seed: 438793322
#> # confidence level: 0.95
#> # budget: 96
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2025-09-20 18:39:09 UTC: Iteration 1 of 3
#> # experimentsUsed: 0
#> # remainingBudget: 96
#> # currentBudget: 32
#> # nbConfigurations: 5
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded.
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          1|          5|          5|     5.696544597|          5|00:00:00|   NA|  NA|    NA|
#> |x|          2|          5|          5|     5.696544597|         10|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|          5|          5|     5.696544597|         15|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          5|          5|     5.696544597|         20|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          1|          5|     5.696544597|         25|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           5    mean value:      5.696544597
#> Description of the best-so-far configuration:
#>   .ID.               x1               x2 .PARENT.
#> 5    5 2.15160071849823 4.15610745549202       NA
#> 
#> # 2025-09-20 18:39:10 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                 x1               x2
#> 5 2.15160071849823 4.15610745549202
#> # 2025-09-20 18:39:10 UTC: Iteration 2 of 3
#> # experimentsUsed: 25
#> # remainingBudget: 71
#> # currentBudget: 35
#> # nbConfigurations: 5
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded.
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          6|          5|          6|     2.058597335|          5|00:00:00|   NA|  NA|    NA|
#> |x|          1|          5|          6|     2.058597335|          9|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          5|          6|     2.058597335|         13|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|          5|          6|     2.058597335|         17|00:00:00|+1.00|1.00|0.0000|
#> |-|          2|          2|          6|     2.058597335|         21|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best configuration for the instances in this race: 6
#> Best-so-far configuration:           5    mean value:      5.696544597
#> Description of the best-so-far configuration:
#>   .ID.               x1               x2 .PARENT.
#> 5    5 2.15160071849823 4.15610745549202       NA
#> 
#> # 2025-09-20 18:39:10 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                 x1               x2
#> 5 2.15160071849823 4.15610745549202
#> 6 9.43809640151962 1.19790239189972
#> # 2025-09-20 18:39:10 UTC: Iteration 3 of 3
#> # experimentsUsed: 46
#> # remainingBudget: 50
#> # currentBudget: 50
#> # nbConfigurations: 7
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded.
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          7|          7|         12|    0.7123056014|          7|00:00:00|   NA|  NA|    NA|
#> |x|          3|          7|         12|    0.7123056014|         13|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|          7|         12|    0.7123056014|         18|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          7|         12|    0.7123056014|         23|00:00:00|+1.00|1.00|0.0000|
#> |-|          1|          3|         12|    0.7123056014|         28|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best configuration for the instances in this race: 12
#> Best-so-far configuration:           6    mean value:      2.058597335
#> Description of the best-so-far configuration:
#>   .ID.               x1               x2 .PARENT.
#> 6    6 9.43809640151962 1.19790239189972        5
#> 
#> # 2025-09-20 18:39:10 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  x1               x2
#> 6  9.43809640151962 1.19790239189972
#> 5  2.15160071849823 4.15610745549202
#> 12 3.00459676598218 1.91049682606301
#> # 2025-09-20 18:39:10 UTC: Iteration 4 of 4
#> # experimentsUsed: 74
#> # remainingBudget: 22
#> # currentBudget: 22
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded.
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          8|          4|         12|    0.7123056014|          4|00:00:00|   NA|  NA|    NA|
#> |x|          7|          4|         12|    0.7123056014|          5|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|          4|         12|    0.7123056014|          7|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|          4|         12|    0.7123056014|          9|00:00:00|+1.00|1.00|0.0000|
#> |-|          2|          3|         12|    0.7123056014|         10|00:00:00|+1.00|1.00|0.0000|
#> |.|          3|          3|         12|    0.7123056014|         10|00:00:00|+1.00|1.00|0.0000|
#> |.|          4|          3|         12|    0.7123056014|         10|00:00:00|+1.00|1.00|0.0000|
#> |.|          1|          3|         12|    0.7123056014|         10|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          12    mean value:     0.7123056014
#> Description of the best-so-far configuration:
#>    .ID.               x1               x2 .PARENT.
#> 12   12 3.00459676598218 1.91049682606301        5
#> 
#> # 2025-09-20 18:39:10 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  x1               x2
#> 12 3.00459676598218 1.91049682606301
#> 6  9.43809640151962 1.19790239189972
#> 5  2.15160071849823 4.15610745549202
#> # 2025-09-20 18:39:10 UTC: Iteration 5 of 5
#> # experimentsUsed: 84
#> # remainingBudget: 12
#> # currentBudget: 12
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded.
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          8|          4|         12|    0.7123056014|          1|00:00:00|   NA|  NA|    NA|
#> |x|          3|          4|         12|    0.7123056014|          2|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|          4|         12|    0.7123056014|          3|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|          4|         12|    0.7123056014|          4|00:00:00|+1.00|1.00|0.0000|
#> |-|          7|          3|         12|    0.7123056014|          5|00:00:00|+1.00|1.00|0.0000|
#> |.|          5|          3|         12|    0.7123056014|          5|00:00:00|+1.00|1.00|0.0000|
#> |.|          6|          3|         12|    0.7123056014|          5|00:00:00|+1.00|1.00|0.0000|
#> |.|          4|          3|         12|    0.7123056014|          5|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          12    mean value:     0.7123056014
#> Description of the best-so-far configuration:
#>    .ID.               x1               x2 .PARENT.
#> 12   12 3.00459676598218 1.91049682606301        5
#> 
#> # 2025-09-20 18:39:11 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  x1               x2
#> 12 3.00459676598218 1.91049682606301
#> 6  9.43809640151962 1.19790239189972
#> 5  2.15160071849823 4.15610745549202
#> # 2025-09-20 18:39:11 UTC: Stopped because there is not enough budget left to race more than the minimum (3).
#> # You may either increase the budget or set 'minNbSurvival' to a lower value.
#> # Iteration: 6
#> # nbIterations: 6
#> # experimentsUsed: 89
#> # timeUsed: 0
#> # remainingBudget: 7
#> # currentBudget: 7
#> # number of elites: 3
#> # nbConfigurations: 3
#> # Total CPU user time: 1.673, CPU sys time: 0.021, Wall-clock time: 1.694
#> # 2025-09-20 18:39:11 UTC: Starting post-selection:
#> # Configurations selected: 12, 6, 5, 1.
#> # Pending instances: 1, 1, 1, 4.
#> # 2025-09-20 18:39:11 UTC: seed: 438793322
#> # Configurations: 4
#> # Available experiments: 7
#> # minSurvival: 1
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded.
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          7|          4|         12|    0.7123056014|          1|00:00:00|   NA|  NA|    NA|
#> |.|          1|          4|         12|    0.7123056014|          1|00:00:00|+1.00|1.00|0.0000|
#> |.|          4|          4|         12|    0.7123056014|          1|00:00:00|+1.00|1.00|0.0000|
#> |.|          5|          4|         12|    0.7123056014|          1|00:00:00|+1.00|1.00|0.0000|
#> |x|          8|          4|         12|    0.7123056014|          2|00:00:00|+1.00|1.00|0.0000|
#> |.|          3|          4|         12|    0.7123056014|          2|00:00:00|+1.00|1.00|0.0000|
#> |.|          2|          4|         12|    0.7123056014|          2|00:00:00|+1.00|1.00|0.0000|
#> |-|          6|          1|         12|    0.7123056014|          3|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          12    mean value:     0.7123056014
#> Description of the best-so-far configuration:
#>    .ID.               x1               x2 .PARENT.
#> 12   12 3.00459676598218 1.91049682606301        5
#> 
#> # 2025-09-20 18:39:11 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  x1               x2
#> 12 3.00459676598218 1.91049682606301
#> # Total CPU user time: 1.967, CPU sys time: 0.031, Wall-clock time: 1.998
#>          x1       x2 configuration  x_domain         y
#>       <num>    <num>         <int>    <list>     <num>
#> 1: 3.004597 1.910497            12 <list[2]> 0.7123056

# best scoring configuration
instance$result
#>          x1       x2 configuration  x_domain         y
#>       <num>    <num>         <int>    <list>     <num>
#> 1: 3.004597 1.910497            12 <list[2]> 0.7123056

# all evaluations
as.data.table(instance$archive)
#>             x1         x2           y  race  step instance configuration
#>          <num>      <num>       <num> <int> <int>    <int>         <int>
#>  1:  4.0266007  9.7811070  69.4498893     1     1        1             1
#>  2: -3.4733993  2.2811072 117.6810288     1     1        1             2
#>  3:  0.2766007 13.5311070  82.6217692     1     1        1             3
#>  4:  7.7766007  6.0311075  31.8604553     1     1        1             4
#>  5:  2.1516007  4.1561075   5.6965446     1     1        1             5
#>  6:  4.0266007  9.7811070  69.4498893     1     1        5             1
#>  7: -3.4733993  2.2811072 117.6810288     1     1        5             2
#>  8:  0.2766007 13.5311070  82.6217692     1     1        5             3
#>  9:  7.7766007  6.0311075  31.8604553     1     1        5             4
#> 10:  2.1516007  4.1561075   5.6965446     1     1        5             5
#> 11:  4.0266007  9.7811070  69.4498893     1     1        8             1
#> 12: -3.4733993  2.2811072 117.6810288     1     1        8             2
#> 13:  0.2766007 13.5311070  82.6217692     1     1        8             3
#> 14:  7.7766007  6.0311075  31.8604553     1     1        8             4
#> 15:  2.1516007  4.1561075   5.6965446     1     1        8             5
#> 16:  4.0266007  9.7811070  69.4498893     1     1        2             1
#> 17: -3.4733993  2.2811072 117.6810288     1     1        2             2
#> 18:  0.2766007 13.5311070  82.6217692     1     1        2             3
#> 19:  7.7766007  6.0311075  31.8604553     1     1        2             4
#> 20:  2.1516007  4.1561075   5.6965446     1     1        2             5
#> 21:  4.0266007  9.7811070  69.4498893     1     1        7             1
#> 22: -3.4733993  2.2811072 117.6810288     1     1        7             2
#> 23:  0.2766007 13.5311070  82.6217692     1     1        7             3
#> 24:  7.7766007  6.0311075  31.8604553     1     1        7             4
#> 25:  2.1516007  4.1561075   5.6965446     1     1        7             5
#> 26:  2.1516007  4.1561075   5.6965446     2     1        4             5
#> 27:  9.4380964  1.1979024   2.0585973     2     1        4             6
#> 28:  2.6662158 10.7114427  66.0482550     2     1        4             7
#> 29: -0.9377328  6.2995126  17.3878051     2     1        4             8
#> 30: -0.2024575  4.5803442  22.4586034     2     1        4             9
#> 31:  9.4380964  1.1979024   2.0585973     2     1        1             6
#> 32:  2.6662158 10.7114427  66.0482550     2     1        1             7
#> 33: -0.9377328  6.2995126  17.3878051     2     1        1             8
#> 34: -0.2024575  4.5803442  22.4586034     2     1        1             9
#> 35:  9.4380964  1.1979024   2.0585973     2     1        2             6
#> 36:  2.6662158 10.7114427  66.0482550     2     1        2             7
#> 37: -0.9377328  6.2995126  17.3878051     2     1        2             8
#> 38: -0.2024575  4.5803442  22.4586034     2     1        2             9
#> 39:  9.4380964  1.1979024   2.0585973     2     1        7             6
#> 40:  2.6662158 10.7114427  66.0482550     2     1        7             7
#> 41: -0.9377328  6.2995126  17.3878051     2     1        7             8
#> 42: -0.2024575  4.5803442  22.4586034     2     1        7             9
#> 43:  9.4380964  1.1979024   2.0585973     2     1        5             6
#> 44:  2.6662158 10.7114427  66.0482550     2     1        5             7
#> 45: -0.9377328  6.2995126  17.3878051     2     1        5             8
#> 46: -0.2024575  4.5803442  22.4586034     2     1        5             9
#> 47:  2.1516007  4.1561075   5.6965446     3     1        9             5
#> 48:  9.4380964  1.1979024   2.0585973     3     1        9             6
#> 49:  7.4659299  3.1320151  16.9226397     3     1        9            10
#> 50:  7.5366533  0.7759566  13.3175340     3     1        9            11
#> 51:  3.0045968  1.9104968   0.7123056     3     1        9            12
#> 52:  4.7920415  1.8426197  11.0168730     3     1        9            13
#> 53:  4.1408296  3.9262104  10.1026354     3     1        9            14
#> 54:  9.4380964  1.1979024   2.0585973     3     1        8             6
#> 55:  7.4659299  3.1320151  16.9226397     3     1        8            10
#> 56:  7.5366533  0.7759566  13.3175340     3     1        8            11
#> 57:  3.0045968  1.9104968   0.7123056     3     1        8            12
#> 58:  4.7920415  1.8426197  11.0168730     3     1        8            13
#> 59:  4.1408296  3.9262104  10.1026354     3     1        8            14
#> 60:  7.4659299  3.1320151  16.9226397     3     1        5            10
#> 61:  7.5366533  0.7759566  13.3175340     3     1        5            11
#> 62:  3.0045968  1.9104968   0.7123056     3     1        5            12
#> 63:  4.7920415  1.8426197  11.0168730     3     1        5            13
#> 64:  4.1408296  3.9262104  10.1026354     3     1        5            14
#> 65:  7.4659299  3.1320151  16.9226397     3     1        2            10
#> 66:  7.5366533  0.7759566  13.3175340     3     1        2            11
#> 67:  3.0045968  1.9104968   0.7123056     3     1        2            12
#> 68:  4.7920415  1.8426197  11.0168730     3     1        2            13
#> 69:  4.1408296  3.9262104  10.1026354     3     1        2            14
#> 70:  7.4659299  3.1320151  16.9226397     3     1        1            10
#> 71:  7.5366533  0.7759566  13.3175340     3     1        1            11
#> 72:  3.0045968  1.9104968   0.7123056     3     1        1            12
#> 73:  4.7920415  1.8426197  11.0168730     3     1        1            13
#> 74:  4.1408296  3.9262104  10.1026354     3     1        1            14
#> 75:  9.4380964  1.1979024   2.0585973     4     1       10             6
#> 76:  2.1516007  4.1561075   5.6965446     4     1       10             5
#> 77:  3.0045968  1.9104968   0.7123056     4     1       10            12
#> 78:  2.3588556  2.7426348   3.2414939     4     1       10            15
#> 79:  2.3588556  2.7426348   3.2414939     4     1        9            15
#> 80:  3.0045968  1.9104968   0.7123056     4     1        4            12
#> 81:  2.3588556  2.7426348   3.2414939     4     1        4            15
#> 82:  3.0045968  1.9104968   0.7123056     4     1        7            12
#> 83:  2.3588556  2.7426348   3.2414939     4     1        7            15
#> 84:  2.3588556  2.7426348   3.2414939     4     1        5            15
#> 85:  3.6196369  3.2723983   3.2717513     5     1       10            16
#> 86:  3.6196369  3.2723983   3.2717513     5     1        8            16
#> 87:  3.6196369  3.2723983   3.2717513     5     1        5            16
#> 88:  3.6196369  3.2723983   3.2717513     5     1        1            16
#> 89:  3.6196369  3.2723983   3.2717513     5     1        9            16
#> 90:  4.0266007  9.7811070  69.4498893     6     1        9             1
#> 91:  4.0266007  9.7811070  69.4498893     6     1       10             1
#> 92:  4.0266007  9.7811070  69.4498893     6     1        4             1
#>             x1         x2           y  race  step instance configuration
#>               timestamp batch_nr
#>                  <POSc>    <int>
#>  1: 2025-09-20 18:39:09        1
#>  2: 2025-09-20 18:39:09        1
#>  3: 2025-09-20 18:39:09        1
#>  4: 2025-09-20 18:39:09        1
#>  5: 2025-09-20 18:39:09        1
#>  6: 2025-09-20 18:39:10        2
#>  7: 2025-09-20 18:39:10        2
#>  8: 2025-09-20 18:39:10        2
#>  9: 2025-09-20 18:39:10        2
#> 10: 2025-09-20 18:39:10        2
#> 11: 2025-09-20 18:39:10        3
#> 12: 2025-09-20 18:39:10        3
#> 13: 2025-09-20 18:39:10        3
#> 14: 2025-09-20 18:39:10        3
#> 15: 2025-09-20 18:39:10        3
#> 16: 2025-09-20 18:39:10        4
#> 17: 2025-09-20 18:39:10        4
#> 18: 2025-09-20 18:39:10        4
#> 19: 2025-09-20 18:39:10        4
#> 20: 2025-09-20 18:39:10        4
#> 21: 2025-09-20 18:39:10        5
#> 22: 2025-09-20 18:39:10        5
#> 23: 2025-09-20 18:39:10        5
#> 24: 2025-09-20 18:39:10        5
#> 25: 2025-09-20 18:39:10        5
#> 26: 2025-09-20 18:39:10        6
#> 27: 2025-09-20 18:39:10        6
#> 28: 2025-09-20 18:39:10        6
#> 29: 2025-09-20 18:39:10        6
#> 30: 2025-09-20 18:39:10        6
#> 31: 2025-09-20 18:39:10        7
#> 32: 2025-09-20 18:39:10        7
#> 33: 2025-09-20 18:39:10        7
#> 34: 2025-09-20 18:39:10        7
#> 35: 2025-09-20 18:39:10        8
#> 36: 2025-09-20 18:39:10        8
#> 37: 2025-09-20 18:39:10        8
#> 38: 2025-09-20 18:39:10        8
#> 39: 2025-09-20 18:39:10        9
#> 40: 2025-09-20 18:39:10        9
#> 41: 2025-09-20 18:39:10        9
#> 42: 2025-09-20 18:39:10        9
#> 43: 2025-09-20 18:39:10       10
#> 44: 2025-09-20 18:39:10       10
#> 45: 2025-09-20 18:39:10       10
#> 46: 2025-09-20 18:39:10       10
#> 47: 2025-09-20 18:39:10       11
#> 48: 2025-09-20 18:39:10       11
#> 49: 2025-09-20 18:39:10       11
#> 50: 2025-09-20 18:39:10       11
#> 51: 2025-09-20 18:39:10       11
#> 52: 2025-09-20 18:39:10       11
#> 53: 2025-09-20 18:39:10       11
#> 54: 2025-09-20 18:39:10       12
#> 55: 2025-09-20 18:39:10       12
#> 56: 2025-09-20 18:39:10       12
#> 57: 2025-09-20 18:39:10       12
#> 58: 2025-09-20 18:39:10       12
#> 59: 2025-09-20 18:39:10       12
#> 60: 2025-09-20 18:39:10       13
#> 61: 2025-09-20 18:39:10       13
#> 62: 2025-09-20 18:39:10       13
#> 63: 2025-09-20 18:39:10       13
#> 64: 2025-09-20 18:39:10       13
#> 65: 2025-09-20 18:39:10       14
#> 66: 2025-09-20 18:39:10       14
#> 67: 2025-09-20 18:39:10       14
#> 68: 2025-09-20 18:39:10       14
#> 69: 2025-09-20 18:39:10       14
#> 70: 2025-09-20 18:39:10       15
#> 71: 2025-09-20 18:39:10       15
#> 72: 2025-09-20 18:39:10       15
#> 73: 2025-09-20 18:39:10       15
#> 74: 2025-09-20 18:39:10       15
#> 75: 2025-09-20 18:39:10       16
#> 76: 2025-09-20 18:39:10       16
#> 77: 2025-09-20 18:39:10       16
#> 78: 2025-09-20 18:39:10       16
#> 79: 2025-09-20 18:39:10       17
#> 80: 2025-09-20 18:39:10       18
#> 81: 2025-09-20 18:39:10       18
#> 82: 2025-09-20 18:39:10       19
#> 83: 2025-09-20 18:39:10       19
#> 84: 2025-09-20 18:39:10       20
#> 85: 2025-09-20 18:39:10       21
#> 86: 2025-09-20 18:39:10       22
#> 87: 2025-09-20 18:39:10       23
#> 88: 2025-09-20 18:39:11       24
#> 89: 2025-09-20 18:39:11       25
#> 90: 2025-09-20 18:39:11       26
#> 91: 2025-09-20 18:39:11       27
#> 92: 2025-09-20 18:39:11       28
#>               timestamp batch_nr
# }