Optimization via Iterated Racing

OptimizerIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi:10.1016/j.orp.2016.09.002 .

Parameters

instances

list()
A list of instances where the configurations executed on.

targetRunnerParallel

function()
A function that executes the objective function with a specific parameter configuration and instance. A default function is provided, see section "Target Runner and Instances".

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with OptimizerIrace.

In contrast to irace::defaultScenario(), we set digits = 15. This represents double parameters with a higher precision and avoids rounding errors.

Target Runner and Instances

The irace package uses a targetRunner script or R function to evaluate a configuration on a particular instance. Usually it is not necessary to specify a targetRunner function when using OptimizerIrace. A default function is used that forwards several configurations and instances to the user defined objective function. As usually, the user defined function has a xs, xss or xdt parameter depending on the used Objective class. For irace, the function needs an additional instances parameter.

fun = function(xs, instances) {
 # function to evaluate configuration in `xs` on instance `instances`
}

Archive

The Archive holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The optimization result (instance$result) is the best performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("irace")
opt("irace")

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerIrace

Methods

Public methods

OptimizerIrace$new()

OptimizerIrace$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15)
)

codomain = ps(y = p_dbl(tags = "minimize"))

# branin function with noise
# the noise generates different instances of the branin function
# the noise values are passed via the `instances` parameter
fun = function(xdt, instances) {
  ys = branin(xdt[["x1"]], xdt[["x2"]], noise = as.numeric(instances))
  data.table(y = ys)
}

# define objective with instances as a constant
objective = ObjectiveRFunDt$new(
 fun = fun,
 domain = domain,
 codomain = codomain,
 constants = ps(instances = p_uty()))

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 1000))

# create instances of branin function
instances = rnorm(10, mean = 0, sd = 0.1)

# load optimizer irace and set branin instances
optimizer = opt("irace", instances = instances)

# modifies the instance by reference
optimizer$optimize(instance)
#> # 2023-10-23 10:53:51 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 3
#> # minNbSurvival: 3
#> # nbParameters: 2
#> # seed: 1756483848
#> # confidence level: 0.95
#> # budget: 1000
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2023-10-23 10:53:51 UTC: Iteration 1 of 3
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 1000
#> # currentBudget: 333
#> # nbConfigurations: 55
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          1|         55|         34|     1.892673351|         55|00:00:00|   NA|  NA|    NA|
#> |x|          2|         55|         34|     1.892673351|        110|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         55|         34|     1.892673351|        165|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         55|         34|     1.892673351|        220|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          1|         34|     1.892673351|        275|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          34    mean value:      1.892673351
#> Description of the best-so-far configuration:
#>    .ID.                x1               x2 .PARENT.
#> 34   34 -2.78157702297904 10.4888794384897       NA
#> 
#> # 2023-10-23 10:53:51 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                   x1               x2
#> 34 -2.78157702297904 10.4888794384897
#> # 2023-10-23 10:53:52 UTC: Iteration 2 of 3
#> # experimentsUsedSoFar: 275
#> # remainingBudget: 725
#> # currentBudget: 362
#> # nbConfigurations: 52
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          6|         52|         56|     1.180358589|         52|00:00:00|   NA|  NA|    NA|
#> |x|          2|         52|         56|     1.180358589|        103|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|         52|         56|     1.180358589|        154|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         52|         56|     1.180358589|        205|00:00:00|+1.00|1.00|0.0000|
#> |-|          3|          2|         56|     1.180358589|        256|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          34    mean value:      1.892673351
#> Description of the best-so-far configuration:
#>    .ID.                x1               x2 .PARENT.
#> 34   34 -2.78157702297904 10.4888794384897       NA
#> 
#> # 2023-10-23 10:53:52 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                   x1               x2
#> 34 -2.78157702297904 10.4888794384897
#> 56 -2.83249038767195 10.9723083132923
#> # 2023-10-23 10:53:52 UTC: Iteration 3 of 3
#> # experimentsUsedSoFar: 531
#> # remainingBudget: 469
#> # currentBudget: 469
#> # nbConfigurations: 60
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          7|         60|        122|     1.048862628|         60|00:00:00|   NA|  NA|    NA|
#> |x|          6|         60|        122|     1.048862628|        118|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|         60|        122|     1.048862628|        177|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|         60|        122|     1.048862628|        235|00:00:00|+1.00|1.00|0.0000|
#> |-|          4|          3|        122|     1.048862628|        293|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          56    mean value:      1.180358589
#> Description of the best-so-far configuration:
#>    .ID.                x1               x2 .PARENT.
#> 56   56 -2.83249038767195 10.9723083132923       34
#> 
#> # 2023-10-23 10:53:53 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                    x1               x2
#> 56  -2.83249038767195 10.9723083132923
#> 34  -2.78157702297904 10.4888794384897
#> 122 -2.85197330709924 11.0887395979662
#> # 2023-10-23 10:53:53 UTC: Iteration 4 of 4
#> # experimentsUsedSoFar: 824
#> # remainingBudget: 176
#> # currentBudget: 176
#> # nbConfigurations: 21
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          8|         21|        173|     1.036854646|         21|00:00:00|   NA|  NA|    NA|
#> |x|          6|         21|        173|     1.036854646|         39|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         21|        173|     1.036854646|         58|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         21|        173|     1.036854646|         76|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          4|        173|     1.036854646|         94|00:00:00|+1.00|1.00|0.0000|
#> |!|          2|          4|        173|     1.036854646|         95|00:00:00|+1.00|1.00|0.0000|
#> |!|          1|          4|        173|     1.036854646|         97|00:00:00|+1.00|1.00|0.0000|
#> |-|          7|          1|        173|     1.036854646|         98|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         173    mean value:      1.036854646
#> Description of the best-so-far configuration:
#>     .ID.                x1               x2 .PARENT.
#> 173  173 -2.85842589878343 11.0983157930833      122
#> 
#> # 2023-10-23 10:53:53 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                    x1               x2
#> 173 -2.85842589878343 11.0983157930833
#> # 2023-10-23 10:53:53 UTC: Iteration 5 of 5
#> # experimentsUsedSoFar: 922
#> # remainingBudget: 78
#> # currentBudget: 78
#> # nbConfigurations: 8
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          9|          8|        187|     1.010471843|          8|00:00:00|   NA|  NA|    NA|
#> |x|          6|          8|        187|     1.010471843|         15|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          8|        187|     1.010471843|         22|00:00:00|+1.00|1.00|0.0000|
#> |x|          7|          8|        187|     1.010471843|         29|00:00:00|+1.00|1.00|0.0000|
#> |-|          2|          2|        187|     1.010471843|         36|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         173    mean value:      1.036854646
#> Description of the best-so-far configuration:
#>     .ID.                x1               x2 .PARENT.
#> 173  173 -2.85842589878343 11.0983157930833      122
#> 
#> # 2023-10-23 10:53:53 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                    x1               x2
#> 173 -2.85842589878343 11.0983157930833
#> 187 -2.85759757521311 11.1254599014628
#> # 2023-10-23 10:53:53 UTC: Iteration 6 of 6
#> # experimentsUsedSoFar: 958
#> # remainingBudget: 42
#> # currentBudget: 42
#> # nbConfigurations: 6
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|         10|          6|        187|     1.010471843|          6|00:00:00|   NA|  NA|    NA|
#> |x|          8|          6|        187|     1.010471843|         11|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|          6|        187|     1.010471843|         15|00:00:00|+1.00|1.00|0.0000|
#> |x|          7|          6|        187|     1.010471843|         19|00:00:00|+1.00|1.00|0.0000|
#> |-|          3|          2|        187|     1.010471843|         24|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         173    mean value:      1.036854646
#> Description of the best-so-far configuration:
#>     .ID.                x1               x2 .PARENT.
#> 173  173 -2.85842589878343 11.0983157930833      122
#> 
#> # 2023-10-23 10:53:53 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                    x1               x2
#> 173 -2.85842589878343 11.0983157930833
#> 187 -2.85759757521311 11.1254599014628
#> # 2023-10-23 10:53:53 UTC: Stopped because there is not enough budget left to race more than the minimum (3)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 7
#> # nbIterations: 7
#> # experimentsUsedSoFar: 982
#> # timeUsed: 0
#> # remainingBudget: 18
#> # currentBudget: 18
#> # number of elites: 2
#> # nbConfigurations: 3
#> # Total CPU user time: 2.345, CPU sys time: 0.032, Wall-clock time: 2.377
#>           x1       x2 configuration  x_domain        y
#> 1: -2.858426 11.09832           173 <list[2]> 1.036855

# best scoring configuration
instance$result
#>           x1       x2 configuration  x_domain        y
#> 1: -2.858426 11.09832           173 <list[2]> 1.036855

# all evaluations
as.data.table(instance$archive)
#>               x1         x2         y race step instance configuration
#>   1:  8.64041519  0.8388212  4.314233    1    1       10             1
#>   2: -0.19274366  0.3548800 54.906344    1    1       10             2
#>   3:  0.03948434 10.6441821 41.748802    1    1       10             3
#>   4:  3.27771394  5.9142289 14.496688    1    1       10             4
#>   5:  6.34298587  9.1058690 83.641018    1    1       10             5
#>  ---                                                                  
#> 978: -2.85759758 11.1254599  1.010472    6    1        2           187
#> 979: -2.86940118 11.0981622  1.034707    6    1        2           190
#> 980: -2.85245366 11.1015568  1.035960    6    1        2           191
#> 981: -2.86402343 11.1005584  1.033044    6    1        2           192
#> 982: -2.86337834 11.1050461  1.028580    6    1        2           193
#>                timestamp batch_nr
#>   1: 2023-10-23 10:53:51        1
#>   2: 2023-10-23 10:53:51        1
#>   3: 2023-10-23 10:53:51        1
#>   4: 2023-10-23 10:53:51        1
#>   5: 2023-10-23 10:53:51        1
#>  ---                             
#> 978: 2023-10-23 10:53:53       33
#> 979: 2023-10-23 10:53:53       33
#> 980: 2023-10-23 10:53:53       33
#> 981: 2023-10-23 10:53:53       33
#> 982: 2023-10-23 10:53:53       33