Optimization via Iterated Racing

OptimizerIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi:10.1016/j.orp.2016.09.002 .

Parameters

instances

list()
A list of instances where the configurations executed on.

targetRunnerParallel

function()
A function that executes the objective function with a specific parameter configuration and instance. A default function is provided, see section "Target Runner and Instances".

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with OptimizerIrace.

In contrast to irace::defaultScenario(), we set digits = 15. This represents double parameters with a higher precision and avoids rounding errors.

Target Runner and Instances

The irace package uses a targetRunner script or R function to evaluate a configuration on a particular instance. Usually it is not necessary to specify a targetRunner function when using OptimizerIrace. A default function is used that forwards several configurations and instances to the user defined objective function. As usually, the user defined function has a xs, xss or xdt parameter depending on the used Objective class. For irace, the function needs an additional instances parameter.

fun = function(xs, instances) {
 # function to evaluate configuration in `xs` on instance `instances`
}

Archive

The Archive holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The optimization result (instance$result) is the best performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("irace")
opt("irace")

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerIrace

Methods

Public methods

OptimizerIrace$new()

OptimizerIrace$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15)
)

codomain = ps(y = p_dbl(tags = "minimize"))

# branin function with noise
# the noise generates different instances of the branin function
# the noise values are passed via the `instances` parameter
fun = function(xdt, instances) {
  ys = branin(xdt[["x1"]], xdt[["x2"]], noise = as.numeric(instances))
  data.table(y = ys)
}

# define objective with instances as a constant
objective = ObjectiveRFunDt$new(
 fun = fun,
 domain = domain,
 codomain = codomain,
 constants = ps(instances = p_uty()))

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 1000))

# create instances of branin function
instances = rnorm(10, mean = 0, sd = 0.1)

# load optimizer irace and set branin instances
optimizer = opt("irace", instances = instances)

# modifies the instance by reference
optimizer$optimize(instance)
#> # 2022-12-07 17:00:40 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 3
#> # minNbSurvival: 3
#> # nbParameters: 2
#> # seed: 1995445295
#> # confidence level: 0.95
#> # budget: 1000
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2022-12-07 17:00:41 UTC: Iteration 1 of 3
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 1000
#> # currentBudget: 333
#> # nbConfigurations: 55
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          1|         55|         46|     1.918256634|         55|00:00:00|   NA|  NA|    NA|
#> |x|          2|         55|         46|     1.918256634|        110|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         55|         46|     1.918256634|        165|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         55|         46|     1.918256634|        220|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          1|         46|     1.918256634|        275|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          46    mean value:      1.918256634
#> Description of the best-so-far configuration:
#>    .ID.               x1               x2 .PARENT.
#> 46   46 8.90124030294828 1.58482683473267       NA
#> 
#> # 2022-12-07 17:00:41 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  x1               x2
#> 46 8.90124030294828 1.58482683473267
#> # 2022-12-07 17:00:41 UTC: Iteration 2 of 3
#> # experimentsUsedSoFar: 275
#> # remainingBudget: 725
#> # currentBudget: 362
#> # nbConfigurations: 52
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          6|         52|         78|    0.5891280296|         52|00:00:00|   NA|  NA|    NA|
#> |x|          2|         52|         78|    0.5891280296|        103|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|         52|         78|    0.5891280296|        154|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         52|         78|    0.5891280296|        205|00:00:00|+1.00|1.00|0.0000|
#> |-|          3|          2|         78|    0.5891280296|        256|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          46    mean value:      1.918256634
#> Description of the best-so-far configuration:
#>    .ID.               x1               x2 .PARENT.
#> 46   46 8.90124030294828 1.58482683473267       NA
#> 
#> # 2022-12-07 17:00:41 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  x1               x2
#> 46 8.90124030294828 1.58482683473267
#> 78 9.43272594273452 2.04474862189777
#> # 2022-12-07 17:00:41 UTC: Iteration 3 of 3
#> # experimentsUsedSoFar: 531
#> # remainingBudget: 469
#> # currentBudget: 469
#> # nbConfigurations: 60
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          7|         60|        122|    0.4248091499|         60|00:00:00|   NA|  NA|    NA|
#> |x|          4|         60|        122|    0.4248091499|        118|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|         60|        122|    0.4248091499|        176|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|         60|        122|    0.4248091499|        235|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          3|        122|    0.4248091499|        293|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:          78    mean value:     0.5891280296
#> Description of the best-so-far configuration:
#>    .ID.               x1               x2 .PARENT.
#> 78   78 9.43272594273452 2.04474862189777       46
#> 
#> # 2022-12-07 17:00:42 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                   x1               x2
#> 78  9.43272594273452 2.04474862189777
#> 46  8.90124030294828 1.58482683473267
#> 122 9.46712727840444 2.37563034970815
#> # 2022-12-07 17:00:42 UTC: Iteration 4 of 4
#> # experimentsUsedSoFar: 824
#> # remainingBudget: 176
#> # currentBudget: 176
#> # nbConfigurations: 21
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          8|         21|        169|    0.4057007412|         21|00:00:00|   NA|  NA|    NA|
#> |x|          5|         21|        169|    0.4057007412|         39|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|         21|        169|    0.4057007412|         57|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|         21|        169|    0.4057007412|         76|00:00:00|+1.00|1.00|0.0000|
#> |-|          6|          4|        169|    0.4057007412|         95|00:00:00|+1.00|1.00|0.0000|
#> |!|          1|          4|        169|    0.4057007412|         96|00:00:00|+1.00|1.00|0.0000|
#> |!|          3|          4|        169|    0.4057007412|         97|00:00:00|+1.00|1.00|0.0000|
#> |-|          7|          1|        169|    0.4057007412|         98|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         169    mean value:     0.4057007412
#> Description of the best-so-far configuration:
#>     .ID.               x1               x2 .PARENT.
#> 169  169 9.40168220320214 2.38311269330349      122
#> 
#> # 2022-12-07 17:00:42 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                   x1               x2
#> 169 9.40168220320214 2.38311269330349
#> # 2022-12-07 17:00:42 UTC: Iteration 5 of 5
#> # experimentsUsedSoFar: 922
#> # remainingBudget: 78
#> # currentBudget: 78
#> # nbConfigurations: 8
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          9|          8|        187|    0.4028494287|          8|00:00:00|   NA|  NA|    NA|
#> |x|          4|          8|        187|    0.4028494287|         15|00:00:00|+1.00|1.00|0.0000|
#> |x|          2|          8|        187|    0.4028494287|         22|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|          8|        187|    0.4028494287|         29|00:00:00|+1.00|1.00|0.0000|
#> |-|          8|          2|        187|    0.4028494287|         36|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         169    mean value:     0.4057007412
#> Description of the best-so-far configuration:
#>     .ID.               x1               x2 .PARENT.
#> 169  169 9.40168220320214 2.38311269330349      122
#> 
#> # 2022-12-07 17:00:42 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                   x1               x2
#> 169 9.40168220320214 2.38311269330349
#> 187 9.40948767942250 2.40016768966248
#> # 2022-12-07 17:00:42 UTC: Iteration 6 of 6
#> # experimentsUsedSoFar: 958
#> # remainingBudget: 42
#> # currentBudget: 42
#> # nbConfigurations: 6
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|         10|          6|        187|    0.4028494287|          6|00:00:00|   NA|  NA|    NA|
#> |x|          1|          6|        187|    0.4028494287|         11|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|          6|        187|    0.4028494287|         16|00:00:00|+1.00|1.00|0.0000|
#> |x|          9|          6|        187|    0.4028494287|         20|00:00:00|+1.00|1.00|0.0000|
#> |-|          8|          2|        187|    0.4028494287|         24|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         169    mean value:     0.4057007412
#> Description of the best-so-far configuration:
#>     .ID.               x1               x2 .PARENT.
#> 169  169 9.40168220320214 2.38311269330349      122
#> 
#> # 2022-12-07 17:00:42 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                   x1               x2
#> 169 9.40168220320214 2.38311269330349
#> 187 9.40948767942250 2.40016768966248
#> # 2022-12-07 17:00:42 UTC: Stopped because there is not enough budget left to race more than the minimum (3)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 7
#> # nbIterations: 7
#> # experimentsUsedSoFar: 982
#> # timeUsed: 0
#> # remainingBudget: 18
#> # currentBudget: 18
#> # number of elites: 2
#> # nbConfigurations: 3
#> # Total CPU user time: 1.907, CPU sys time: 0.024, Wall-clock time: 1.932
#>          x1       x2 configuration  x_domain         y
#> 1: 9.401682 2.383113           169 <list[2]> 0.4057007

# best scoring configuration
instance$result
#>          x1       x2 configuration  x_domain         y
#> 1: 9.401682 2.383113           169 <list[2]> 0.4057007

# all evaluations
as.data.table(instance$archive)
#>              x1       x2           y race step instance configuration
#>   1: -3.6889261 3.413431 106.1599750    1    1        6             1
#>   2:  0.5437219 4.731753  18.4119341    1    1        6             2
#>   3: -1.2836489 8.734496  12.9485846    1    1        6             3
#>   4:  8.5486012 2.048512   3.8992982    1    1        6             4
#>   5:  9.5657450 3.935385   2.2858136    1    1        6             5
#>  ---                                                                 
#> 978:  9.4039970 2.400386   0.4032257    6    1        3           193
#> 979:  9.3960858 2.411311   0.4034071    6    1        5           190
#> 980:  9.3908836 2.377121   0.4082239    6    1        5           191
#> 981:  9.4022699 2.393959   0.4041785    6    1        5           192
#> 982:  9.4039970 2.400386   0.4032257    6    1        5           193
#>                timestamp batch_nr
#>   1: 2022-12-07 17:00:41        1
#>   2: 2022-12-07 17:00:41        1
#>   3: 2022-12-07 17:00:41        1
#>   4: 2022-12-07 17:00:41        1
#>   5: 2022-12-07 17:00:41        1
#>  ---                             
#> 978: 2022-12-07 17:00:42       32
#> 979: 2022-12-07 17:00:42       33
#> 980: 2022-12-07 17:00:42       33
#> 981: 2022-12-07 17:00:42       33
#> 982: 2022-12-07 17:00:42       33