Optimization via Local Search

OptimizerBatchLocalSearch class that implements a simple Local Search. Local Search starts by determining the n_initial_points initial best points present in the Archive of the OptimInstance. If fewer points than n_initial_points are present, additional initial_random_sample_size points sampled uniformly at random are evaluated and the best n_initial_points initial points are determined.

In each iteration, for each of the n_initial_points initial best points, neighbors_per_point neighbors are generated by local mutation. Local mutation generates a neighbor by sampling a single parameter that is to be mutated and then proceeds as follows: Double parameters (paradox::p_dbl()) are mutated via Gaussian mutation (with a prior standardization to [0, 1] and retransformation after mutation). Integer parameters (paradox::p_int()) undergo the same mutation but are rounded to the closest integer after mutation. Categorical parameters (paradox::p_fct() and paradox::p_lgl()) are mutated via uniform mutation. Note that parameters that are conditioned on (i.e., they are parents of a paradox::Condition, see the dependencies of the search space) are not mutated.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("local_search")
opt("local_search")

Parameters

n_initial_points

integer(1)
Size of the set of initial best points which are used as starting points for the Local Search. Default is 10.

initial_random_sample_size

integer(1)
Number of points that are sampled uniformly at random before the best n_initial_points initial points are determined, if fewer points than n_initial_points are present in the Archive of the OptimInstance. Default is 100.

neighbors_per_point

integer(1)
Number of neighboring points to generate for each of the n_initial_points best starting points in each iteration. Default is 100.

mutation_sd

numeric(1)
Standard deviation used to create neighbors during mutation of numeric parameters on the standardized [0, 1] scale. Default is 0.1.

Archive

The Archive holds the following additional column that is specific to the algorithm:

.point_id (integer(1))
The id (1, ..., n_initial_points) indicating from which of the n_initial_points best points the evaluated point was generated from.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super classes

bbotk::Optimizer -> bbotk::OptimizerBatch -> OptimizerBatchLocalSearch

Methods

Public methods

OptimizerBatchLocalSearch$new()

OptimizerBatchLocalSearch$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$print()

bbotk::OptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerBatchLocalSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerBatchLocalSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

search_space = domain = ps(x = p_dbl(lower = -1, upper = 1))

codomain = ps(y = p_dbl(tags = "minimize"))

objective_function = function(xs) {
  list(y = as.numeric(xs)^2)
}

objective = ObjectiveRFun$new(
 fun = objective_function,
 domain = domain,
 codomain = codomain)

instance = OptimInstanceBatchSingleCrit$new(
 objective = objective,
 search_space = search_space,
 terminator = trm("evals", n_evals = 100))

# evaluate an initial sample of 10 points uniformly at random
# choose the best 3 points as the initial points
# for each of these points generate 10 neighbors
# repeat this process
optimizer = opt("local_search",
  n_initial_points = 3,
  initial_random_sample_size = 10,
  neighbors_per_point = 10)

# modifies the instance by reference
optimizer$optimize(instance)
#>               x  x_domain            y
#>           <num>    <list>        <num>
#> 1: -0.003576096 <list[1]> 1.278846e-05

# returns best scoring evaluation
instance$result
#>               x  x_domain            y
#>           <num>    <list>        <num>
#> 1: -0.003576096 <list[1]> 1.278846e-05

# allows access of data.table of full path of all evaluations
as.data.table(instance$archive$data)
#>                 x            y  x_domain           timestamp batch_nr .point_id
#>             <num>        <num>    <list>              <POSc>    <int>     <int>
#>   1: -0.954400192 9.108797e-01 <list[1]> 2024-11-26 15:25:27        1        NA
#>   2:  0.577125689 3.330741e-01 <list[1]> 2024-11-26 15:25:27        1        NA
#>   3:  0.946978642 8.967685e-01 <list[1]> 2024-11-26 15:25:27        1        NA
#>   4: -0.188561180 3.555532e-02 <list[1]> 2024-11-26 15:25:27        1        NA
#>   5:  0.944853830 8.927488e-01 <list[1]> 2024-11-26 15:25:27        1        NA
#>   6: -0.497483826 2.474902e-01 <list[1]> 2024-11-26 15:25:27        1        NA
#>   7: -0.407919032 1.663979e-01 <list[1]> 2024-11-26 15:25:27        1        NA
#>   8: -0.611497788 3.739295e-01 <list[1]> 2024-11-26 15:25:27        1        NA
#>   9: -0.780039704 6.084619e-01 <list[1]> 2024-11-26 15:25:27        1        NA
#>  10:  0.983069457 9.664256e-01 <list[1]> 2024-11-26 15:25:27        1        NA
#>  11: -0.039024270 1.522894e-03 <list[1]> 2024-11-26 15:25:28        2         1
#>  12:  0.063826659 4.073842e-03 <list[1]> 2024-11-26 15:25:28        2         1
#>  13: -0.532466579 2.835207e-01 <list[1]> 2024-11-26 15:25:28        2         1
#>  14: -0.409938822 1.680498e-01 <list[1]> 2024-11-26 15:25:28        2         1
#>  15:  0.026710983 7.134766e-04 <list[1]> 2024-11-26 15:25:28        2         1
#>  16: -0.288154464 8.303300e-02 <list[1]> 2024-11-26 15:25:28        2         1
#>  17:  0.074868694 5.605321e-03 <list[1]> 2024-11-26 15:25:28        2         1
#>  18:  0.017018912 2.896434e-04 <list[1]> 2024-11-26 15:25:28        2         1
#>  19: -0.199330053 3.973247e-02 <list[1]> 2024-11-26 15:25:28        2         1
#>  20: -0.201953347 4.078515e-02 <list[1]> 2024-11-26 15:25:28        2         1
#>  21: -0.271722883 7.383333e-02 <list[1]> 2024-11-26 15:25:28        2         2
#>  22: -0.503341945 2.533531e-01 <list[1]> 2024-11-26 15:25:28        2         2
#>  23: -0.183549729 3.369050e-02 <list[1]> 2024-11-26 15:25:28        2         2
#>  24: -0.477364038 2.278764e-01 <list[1]> 2024-11-26 15:25:28        2         2
#>  25: -0.533855957 2.850022e-01 <list[1]> 2024-11-26 15:25:28        2         2
#>  26: -0.547760347 3.000414e-01 <list[1]> 2024-11-26 15:25:28        2         2
#>  27: -0.240274009 5.773160e-02 <list[1]> 2024-11-26 15:25:28        2         2
#>  28:  0.102163246 1.043733e-02 <list[1]> 2024-11-26 15:25:28        2         2
#>  29: -0.595853461 3.550413e-01 <list[1]> 2024-11-26 15:25:28        2         2
#>  30: -0.565952661 3.203024e-01 <list[1]> 2024-11-26 15:25:28        2         2
#>  31: -0.534537585 2.857304e-01 <list[1]> 2024-11-26 15:25:28        2         3
#>  32: -0.338341833 1.144752e-01 <list[1]> 2024-11-26 15:25:28        2         3
#>  33: -0.809975336 6.560600e-01 <list[1]> 2024-11-26 15:25:28        2         3
#>  34: -0.671739311 4.512337e-01 <list[1]> 2024-11-26 15:25:28        2         3
#>  35: -0.759721489 5.771767e-01 <list[1]> 2024-11-26 15:25:28        2         3
#>  36: -0.544639490 2.966322e-01 <list[1]> 2024-11-26 15:25:28        2         3
#>  37: -0.460214283 2.117972e-01 <list[1]> 2024-11-26 15:25:28        2         3
#>  38: -0.058247365 3.392756e-03 <list[1]> 2024-11-26 15:25:28        2         3
#>  39: -0.084526772 7.144775e-03 <list[1]> 2024-11-26 15:25:28        2         3
#>  40: -0.483267941 2.335479e-01 <list[1]> 2024-11-26 15:25:28        2         3
#>  41:  0.129157773 1.668173e-02 <list[1]> 2024-11-26 15:25:28        3         1
#>  42: -0.039350442 1.548457e-03 <list[1]> 2024-11-26 15:25:28        3         1
#>  43: -0.044046534 1.940097e-03 <list[1]> 2024-11-26 15:25:28        3         1
#>  44:  0.323200747 1.044587e-01 <list[1]> 2024-11-26 15:25:28        3         1
#>  45:  0.233518454 5.453087e-02 <list[1]> 2024-11-26 15:25:28        3         1
#>  46: -0.133642779 1.786039e-02 <list[1]> 2024-11-26 15:25:28        3         1
#>  47:  0.356372152 1.270011e-01 <list[1]> 2024-11-26 15:25:28        3         1
#>  48:  0.236340695 5.585692e-02 <list[1]> 2024-11-26 15:25:28        3         1
#>  49: -0.036530442 1.334473e-03 <list[1]> 2024-11-26 15:25:28        3         1
#>  50:  0.050052539 2.505257e-03 <list[1]> 2024-11-26 15:25:28        3         1
#>  51:  0.269299095 7.252200e-02 <list[1]> 2024-11-26 15:25:28        3         2
#>  52:  0.365496632 1.335878e-01 <list[1]> 2024-11-26 15:25:28        3         2
#>  53:  0.408084580 1.665330e-01 <list[1]> 2024-11-26 15:25:28        3         2
#>  54: -0.013708960 1.879356e-04 <list[1]> 2024-11-26 15:25:28        3         2
#>  55:  0.088106471 7.762750e-03 <list[1]> 2024-11-26 15:25:28        3         2
#>  56:  0.146333079 2.141337e-02 <list[1]> 2024-11-26 15:25:28        3         2
#>  57: -0.030571781 9.346338e-04 <list[1]> 2024-11-26 15:25:28        3         2
#>  58:  0.114121060 1.302362e-02 <list[1]> 2024-11-26 15:25:28        3         2
#>  59: -0.122045296 1.489505e-02 <list[1]> 2024-11-26 15:25:28        3         2
#>  60:  0.436964397 1.909379e-01 <list[1]> 2024-11-26 15:25:28        3         2
#>  61: -0.339305263 1.151281e-01 <list[1]> 2024-11-26 15:25:28        3         3
#>  62: -0.035420468 1.254610e-03 <list[1]> 2024-11-26 15:25:28        3         3
#>  63:  0.180627334 3.262623e-02 <list[1]> 2024-11-26 15:25:28        3         3
#>  64: -0.150222336 2.256675e-02 <list[1]> 2024-11-26 15:25:28        3         3
#>  65: -0.154516859 2.387546e-02 <list[1]> 2024-11-26 15:25:28        3         3
#>  66: -0.003576096 1.278846e-05 <list[1]> 2024-11-26 15:25:28        3         3
#>  67: -0.191916635 3.683199e-02 <list[1]> 2024-11-26 15:25:28        3         3
#>  68: -0.098271054 9.657200e-03 <list[1]> 2024-11-26 15:25:28        3         3
#>  69:  0.072369314 5.237318e-03 <list[1]> 2024-11-26 15:25:28        3         3
#>  70: -0.247617007 6.131418e-02 <list[1]> 2024-11-26 15:25:28        3         3
#>  71:  0.005598677 3.134518e-05 <list[1]> 2024-11-26 15:25:28        4         1
#>  72:  0.275859813 7.609864e-02 <list[1]> 2024-11-26 15:25:28        4         1
#>  73: -0.006493060 4.215983e-05 <list[1]> 2024-11-26 15:25:28        4         1
#>  74:  0.285797544 8.168024e-02 <list[1]> 2024-11-26 15:25:28        4         1
#>  75:  0.249355769 6.217830e-02 <list[1]> 2024-11-26 15:25:28        4         1
#>  76: -0.109923470 1.208317e-02 <list[1]> 2024-11-26 15:25:28        4         1
#>  77:  0.247098783 6.105781e-02 <list[1]> 2024-11-26 15:25:28        4         1
#>  78: -0.273830197 7.498298e-02 <list[1]> 2024-11-26 15:25:28        4         1
#>  79: -0.302970558 9.179116e-02 <list[1]> 2024-11-26 15:25:28        4         1
#>  80:  0.080028696 6.404592e-03 <list[1]> 2024-11-26 15:25:28        4         1
#>  81: -0.216127987 4.671131e-02 <list[1]> 2024-11-26 15:25:28        4         2
#>  82:  0.074925420 5.613819e-03 <list[1]> 2024-11-26 15:25:28        4         2
#>  83:  0.038594841 1.489562e-03 <list[1]> 2024-11-26 15:25:28        4         2
#>  84: -0.109331457 1.195337e-02 <list[1]> 2024-11-26 15:25:28        4         2
#>  85:  0.219836215 4.832796e-02 <list[1]> 2024-11-26 15:25:28        4         2
#>  86: -0.233435016 5.449191e-02 <list[1]> 2024-11-26 15:25:28        4         2
#>  87: -0.352763435 1.244420e-01 <list[1]> 2024-11-26 15:25:28        4         2
#>  88: -0.125644332 1.578650e-02 <list[1]> 2024-11-26 15:25:28        4         2
#>  89:  0.153541888 2.357511e-02 <list[1]> 2024-11-26 15:25:28        4         2
#>  90:  0.228630552 5.227193e-02 <list[1]> 2024-11-26 15:25:28        4         2
#>  91: -0.039064898 1.526066e-03 <list[1]> 2024-11-26 15:25:28        4         3
#>  92: -0.168998198 2.856039e-02 <list[1]> 2024-11-26 15:25:28        4         3
#>  93:  0.377602571 1.425837e-01 <list[1]> 2024-11-26 15:25:28        4         3
#>  94: -0.073921292 5.464357e-03 <list[1]> 2024-11-26 15:25:28        4         3
#>  95: -0.108552879 1.178373e-02 <list[1]> 2024-11-26 15:25:28        4         3
#>  96: -0.066281266 4.393206e-03 <list[1]> 2024-11-26 15:25:28        4         3
#>  97: -0.279674970 7.821809e-02 <list[1]> 2024-11-26 15:25:28        4         3
#>  98: -0.032297021 1.043098e-03 <list[1]> 2024-11-26 15:25:28        4         3
#>  99: -0.188163972 3.540568e-02 <list[1]> 2024-11-26 15:25:28        4         3
#> 100: -0.192753651 3.715397e-02 <list[1]> 2024-11-26 15:25:28        4         3
#>                 x            y  x_domain           timestamp batch_nr .point_id