Optimization via Non-linear Optimization

OptimizerNLoptr class that implements non-linear optimization. Calls nloptr::nloptr() from package nloptr.

Usage,NULL

Source

Johnson, G S (2020). “The NLopt nonlinear-optimization package.” https://github.com/stevengj/nlopt.

Parameters

algorithm

character(1)

eval_g_ineq

function()

xtol_rel

numeric(1)

xtol_abs

numeric(1)

ftol_rel

numeric(1)

ftol_abs

numeric(1)

start_values

character(1)
Create random start values or based on center of search space? In the latter case, it is the center of the parameters before a trafo is applied.

For the meaning of the control parameters, see nloptr::nloptr() and nloptr::nloptr.print.options().

The termination conditions stopval, maxtime and maxeval of nloptr::nloptr() are deactivated and replaced by the Terminator subclasses. The x and function value tolerance termination conditions (xtol_rel = 10^-4, xtol_abs = rep(0.0, length(x0)), ftol_rel = 0.0 and ftol_abs = 0.0) are still available and implemented with their package defaults. To deactivate these conditions, set them to -1.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super class

bbotk::Optimizer -> OptimizerNLoptr

Methods

Public methods

OptimizerNLoptr$new()

OptimizerNLoptr$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerNLoptr$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerNLoptr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
if (requireNamespace("nloptr")) {

  search_space = domain = ps(x = p_dbl(lower = -1, upper = 1))

  codomain = ps(y = p_dbl(tags = "minimize"))

  objective_function = function(xs) {
    list(y = as.numeric(xs)^2)
  }

  objective = ObjectiveRFun$new(
    fun = objective_function,
    domain = domain,
    codomain = codomain)

  # We use the internal termination criterion xtol_rel
  terminator = trm("none")
  instance = OptimInstanceSingleCrit$new(
    objective = objective,
    search_space = search_space,
    terminator = terminator)

  optimizer = opt("nloptr", algorithm = "NLOPT_LN_BOBYQA")

  # Modifies the instance by reference
  optimizer$optimize(instance)

  # Returns best scoring evaluation
  instance$result

  # Allows access of data.table of full path of all evaluations
  as.data.table(instance$archive)
}
#> Loading required namespace: nloptr
#>                 x            y           timestamp batch_nr    x_domain_x
#>  1:  7.635295e-01 5.829772e-01 2021-12-01 07:58:57        1  7.635295e-01
#>  2:  7.635295e-01 5.829772e-01 2021-12-01 07:58:57        2  7.635295e-01
#>  3:  7.635295e-01 5.829772e-01 2021-12-01 07:58:57        3  7.635295e-01
#>  4:  9.408824e-01 8.852596e-01 2021-12-01 07:58:57        4  9.408824e-01
#>  5:  5.861766e-01 3.436030e-01 2021-12-01 07:58:57        5  5.861766e-01
#>  6:  4.088237e-01 1.671368e-01 2021-12-01 07:58:57        6  4.088237e-01
#>  7:  5.411785e-02 2.928742e-03 2021-12-01 07:58:57        7  5.411785e-02
#>  8:  0.000000e+00 0.000000e+00 2021-12-01 07:58:57        8  0.000000e+00
#>  9: -1.773529e-02 3.145405e-04 2021-12-01 07:58:57        9 -1.773529e-02
#> 10:  1.773529e-03 3.145405e-06 2021-12-01 07:58:57       10  1.773529e-03
#> 11: -1.773529e-04 3.145405e-08 2021-12-01 07:58:57       11 -1.773529e-04
#> 12:  1.773529e-05 3.145405e-10 2021-12-01 07:58:57       12  1.773529e-05
#> 13: -8.790079e-22 7.726548e-43 2021-12-01 07:58:58       13 -8.790079e-22
# }