Optimization via Non-linear Optimization

OptimizerBatchNLoptr class that implements non-linear optimization. Calls nloptr::nloptr() from package nloptr.

Source

Johnson, G S (2020). “The NLopt nonlinear-optimization package.” https://github.com/stevengj/nlopt.

Parameters

algorithm

character(1)

eval_g_ineq

function()

xtol_rel

numeric(1)

xtol_abs

numeric(1)

ftol_rel

numeric(1)

ftol_abs

numeric(1)

start_values

character(1)
Create random start values or based on center of search space? In the latter case, it is the center of the parameters before a trafo is applied.

For the meaning of the control parameters, see nloptr::nloptr() and nloptr::nloptr.print.options().

The termination conditions stopval, maxtime and maxeval of nloptr::nloptr() are deactivated and replaced by the Terminator subclasses. The x and function value tolerance termination conditions (xtol_rel = 10^-4, xtol_abs = rep(0.0, length(x0)), ftol_rel = 0.0 and ftol_abs = 0.0) are still available and implemented with their package defaults. To deactivate these conditions, set them to -1.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super classes

bbotk::Optimizer -> bbotk::OptimizerBatch -> OptimizerBatchNLoptr

Methods

Public methods

OptimizerBatchNLoptr$new()

OptimizerBatchNLoptr$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$print()

bbotk::OptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerBatchNLoptr$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerBatchNLoptr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
if (requireNamespace("nloptr")) {

  search_space = domain = ps(x = p_dbl(lower = -1, upper = 1))

  codomain = ps(y = p_dbl(tags = "minimize"))

  objective_function = function(xs) {
    list(y = as.numeric(xs)^2)
  }

  objective = ObjectiveRFun$new(
    fun = objective_function,
    domain = domain,
    codomain = codomain)

  # We use the internal termination criterion xtol_rel
  terminator = trm("none")
  instance = OptimInstanceBatchSingleCrit$new(
    objective = objective,
    search_space = search_space,
    terminator = terminator)

  optimizer = opt("nloptr", algorithm = "NLOPT_LN_BOBYQA")

  # Modifies the instance by reference
  optimizer$optimize(instance)

  # Returns best scoring evaluation
  instance$result

  # Allows access of data.table of full path of all evaluations
  as.data.table(instance$archive)
}
#>                 x            y           timestamp batch_nr    x_domain_x
#>             <num>        <num>              <POSc>    <int>         <num>
#>  1:  6.432053e-01 4.137131e-01 2024-11-08 08:25:18        1  6.432053e-01
#>  2:  6.432053e-01 4.137131e-01 2024-11-08 08:25:18        2  6.432053e-01
#>  3:  6.432053e-01 4.137131e-01 2024-11-08 08:25:18        3  6.432053e-01
#>  4:  9.108013e-01 8.295591e-01 2024-11-08 08:25:18        4  9.108013e-01
#>  5:  3.756093e-01 1.410824e-01 2024-11-08 08:25:18        5  3.756093e-01
#>  6:  1.080133e-01 1.166687e-02 2024-11-08 08:25:18        6  1.080133e-01
#>  7:  0.000000e+00 0.000000e+00 2024-11-08 08:25:18        7  0.000000e+00
#>  8: -2.675960e-02 7.160763e-04 2024-11-08 08:25:18        8 -2.675960e-02
#>  9:  2.675960e-03 7.160763e-06 2024-11-08 08:25:18        9  2.675960e-03
#> 10: -2.675960e-04 7.160763e-08 2024-11-08 08:25:18       10 -2.675960e-04
#> 11:  2.675960e-05 7.160763e-10 2024-11-08 08:25:18       11  2.675960e-05
#> 12: -3.091070e-21 9.554716e-42 2024-11-08 08:25:18       12 -3.091070e-21
# }