Optimization via Random Search

OptimizerBatchRandomSearch class that implements a simple Random Search.

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281–305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("random_search")
opt("random_search")

Parameters

batch_size

integer(1)
Maximum number of points to try in a batch.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Super classes

bbotk::Optimizer -> bbotk::OptimizerBatch -> OptimizerBatchRandomSearch

Methods

Public methods

OptimizerBatchRandomSearch$new()

OptimizerBatchRandomSearch$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$print()

bbotk::OptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerBatchRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerBatchRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

search_space = domain = ps(x = p_dbl(lower = -1, upper = 1))

codomain = ps(y = p_dbl(tags = "minimize"))

objective_function = function(xs) {
  list(y = as.numeric(xs)^2)
}

objective = ObjectiveRFun$new(
 fun = objective_function,
 domain = domain,
 codomain = codomain)

instance = OptimInstanceBatchSingleCrit$new(
 objective = objective,
 search_space = search_space,
 terminator = trm("evals", n_evals = 10))

optimizer = opt("random_search")

# modifies the instance by reference
optimizer$optimize(instance)
#>              x  x_domain           y
#>          <num>    <list>       <num>
#> 1: -0.09211178 <list[1]> 0.008484579

# returns best scoring evaluation
instance$result
#>              x  x_domain           y
#>          <num>    <list>       <num>
#> 1: -0.09211178 <list[1]> 0.008484579

# allows access of data.table of full path of all evaluations
as.data.table(instance$archive$data)
#>               x           y  x_domain           timestamp batch_nr
#>           <num>       <num>    <list>              <POSc>    <int>
#>  1: -0.45416286 0.206263903 <list[1]> 2025-10-07 08:56:57        1
#>  2:  0.51729917 0.267598435 <list[1]> 2025-10-07 08:56:57        2
#>  3: -0.83821702 0.702607771 <list[1]> 2025-10-07 08:56:57        3
#>  4: -0.63059655 0.397652007 <list[1]> 2025-10-07 08:56:57        4
#>  5: -0.09211178 0.008484579 <list[1]> 2025-10-07 08:56:57        5
#>  6: -0.84805343 0.719194628 <list[1]> 2025-10-07 08:56:57        6
#>  7:  0.65945804 0.434884903 <list[1]> 2025-10-07 08:56:57        7
#>  8:  0.60970179 0.371736269 <list[1]> 2025-10-07 08:56:57        8
#>  9:  0.51787971 0.268199391 <list[1]> 2025-10-07 08:56:57        9
#> 10: -0.99427599 0.988584748 <list[1]> 2025-10-07 08:56:57       10