Dictionary of Terminators

A simple mlr3misc::Dictionary storing objects of class Terminator. Each terminator has an associated help page, see mlr_terminators_[id].

This dictionary can get populated with additional terminators by add-on packages.

For a more convenient way to retrieve and construct terminator, see trm()/trms().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict, ..., objects = FALSE)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "label", "properties" and "unit" as columns. If objects is set to TRUE, the constructed objects are returned in the list column named object.

See also

Examples

as.data.table(mlr_terminators)
#> Key: <key>
#>                 key                label             properties        unit
#>              <char>               <char>                 <list>      <char>
#> 1:       clock_time           Clock Time single-crit,multi-crit     seconds
#> 2:            combo          Combination single-crit,multi-crit     percent
#> 3:            evals Number of Evaluation single-crit,multi-crit evaluations
#> 4:             none                 None single-crit,multi-crit     percent
#> 5:     perf_reached    Performance Level            single-crit     percent
#> 6:         run_time             Run Time single-crit,multi-crit     seconds
#> 7:       stagnation           Stagnation            single-crit     percent
#> 8: stagnation_batch     Stagnation Batch            single-crit     percent
mlr_terminators$get("evals")
#> <TerminatorEvals>: Number of Evaluation
#> * Parameters: n_evals=100, k=0
trm("evals", n_evals = 10)
#> <TerminatorEvals>: Number of Evaluation
#> * Parameters: n_evals=10, k=0