Combine Terminators

This class takes multiple Terminators and terminates as soon as one or all of the included terminators are positive.

Dictionary

This Terminator can be instantiated via the dictionary mlr_terminators or with the associated sugar function term():

mlr_terminators$get("combo")
term("combo")

Parameters

any (logical(1))
Terminate iff any included terminator is positive? (not all), default is `TRUE.

See also

Super class

bbotk::Terminator -> TerminatorCombo

Public fields

terminators

(list())
List of objects of class Terminator.

Methods

Public methods

TerminatorCombo$new()

TerminatorCombo$is_terminated()

TerminatorCombo$print()

TerminatorCombo$clone()

Inherited methods

bbotk::Terminator$format()

Method new()

Creates a new instance of this R6 class.

Usage

TerminatorCombo$new(terminators = list(TerminatorNone$new()))

Arguments

terminators

(list())
List of objects of class Terminator.

Method is_terminated()

Is TRUE iff the termination criterion is positive, and FALSE otherwise.

Usage

TerminatorCombo$is_terminated(archive)

Arguments

archive

(Archive).

Returns

logical(1).

Method print()

Printer.

Usage

TerminatorCombo$print(...)

Arguments

...

(ignored).

Method clone()

The objects of this class are cloneable with this method.

Usage

TerminatorCombo$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

term("combo",
  list(term("clock_time", stop_time = Sys.time() + 60), term("evals", n_evals = 10)),
  any = FALSE
)

#> <TerminatorCombo>
#> * Parameters: any=FALSE
#> * Terminators: <TerminatorClockTime>,<TerminatorEvals>